New functionality added to ParMetis & Jostle interface: If the new parameter SCATTER_GRAPH is turned on, then the ParMetis-style graph data structure will be distributed (scattered) onto all processors. Currently each processor receives equally many vertices. Weights are not used in this version.
The partitioner, ParMetis or Jostle, is then invoked using the new
temporary data distribution. The resulting partition is mapped
back to the initial distribution.

The SCATTER_GRAPH option is particularly useful if all data
is initially located on one processor. With SCATTER_GRAPH one
can circumvent current bugs in ParMetis.

The JOSTLE_SCATTER parameter is now obsolete and will be removed shortly.