New_IDs_Branch merged with main trunk of Zoltan!
The data types for IDs in Zoltan are no longer compiled into the library.
Instead, the IDs are now arrays of unsigned integers.  The number of
integer entries in an ID can be set through LB_Set_Param.  Global and local
IDs may have different numbers of entries.  The LB_ID_TYPE is currently
defined to be unsigned integers; in the F90 interface, IDs are integers.

driver:
+ Changes to support new argument lists for Zoltan functions and query
  functions.  Query functions modified to support the new IDs format.
  zdrive currently uses one integer each for global and local IDs.

Still remaining to do:
+ Testing, testing, testing (any volunteers?).