New_IDs_Branch merged with main trunk of Zoltan!
The data types for IDs in Zoltan are no longer compiled into the library.
Instead, the IDs are now arrays of unsigned integers.  The number of
integer entries in an ID can be set through LB_Set_Param.  Global and local
IDs may have different numbers of entries.  The LB_ID_TYPE is currently
defined to be unsigned integers; in the F90 interface, IDs are integers.

oct:
+ Removed use of LB_TAG data structure within Region structure.  Added use
  of dynamically allocated IDs.

Still remaining to do:
+ Incorporate new octree code from RPI.
+ Testing, testing, testing (any volunteers?).