----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	build/TRILINOS_TARGET_DEFS
 	packages/epetra/example/petra_nonlinear/B.cc
 	packages/epetra/example/petra_nonlinear/basis.cc
 	packages/epetra/example/petra_nonlinear/basis.h
 	packages/epetra/example/petra_nonlinear/cc_main.cc
 	packages/epetra/example/petra_nonlinear/main.cc
 	packages/epetra/src/Epetra_BlockMap.cpp
 	packages/epetra/src/Epetra_BlockMap.h
 	packages/epetra/src/Epetra_CrsGraph.cpp
 	packages/epetra/src/Epetra_CrsMatrix.cpp
 	packages/epetra/src/Epetra_LocalMap.h
 	packages/epetra/src/Epetra_Map.h
 	packages/epetra/src/Epetra_MultiVector.cpp
 	packages/epetra/src/Epetra_VbrMatrix.cpp
 	packages/epetra/src/Epetra_VbrMatrix.h
 	packages/epetra/src/Epetra_Vector.cpp
 	packages/epetra/test/BlockMap/checkmap.cpp
 	packages/epetra/test/Map/checkmap.cpp
 Added Files:
 	build/Eqn2Point.sed build/foreachEqn2Point
 	packages/epetra/test/ReferenceLogs/log.04Mar2002:11:17:23
 	packages/epetra/test/ReferenceLogs/log.mpi.04Mar2002:11:17:03
 ----------------------------------------------------------------------
The primary purpose of this update is to improve some terminology in the BlockMap class.  Specifically:
-What was formerly referred to as an equation is now called a point.  Equation was deemed to have too many differing usages.
-What was formerly referred to as a Block within a BlockMap is now an Element.
-What was formerly labeled a FirstElementEntry is now a FirstPointInElement.

PLEASE NOTE:
These changes led to the updating of some method names.  Specifically:
FindLocalBlockID is now FindLocalElementID
NumGlobalEquations is now NumGlobalPoints
NumMyEquations is now NumMyPoints
FirstElementEntryList is now FirstPointInElementList
EquationToBlockList is now PointToElmentList

Outside calls to these methods will NEED TO be updated.  A simple way to do this is to use the Eqn2Point.sed sed file that was checked in along with this update.
To begin to update your method calls, goto a directoy in which there are files that need to be updated.  Then use the "foreachEqn2Point"
script in Trilinos/build that can be sourced to execute the
Eqn2Point script on each file in a directory.  This foreach script
does the following in the current working directory:

- Creates two directories: "old" and "new".
- Then for each file (*.h *.cc *.cpp) it:
- Copies the file to the "old" directory.
- Executes the sed script, putting a new version of the file in "new".
- Prints a "diff" of the old and new version of the file.

You will then need to manually move the files from the "new" directory to the original directory.  Be sure to retain the "old" directory until everything is working perfectly.  This process is identical, though more simple in complexity, to the update from Petra to Epetra.  I used the scripts to update a second copy of the Epetra tests on my machine, and had no problems at all.
If problems arise for you as a result of this update, please contact me at 320-685-4139, or jmwille@sandia.gov.

The updates to the other files include documentation improvements and updated calls to the methods listed above.

I ran all serial and MPI tests, all tests passed, logs are included in the update.