----------------------------------------------------------------------
 Modified Files:
 	Epetra_Comm.h Epetra_DistObject.cpp Epetra_MpiComm.h
 	Epetra_MpiSmpComm.h Epetra_SerialComm.cpp Epetra_SerialComm.h
-------------------------------------------------------------------

Changed the name of the virtual print method in Epetra_Comm.h from Print() to PrintInfo().
The purpose of this was to remove possible ambiguities with the Print() method in Epetra_Object.h
and the implementation of Print() in the concrete comm classes: Epetra_SerialComm, Epetra_MpiComm
and Epetra_MpiSmpComm.

Had to change Epetra_DistObject.cpp to reflect this change.