----------------------------------------------------------------------  Modified Files:  	makefile.SGI32.MPI makefile.SGI32.MPI_DEBUG  	makefile.SGI32.MPI_PURIFY makefile.SGI32.SERIAL  	makefile.SGI32.SERIAL_DEBUG makefile.SGI32.SERIAL_PURIFY  	makefile.SGI64.MPI makefile.SGI64.MPI_DEBUG  	makefile.SGI64.SERIAL makefile.SGI64.SERIAL_DEBUG -------------------------------------------------------------------
Commented out the addition of -ptused since Rob Hoekstra found a less heavy-handed
approach to solve the problem with Epetra_SerialComm by inlining the method that was causing
problems.  We thought that requiring apps to use -ptused might have unintended side-effects.
However, the flag is only commented out, not removed, in case we need it in the future.