----------------------------------------------------------------------  Committing in Trilinos/packages
 Modified Files:
 	aztecoo/example/AztecOO/cxx_main.cpp
 	epetra/src/Epetra_BlockMap.cpp epetra/src/Epetra_BlockMap.h
 	epetra/src/Epetra_CrsMatrix.cpp
 	epetra/src/Epetra_MultiVector.cpp
 	epetra/src/Epetra_MultiVector.h epetra/src/Epetra_Vector.cpp
 	epetra/src/Epetra_Vector.h
 	epetra/test/MultiVector/ExecuteTestProblems.cpp
 	epetra/test/MultiVector/cxx_main.cpp
 ----------------------------------------------------------------------

The mods here add support for post-construction modification of vector
and multivector values.

In the MultiVector class I have added methods that allow a single
value in the multivector to be replaced or summed into, using either
global or local index space.  Furthermore, there is an interface that
is oriented  for block maps and plain maps.  The sum total of methods
here is 8.

Complimenting these 8 methods, I have added 8 methods to the vector
class that allow one to sum in or replace a list of values.  This also
works for global or local index space and block maps where the Block
offset is applied uniformly to all indices in the list.

Note that there is no similar set of methods for multivectors.  I
think that supporting a list of updates for a multivector is too
cumbersome and unlikely to be used.  Instead I have overloaded
operator(int i) in the Epetra_MultiVector class to return the ith vector
in an Epetra_MultiVector as an Epetra_Vector pointer.  Having this feature,
one can use the list-style modification capabilities for each
vector of a multivector.  I think it is far more likely that a user
will make changes to a single vector at a time, even when working with
multivectors.

I have added tests for all of the above features and all tests pass in
serial and parallel.