----------------------------------------------------------------------  Modified Files:  	example/aztec_app/Makefile src/az_aztec.h src/az_aztec_defs.h  	src/az_aztecf.h src/az_cg.c src/az_comm.c src/az_converge.c  	src/az_dd_overlap.c src/az_matrix_util.c src/az_matvec_mult.c  	src/az_solve.c src/az_tools.c src/az_util.c  ----------------------------------------------------------------------
Merged the minor differences that were part of the Aztec 2.1 source
repository into the Trilinos AztecOO repository.  Although AztecOO was
developed starting with Aztec 2.1, minor change occurred in a number
of the files that were common to both instances.

In order for this set of mods to have lasting meaning one of the two
following actions MUST happen:

1) NO FEATURES SHOULD BE ADDED TO THE AZTEC 2.1 REPOSITORY.

2) ANY CRITICAL BUG FIXES TO AZTEC 2.1 MUST BE ADDED TO AZTECOO.

Anyone who does not heed the above will be sent a MASSIVE snowstorm
from Minnesota!!!

In addition, it is to everyone's advantange to STOP using Aztec 2.1
and switch to AztecOO, EVEN IF you are using only the old Aztec
interface.  Doing so will help us see how well AztecOO supports
existing Aztec users, and minimizes the likelihood that Aztec 2.1 will
get out-of-sync with AztecOO.

All of the above changes have passed all the tests that I can run.  However,
there are many "features" in Aztec that have no explicit test
coverage.  Please inform me of strange behavior.