Added an update norm status test called NOX::StatusTest::NormUpdate. This is required for Salsa and Alegra.
Changed the name of WRMS to NormWRMS to be consistent with other tests.  Cleaned up the documentation and implementation.

Added the ability to print the status test results at each outer iteration for the LineSearchBased and TrustRegion solvers.  This can be turned on by adding the flag "OuterIterationStatusTest" to the print level.  I have kept this separate from the "OuterIteration" flag since some simulations don't want all this data printed at each iteration while other require it.

Examples of usage for each of these changes is shown in the 1Dfem_nonlinear test problem.