----------------------------------------------------------------------  Modified Files:  	example/ReducedLinearProblem/cxx_main.cpp  	src/Epetra_BlockMap.cpp src/Epetra_CrsSingletonFilter.cpp  	src/Epetra_CrsSingletonFilter.h src/Epetra_MpiComm.h  	src/Makefile  ----------------------------------------------------------------------
Latest changes to support singleton filter capabilities.  The class Epetra_CrsSingletonFilter appears to be quite
robust, working across a broad set of problems from Xyce, ChileSpice and test matrices from Tim Davis.  This class
still has one major shortcoming in that it cannot handle nonsymmetric filtering.  In other words, the list of rows
and columns that are removed must be identical.  Many problems have this characteristic, but not all.

The analysis works for support nonsymmetric filtering and I can form a reduced system, but none of our solvers allow
the domain and range maps to be different.  Therefore, I need to add permutation support to Epetra in order to
effectively implement a reduced system that our solvers can handle.

If the CrsSingletonFilter class detects a nonsymmetric filter, it will return with a fatal error code in the analysis
phase.  I hope to get rid of this restriction soon, but not right now.