----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	makefile makefile.WIN2000.SERIAL_DEBUG
 ----------------------------------------------------------------------
Fixed the problem that Trilinos rebuilds all libraries everytime make is run with the help of Kevin L, Mike H, and Rich L.  This was broken in a previous commit.

Also updated makefile.WIN2000.SERIAL_DEBUG to use ln for linking instead of ln -s, as ln -s is not portable to Windows.