Lots of changes here.  Basically, everything is pretty much self-contained now.  There are 2 line search options (Standard or Curvilinear) that are contained within the class.  Code is also cleaned up and commented better.  I tried to follow the format layed out in Linesearch::Quadratic since that one has had the most number of eyeballs looking at it...
To use this solver, instantiate a new object of type
NOX::Solver::TensorBased since the solver is not an option from within
the Solver::Manager.

Doxygen comments are not fully tested, but I think it is okay.  If you
find, wrong comments, feel free to modify.  If any algorithmic
bugs/changes, then please let me know.