Fixed some bugs and added some functionality for the curvilinear linesearch.  I have a public function called computeCurvilinearStep(dir, lambda) that will return the new step. Since the curvilinear LS needs to call this public function, I could not use the direction::manager inside of Solver::TensorBased. Instead, I instantiate a Direction::Tensor object directly and use that.
The one thing that I am aware will fail is during a "breakdown" of the
Arnoldi process. I will work on a solution to that (part of my
thesis), but I don't ever expect to encounter this scenario in practice,
especially on such large problems.

I will probably continue to trim this file down, since there are a lot
of debug statements that are cluttering the file.  So you'll probably
see another checkin later...