----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	triutils/src/Trilinos_Util.h
 ----------------------------------------------------------------------

This update is the beginning of the second phase of integrating the
new Autotools configure/make system into the repository.  This phase
updates header files and makes other small changes to
source code.  Most of the changes made in the phase are necessary to
make Trilinos receptive to Autoconf's config.h files.