----------------------------------------------------------------------
 Modified Files:
 	Epetra_CrsGraph.cpp Epetra_CrsGraph.h Epetra_CrsMatrix.cpp
 	Epetra_CrsMatrix.h Epetra_VbrMatrix.cpp Epetra_VbrMatrix.h
 ----------------------------------------------------------------------

Fixes bugs 34 and 137.

The following mods include a complete re-write of MakeColMap() in
Epetra_CrsGraph.  This method had a major flaw:  It force all row GIDs
to be column GIDs, even if there were no column entries associated with
a given row GID.  This created extra communication, but more importantly
caused incorrect construction for tall, skinny matrices.

One additional change was made to the CheckSizes() method in the Epetra_CrsGraph,
Epetra_CrsMatrix and Epetra_VbrMatrix.  While testing the tall skinny matrices,
I had a simple error in my test code that CheckSizes() should have been able
to catch, namely I had not called TransformToLocal() on a matrix prior to trying
to use the matrix in an Export() operation.  This check is now done.