Added attributes to Epetra_FECrsMatrix and Epetra_FEVector to allow them to optionally ignore off-processor contributions. In this case, the GlobalAssemble method is also a no-op.
This 'ignoreNonLocalEntries' attribute is a default constructor argument,
default value being false (by default the behavior is to NOT ignore
off-processor contributions) and there is also a method for setting the
attribute after construction.

This addresses bug 191.

Epetra tests passed.

(Fixed a bug in test/FECrsMatrix and test/FEVector whereby the tests
would hang when run on multiple processors in verbose mode, because there
was a matrix-print and vector-print being called within an "if(verbose)" block.
The tests now complete without hanging in verbose mode.)