Provided an applyJacobianInverseMulti for LAPACK groups.
I also forgot to mention in the NOX LAPACK commit that I added data members
to store the LU factorization of the jacobian matrix, the pivot array, and
an isValid flag for the LU factorization.  This allows repeated solves of the
Jacobian to reuse the LU factorization.  Note that these members are stored
as mutable since they must be updated in applyJacobianInverse, a const method.