Added new logic to LOCA::Stepper::finish() to do one additional step of natural continuation to hit the target parameter value.  Note that this additional step uses the random predictor (instead of the constant predictor) since the constant predictor often does not work for continuing turning points when the system Jacobian does not depend on the continuation parameter.
Modified LOCA::Stepper::stop() so that when current parameter is checked
against maxValue and minValue, the sign of the parameter step is also taken
into account.  This fixes the bug for when the starting value is equal to
one of the bounds.

Made several minor changes to eliminate warnings from "g++ -Wall".  Most of
these involved changing int's to unsigned int's, but a few actual bugs were
discovered as well.