----------------------------------------------------------------------
Committing in .

Modified Files:
	test/BlockCG/classicMakefile
Added Files:
	test/classicMakefile
----------------------------------------------------------------------
Created a small belos test library with the code that is to be common
amongst multiple tests in the future.  This code is located in belos/test.

First build this library, then build the test exe.  If there is a better
solution to the multi-layer test directory structure, this can be changed.

The test exe now builds on Linux.