----------------------------------------------------------------------
Committing in .

Modified Files:
	TrilinosDevGuide.tex
----------------------------------------------------------------------
Partially addresses bug 276 (addresses only part 3)

Added discussion about the "-P" flag for CVS.

Also moved the CVS command section into the CVS appendix.  It is really
additional info for those who are not familiar with CVS, so it is
not appropriate for the main body of the document.  However, it makes
sense to keep it as an appendix.  The appendix should have a description
of a few more CVS commands added to it.