----------------------------------------------------------------------
Committing in .

 Modified Files:
 	AnasaziBLAS.hpp AnasaziBlockArnoldi.hpp AnasaziDenseMatrix.hpp
 	AnasaziLAPACK.hpp AnasaziLOCAInterface.hpp AnasaziMatrix.hpp
 	AnasaziMultiVec.hpp AnasaziPetraInterface.hpp
 	BlockArnoldiPetraEx.cpp README
 Added Files:
 	AnasaziDescription.hpp classicMakefile doxygen_config
----------------------------------------------------------------------

The commenting methods in these files were changed to allow for the creation of
doxygen documentation for Anasazi.  The classicMakefile was erroneously omitted
from the first commitment of Anasazi.  The AnasaziMultiVec and its current
interfaces were changed to include initialization of the multivector elements.
The AnasaziDenseMatrix frobenius norm was corrected.