----------------------------------------------------------------------
Committing in .

Modified Files:
	test-harness
----------------------------------------------------------------------
Minor change to the test-harness that should help to clarify output.

In the past, there were sometimes two attachments called summ.txt and
summary.txt.  I changed summary.txt to EmailBody.txt.  This file is
supposed to the be the body of the email, but sometimes it is not; I
think it depends on your MUA (Mail User Agent).  In any case, summ.txt
contains a summary of the results (passed or not) while EmailBody.txt
shows the results for each individual test and lists the script name and
owner.