----------------------------------------------------------------------
Committing in .

Modified Files:
	configure.ac configure aclocal.m4
----------------------------------------------------------------------
Fixes bug 421.

The malloc header for a DEC is not recognized by the AC_FUNC_MALLOC
macro.  We will check for the header, and if not present, use
AC_FUNC_MALLOC (this is useful for a MAC).

Linux tests passed, must now test on DEC and MAC.