Moved AC_ARG_WITH for epetraext support further down in the configure.ac file so that the help would line up better.
Also, changed --with-nox-epetraext to --with-epetraext.

Slightly changed the help line for loca to say it will compile
libraries (as opposed to a single library).

Removed all help strings related to Anasazi. Note that I had to
comment with "dnl" rather than a "#" because of the way that the M4
macros work.

Reordered the "with" statements so that the incdir and libdir
statements were together.

Removed ALL configure support for y12m. We don't need it any more. We
were just using it for some of our examples. This does not mean that
users cannot use y12m with NOX - just that we don't use it in any of
our examples.

Among other things, this fixes Bug #442.