Added a check taken from Rob Hoekstra's epetraext library that will define MUST_CONST_STL_MAP_KEY if required.
This is due to an STL bug on solaris. Solaris requires a const argument
in constructing maps. This is wrong according to the standard!
DEC on the other hand fails if the argument is const. Therefore, we
have to wrap multimap construction with

#ifdef MUST_CONST_STL_MAP_KEY
  ....
#else
  ....
#endif

for compilation on both Dec and Solaris.