Implemented the multiple right-hand-side version of applyJacobianInverse() for the modified turning point bordering group.  This reduces the number of linear solves per continuation step from 10 to 7 for arc-length continuation of turning points using this method.
Changed how the orientation of the tangent predictor is set in
LOCA::Continuation::ExtendedGroup::computeTangent().  Previously it was
determined by computing the dot product of the tangent vector with the
secant vector.  Now it merely looks at the sign of the parameter component
of the secant vector.  This fixes problems where the dot product was the
wrong sign because of the solution component dominating the parameter
component.