Fixed a memory error in FECrsMatrix that was causing a bug reported by Clif Drumm.
In the course of finding that bug, I used purify and noticed some
free-memory-read errors. One was caused by the fact that the
BlockMapData class held a Directory member, and that Directory member held
a pointer to a BlockMap. The BlockMapData class is reference-counted so that
its lifespan doesn't have to coincide with the lifespan of the Map that
created it. But this leads to cases where the Directory is pointing to a
Map that no longer exists. The Directory object was only used by the
method BlockMap::RemoteIDList, so the solution is to create the Directory
within the RemoteIDList method and not have it as a BlockMapData member.

Also fixed a minor memory leak in Epetra_HashTable and removed another
un-necessary dynamic_cast operation from Epetra_MpiComm.

Verified all epetra tests pass in serial and parallel on linux. Used purify
on a couple of different epetra test executables running on solaris.