(1) Changed Error::raise calls to TEST_FOR_EXCEPTION() throughout. Since TEST_FOR_EXCEPTION has different semantics than Error::raise, this required some restructuring of the error tests.
(2) Rewrote Teuchos_Array so that it derives from std::vector. Array is now
an extension of vector<> that does the following: (a) provides
conditionally-compilable boundschecking on the [] operators, (b) provides
some more obvious method names, such as append() instead of push_back(),
and (c) provides auxiliary methods such as toString(), hashCode(), and the
tuple() methods.

(3) Replaced Teuchos_Stack with std::stack in Teuchos_TreeBuildingXMLHandler.
Since Teuchos_Stack provided no capability not in std::stack, and had a near-identical API, there was no reason to maintain it.