----------------------------------------------------------------------  Added Files:     henricson_nyquist.pdf  ----------------------------------------------------------------------
Added a pdf of the Henricson & Nyquist "C++ Rules and Recommendations" file to the repository. This way the Coding Guidelines can link to a local copy
of it instead of to a (potentially non-permanent) copy on an external website.

(The Henricson & Nyquist document gives permission to redistribute/modify/etc. as long as the original copyright notice is preserved.)