When I committed my latest changes for multivector and Teuchos::any support in NOX, I accidentally put the TRILINOS_HEADER and TRILINOS_LIB configure checks for Teuchos inside an "if" block that was only executed if NOX Epetra support was enabled.  Therefore the HAVE_NOX_TEUCHOS preprocessor macro and the Teuchos include/lib paths were only defined by configure if NOX epetra support was enabled, causing the compilation of NOX to fail in many circumstances.  I didn't detect this in my testing before committing because I always configure NOX with epetra support enabled.
So, these changes should fix this and should fix bugs #645 and #647.  I have
tested configure with all of the different option combinations and it seems
to function as expected.

And just for the record, NOX needs teuchos if and only if NOX is configured
with --enable-loca (the default) or --enable-nox-multivectors or
--enable-nox-any (both off by default).  Currently --enable-teuchos is not
enabled by default in the top-level Trilinos configure nor is it turned on
if LOCA is enabled, so to build NOX correctly, Trilinos must be configured
with --enable-teuchos OR --disable-loca.  And finally, NOX does not need
Teuchos-experimental.