----------------------------------------------------------------------
 Modified Files:
 	src/Ifpack_CrsRiluk.cpp
 ----------------------------------------------------------------------

Fixes bug 667.

Added a line of code to zero out the diagonal of the incomplete factor prior to
initializing the values of the L, D and U factors.  This addresses a problem that
could be seen in repeated calls to the InitValues method, since the diagonal was
being added to, and there were inversed diagonal values still present in D.

Although this bug is not fatal in practice, it is likely that changes will be seen
in the behavior, mostly for the better (since the diagonal will now be a better representation
of the original matrix.  However, in some cases performance may degrade.

This change has been tested in serial and parallel.