----------------------------------------------------------------------
 Modified Files:
 	AztecOO.cpp
 ----------------------------------------------------------------------

Fixes bug 727.

Added a check for nonzero return value when calling the Apply and ApplyInverse
methods in AztecOO.  Because Aztec's matrix-free mechanism does not allow error code returns,
I had to add a throw for the cases where non-zero error codes were returned.  At least this
can be caught if the user desires.