Committing initial redesign of Belos framework.
This is only intended to be a place to keep changes until I'm confident
enough to use this as the Belos distribution.

This redesign includes a StatusTest base class with derived classes like
MaxIters, MaxRestarts, and ResNorm.  It also includes an virtual
IterativeSolver base class and LinearProblemManager class.

Not all the functionality has been tested, nor is the design finished.