----------------------------------------------------------------------  Committing in .
 Modified Files:
 	src/Epetra_VbrMatrix.cpp test/VbrMatrix/cxx_main.cpp
 ----------------------------------------------------------------------

Fixes bug 649.

Fixed a problem in ExtractEntryCopy in Epetra_VbrMatrix.cpp.  The code assumed
that the leading dimension equals the row dimension.  This problem is very old,
but was not picked up in the tests because there was another error which caused
the leading dimension != row dimension case to not execute like it should.

The test driver was also fixed. All tests were run and they passed.