----------------------------------------------------------------------  Modified Files:  	example/EpetraTpetraProfile/cxx_main.cpp  	src/Tpetra_CisMatrix.hpp src/Tpetra_ElementSpace.hpp  	src/Tpetra_Vector.hpp  ----------------------------------------------------------------------
Slightly tweaked the temporary fix in fillComplete. (secondaryDist_ is cpy constructed equal to primaryDist_, so we don't need to assign it.)

Changed the print functions in Vector, and commented out some unused variables in ES and Vector.

Commented out some extra diagnostic outputs in the testdriver, including a tpetraOnly method I was using.

Also changed how the residuals are computed so that it computed them for Tpetra properly.

The testdriver should work properly for (most all) cases now.