Fixed a bug that shows up in MPSalsa.  Eric - I'm not sure if this is how you want to fix this problem but it gets salsa working again.
In MPSalsa we always construct a LOCA::EpetraNew::Group even if only
using the group for a nox solve.  This fails with the latest version
of loca because a base class constructor for LOCA::EpetraNew::Group
found in the file: LOCA_Bifurcation_TPBord_SingularSolveGroup.H is
accepting a const parameter list and trying to access the "Singular
Solve" sublist.  This sublist does not exist unless you are doing a
singular solve and so the sublist() call fails due to the list being
const.  I just added a const_cast so if the list does not exist, it
can be created.