I have removed the runtime error checking for ScalarTraits<>.  As Scott Myers states in Item 46 in his book "Effective C++", "Prefer compile-time and link-time errors to runtime errors".  Now if you try to instantiate the default ScalarTraits class you will get a compile-time error that will will contain the string \'T::this_type_is_missing_a_specialization\'.  I put this in because I compiled a piece of test code that used ScalarTraits<std::complex<> > and then got a strang runtime error.  It took me an hour or more to figure out that I had to configure teuchos with --enable-teuchos--complex to fix the problem.  Now that code would not have compiled.
I also removed the functions eps(), sfmin(), base(), prec(), t(), rnd(), emin(), rmin(), emax() and rmax() from the specializations for ScalarTraits<int> and ScalarTraits<mp_real> since these are not defined for these types.  I just could not find a good portable way to allow then to be defined but still result in a nice compile-time error stating that they were not defined for this type.  Hopefully when someone trys to call these functions they will look at the header file and see the comment that these functions are purposefully not defined.  The test suite build and ran sucessfully (except for the test for the packages EpetraExt and NOX have not not worked for some time).

If there is any problem with using this new compile-time safe implementation of ScalarTraits<> please come talk to me (Roscoe Bartlett, rabartl@sandia.gov) and I will show (or tell) you how to resturcture your code to do the correct thing at compile time.

Strive for compile-time checking.