: Modified Files: : 	Teuchos_ScalarTraits.hpp
seedrandom(), specialized for the mp_real datatype, was set up to call an mp_real function that doesn't exist -- mp_srand(). There does not appear to be a function native to ARPREC that seeds the random() function.

Since the mp_real datatype is made up of doubles, seeding the random function for doubles should hopefully be sufficient. Upon a suggestion from Heidi, I changed seedrandom() to cast the input unsigned int parameter to a long int, then call srand48() with that value. This appears to work correctly based on limited testing.