----------------------------------------------------------------------
 Committing in .

 Added Files:
 	reference-options
 ----------------------------------------------------------------------
This file contains all of the options that I could find for Trilinos
as of a couple of weeks ago.  We may never be able to use all options
on all platforms, but it still will be nice to have a reference list
that we can update periodically.  If nothing else, we can start with
this list of options when we are installing the test harness on a new
machine.