Added option to use a provided row and column permutation, or the natural ordering.  Added an option to control the row scaling used (no scaling, row abs. sum, row abs. max).  Control parameters and and statistical output arguments simplified.
Added all of the MATLAB mexFunction interfaces to btf, klu_kernel, and klu_btf,
to provide examples for developers of other packages.  The Makefiles and
GNUmakefiles in the btf, klu_kernel, and klu_btf directories compile the
mexFunctions.  This allows MATLAB to call the klu codes.  For example,
with the btf mexFunction the MATLAB statement "[p,q,r] = btf (A)" is
the same as "[p,q,r] = dmperm (A)", except that the exact permutations
can differ.

Added support for KLU's refactorization (factor a matrix using the pivot
order from a prior numerical factorization), with help from Marzio Sala.