Added A'x=b solve (klu_btf_tsolve); not yet in Amesos interface, however.
If you have a compressed-sparse-row matrix, you could skip the transpose
in Amesos and can pass the matrix directly to KLU (factorizing A'), and follow
this by a solve of A'x=b.

Improved run time of forward/backsolve with multiple right-hand-sides, via
stride-1 access in innermost loops.  Example with Hamm/scircuit
matrix:  forward/backsolve of Ax=b takes 0.04 + (k-1)*0.02 seconds
for k right-hand-sides, with k = 1 to 4.  The 5th, 9th, etc.
right-hand-side takes 0.04 seconds because the solve is done in
chunks of 4 columns at a time.  This is limited by the 4n workspace.

The solve of A'x=b is slightly faster (by about 10%) than the solve of Ax=b.
Both forward/backsolves for Ax=b and A'x=b have stride-1 access when b
is a matrix instead of merely a column vector.

A default compile-time parameter added to remove as many floating-point
divisions as possible.  Prescaling and dividing a column of L by the pivot
is now done via multiplication by the inverse.  The inverse of the diagonal
is now stored, eliminating divisions in the forward/back solves.  To turn
this off and get the conventional behavior, compile with -DNRECIPROCAL.