Minor modifications to residual norm and multivector file.
Changes to TSFCore interface file include changing 0-based indexing to 1-based
indexing used by TSFCore.  The cloning methods were corrected and the MvAddMv
checks to see if any of the inputs are the same as *this.

Also committing new configure.ac/configure files that use the macros that were
just checked in.