Fix some small problems with Anasazi functions (still to be tested).
Now the code can use Anasazi to estimate the eigen-stuff, then turn to
the ml_ggb.c files to build the GGB correction.

Add another option (a.k.a. "fattening", I called it "enhanced"
mode). The code can create a hierarchy using METIS and the specified
null space. A couple of prelimiary test may suggest that this can boost
the converge quite a while.