Changes to use both serial and parallel coloring supported in EpetraExt.
It should be noted that "serial" coloring can be used in parallel so long as
elements are ghosted by the application.  The application must then support
two types of residual fills.  The parallel coloring capability
utilizes the same interface to fill residuals as is used outside the
Finite-Difference context, thus imposing no added requirements on the
application.

Examples of the use of each are provided in the recently added
epetra-examples/Brusselator_EpetraNew problem.