Added Belos to the test harness, so more information can be gathered about its portability, etc. while it's being developed.  Thus, the current tests had to be changed to use the new Belos interface to determine if the test passed.
Needed to add Dummy files to allow for copying of some HB matrices for the tests
during building.

Fixed tac_arg_check_default_feature.m4 macro.  A error in logic meant that this
test didn't act properly.

Removed some HB matrices that aren't being used in the test and don't need to be
kept in the Trilinos repository.