Commiting the current state of the refactoring for Belos to eventually be able to directly use TSFCore.  Currently all persisting relationships that used to use raw pointers now use RefCountPtr.  We have also created the MultiVecTraits class and its specialization for Belos::MultiVec.  All of BlockGmres is not refactored to use this new traits class.
We are just part of the way through doing this refactoring.

I have added new shell scripts to copy the source files back and forth to the main directories so that they can be built (see copy_to_main.x and copy_from_main.x).  This is done in such a way as to insure that changes to the main source files are not commited by acident.