Initial 2D data distribution appears to be working.  More testing would be nice; I'll continue testing but you can try this first version.
Currently using scheme (A) from Bob's notes for hindex, vindex.
Macros defining mappings between global and local numbers
and global numbers and processors are provided in phg.h;
algorithms that use those macros should need little or no
changes if we switch to scheme (B).

Mapping from the 2D data distribution back to the Zoltan GID/LIDs and
RETURN_LISTS is not yet implemented; since we can't generate partitions
yet I still have time to work on this part.  Currently, you might see
SEGV from this part of the code, but it shouldn't interfere with your
development.

I consolidated the include files; the previous structure was an artifact
of using hg_exec for the serial implementation.