----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	cxx_main.cpp
 ----------------------------------------------------------------------
In a recent change I created some positive return values that indicate
that a particular method worked, but was done in what might have been
an inefficient manner.  Returning a positive error code is not compatible
with the Trilinos Test Harness.  I changed the test file so that the
positive return code from the source code is reported in the test output,
but then if the final return code is positive, a zero error code is returned
by the test script.