----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	src/Epetra_CrsMatrix.cpp test/CrsMatrix/cxx_main.cpp
 ----------------------------------------------------------------------
Moved a method call out of a for loop.  Instead I saved the return value of
the method call and put it in a variable and then put that variable in the
loop.

Removed a warning from the leftScale method because, although there is
a potential inefficiency, there may not be an easy way to avoid this
in the vast majority of cases.

Updated the tests to reflect the removal of the warning.