Fixed the problems recently occurred in the test harness.
Introduced the -v option in TestMultiLevelPreconditioner. Now the examples
have almost no output by default. If an error is caught, the example is re-run
with -v (producing all output).

Introduced a new test (called TestExamples). This test jumps into the example
subdirectories, and executes most of the .exe that are there. The test cannot
deal in an easy way with the examples that require an input (either in the
command line, or an user-input).
The serial version runs on 1 process (of course). The parallel version on 2
processors. Goal of this test suite is to be sure that examples can run,
successfully, till the end. It does not compare the iteration number or any
other algorithmic feature of ML.

Authors of new examples should consider the following:
- possibly, the example should run without any input parameter specified at
  the command line;
- if an input parameter on the command line is necessary, then this example
  must be disabled from TestExample (serial and mpi)