----------------------------------------------------------------------
 Committing in .

 Modified Files:
 	Hello_Test.cpp
 ----------------------------------------------------------------------
The "mpi" test, which doesn't do anything that takes advantage of mpi at
this time, could not be run through mpirun.  I added some statements to
allow the test to be run through mpirun so that we at least have an
example for the test harness.  We should consider creating a test
that actually uses mpi in some way in the future.