Bug 371 reported that Epetra_VbrMatrix::Multiply failed if the input and result vectors are the same vector. In that situation, the vector came back zero'd out, regardless of the contents of the matrix.
It turns out that if the matrix was holding a non-null Exporter, then it
would do the right thing because it would create a temporary export vector
and form the matrix-vector product in the temporary, then copy the contents
to the result vector before returning. However, if the matrix is not holding
a non-null Exporter (e.g., if running on only one processor), then the result
vector is zero'd before calculations begin, causing the problem in the
bug-report.

I've now added a check for the case where the input and result vectors are
the same. If they are, then a temporary vector is created regardless of
whether an exporter exists. The matrix-vector product is formed in the
temporary and copied to the result vector before returning. This represents
a small performance penalty but it is unavoidable if the vectors are the
same. If vectors are not the same, no extra temporary is created and there
is no performance penalty.

Also began expanding and updating the doxygen documentation in
Epetra_VbrMatrix.h, and corrected a minor error in the comments in
Epetra_BlockMap.h.