Several changes to the Tensor-Krylov method code that hopefully cleans it up a bit more and provides a few more options to the user....
I added a macro for calculating a stopping condition that is closer in
style to the way that GMRES stops.  I showed in my thesis that are at
least 2 ways of calculating a stopping condition.  I used one in my
thesis research but never coded up the new option.

I also added an option to protect the curvilinear linesearch when
using restarts in the Tensor-Krylov method.  Previously, a tensor
direction computed with restarts would not guarantee that the
corresponding Newton direction would be a descent direction.  This
option now guarantees that the Newton step is a descent direction