zdrive will now interpret input in *.assign as partition and/or processor assignments.  If the max assignment value in *.assign is less than the number of processors being used, the assignment values designate both partition and processor assignment.  Otherwise, they designate only partition assignments, and INITIAL_LINEAR distribution is used to place objects on processors.
This is the first step in allowing Umit to test refinement from
a given partition distribution.  HG and PHG do not currently
use an input distribution; I will modify them to store the
input distribution and maintain it when no coarsening and
no coarse partitioning are specified.  Then Umit can run his
tests.