Fix for bug 1100.
Builds were failing on qt due to return type on Epetra_Operator
derived classes.  at some point a const was added to the Label()
command and we did not update the derived classes.  the compiler on qt
didn't like this.

Also re-organized the constructor initializer lists in the colored
objects to remove warnings about declaration order.