Fix bug 1120. Andy Salinger reported that calling AZ_iterate and then AZ_scale_f was failing in AZ_scale_f because it couldn't find the scaling that had been allocated during AZ_iterate. This is fixed now. It had been broken by changes I made in December to allow AZ_iterate to be used recursively (AZ_iterate calls AZ_precondition, which leads to an inner call to AZ_iterate).
Aztec's memory management is basically as follows:
  The function AZ_manage_memory obtains heap-allocated memory and stores it
  in a global linked-list along with 'type' and 'label' attributes.
  Subsequent calls to AZ_manage_memory with the same type and label will
  reuse the previously allocated memory rather than allocating new memory.

  Example: When preconditioning is performed during a call to AZ_iterate,
  memory for factorizations is allocated and tagged with data_org[AZ_name].
  If options[AZ_keep_info]==1 then this memory is not freed when AZ_iterate
  exits. On subsequent calls to AZ_iterate, if options[AZ_pre_calc]==AZ_reuse,
  the previously allocated factorizations can be reused.

Before the changes made in December, if AZ_iterate was called recursively,
the memory would be corrupted because the different invocations of AZ_iterate
were working with the same memory. The fix for that was to add a static int
that was incremented and decremented by AZ_iterate on entry and exit, and used
in the tags when allocating memory. This solved the recursive problem, but had
the side-effect that later calls to AZ_scale_f didn't know what tag to look for,
and couldn't find the previously allocated scaling.

The fix is to make use of the AZ_SCALING->mat_name attribute to store the
name of the matrix-name-tag, so that AZ_scale_f can find it if called from
outside of AZ_iterate.

I've expanded the test program in test/AztecOO to try to cover more of Aztec,
but it's still hard to be sure that no un-anticipated side-effects will
arise as a result of this change. But the test does cover basic use-cases
involving recursive calls to AZ_iterate, as well as successive calls using
AZ_keep_info and AZ_reuse. It also tests the case of calling AZ_scale_f
after AZ_iterate, and everything appears to be in order. I also used purify
to make sure I hadn't created a bunch of leaks or other memory errors.