This fixes bug#1137.
It looks like the removal of ML_EXTERNAL broke a part of the user-defined
prolongation capability.

Previously, the call to the user-defined function user_gen_prolongator would
use as the fourth (void *) data pointer either ag (type ML_Aggregate*) or data
(type void *), depending on the ML_EXTERNAL flag.   When this was removed, the
data parameter was incorrectly determined to be unused.  I have added an if
check to determine whether the data parameter is NULL.  If it is, then
user_gen_prolongator is called with ag as the fourth parameter.  Otherwise,
user_gen_prolongator is called with data.