Hello Zoltan members,
this is a first commit from me after a long time...
In hg.c I removed an small error which now clearly notices when
'fmkway' or 'grkway' is chosen. In hg_pack I fixed a problem when
an optimizer for the packing like aug2 is used after a calculation of
a packing is done where the last chosen hyperedge was not completely
packed due to the limit-parameter. Some other minor changes only
manipulate some information output.

Greetings, Robert