Removed the field    ML_GridAGX    *grid_info from the definition of ML_Operator.
Now, grid information (like coordinates) for a square ML_Operator, A, can be
accessed via the to or from field:

   grid_info = (ML_Aggregate_Viz_Stats *) A->to->Grid->Grid;
   for (i=0; i<Amat->outvec_leng; i++)
     printf("(x%d,y%d) = %d\n",i,i,grid_info->x[i],grid_info->y[i]);

The Grid field in "Amat->to" is really just a pointer into the array Grid of
the main ML structure.  Grid is of type ML_Grid, which has a void * data field.
This is where the ML_Aggregate_Viz_Stats* data now lives.