Merging change from trilinos-release-6-branch.
This will hopefully fix the nightly link failure on atlantis for
the example ml_aggr_user.exe.  I believe the linker is looking downstream only
for undefined symbols.   On the link line, libumfpack_atlantis.a (the
3rd party library that contains the undefined symbol) occurs after libscs.a,
the blas/lapack library on atlantis.

The proposed fix is to put a second instance of @LAPACK_LIBS@ @BLAS_LIBS@ in
the definition of ML_LIBS.  This is after @LIBS@ @FLIBS@, one of which I
believe contains the Trilinos 3rd party libaries.