Removed ... for now ... the while (kstart < nTotal) loop.
Since nTotal can vary within a processor column, and since
there is column communication within this while loop, the
while condition causes the code to hang when some processor
has nTotal=0.

Moreover, the while loop is not currently needed, as it
was executed more than once only when the partial-IP send
buffer was full.  Since we currently realloc that buffer,
the while loop is not needed.

We'll have to revisit this problem when we reactivate the
fixed-size send buffer (i.e., when we remove the realloc).

Erik's xyce problem now runs -- YEA!