Here I have done a few things:
1) I moved the experimental code from Teuchos_VerboseObject2.* and Teuchos_FancyOStream2.hpp to Teuchos_VerboseObject.* and Teuchos_FancyOStream.hpp and added this to the core library.

2) I removed the xxx1.* version of this code.

3) I added test/FancyOutputting/FancyOutputting_test.exe to test/defintion.  I need to add an output comparison to make sure this really is working.

4) I made a few changes suggested by Kevin Long:

4.a) I broke the current Teuchos::VerboseObject into a non-templated base class Teuchos::VerboseObjectBase that just as the OStream stuff and a templated class Teuchos::VerboseOBject that has class-specific verbosity setting stuff.

4.b) I added the ability to temporarily turn off tabbing and this is access through the Teuchos::OSTab object (so that destructor will turn this off automatically), even if an exception is thrown.

I also tried to fix the static initailization problem that Kevin mentioned but it could not work with the current setxxx(...) and getxxx(...) versions of the functions that I have in there.  The solution that I tried did not seem to work so I commented it out.  We need to revisit this at some point if this becomes a problem but I suspect that there will be a way to fix this and still keep the setxxx(...) and getxxx(...) versions.