Several changes:
1) Moved all LOCA source code over to using NOX::Utils instead of LOCA::Utils
and removed LOCA_Utils.{H,C}.  It is silly to have a separate LOCA::Utils
class when all that is different is two additions to the MsgType enum.
NOX::Utils doesn't use static methods (for good reason), so this change
required passing the LOCA::GlobalData object (which has a RCP to a
NOX::Utils object) to many new objects.  In particular, the LOCA groups
now need to a RCP to a global data object in their constructors.  There is
now a global function, LOCA::createGlobalData() which can be used to create
this object.

2) Made the LOCA::ErrorCheck methods not static, which required similar changes
as with NOX::Utils.  Now LOCA has no static methods, making it thread-safe
and allowing it to be called recursively.  In some places,
LOCA::ErrorCheck::throwError() was replaced by Teuchos' TEST_FOR_EXCEPTION(),
which doesn't require NOX::Utils eliminating the need for LOCA::GlobalData.
The upshot of this is LOCA may now throw non-string exceptions, so any
code calling LOCA should have:

   catch (std::exception& e) { ... }

after the try block.  Ultimately, I would like to transition everything over
in LOCA to using TEST_FOR_EXCEPTION(), but more needs to be worked out with
regards to NOX::Abstract::Group::ReturnType management.

3) Replaced all instances of cout and cerr with globalData->locaUtils->out()
in the LOCA source.  There are still cout/cerr calls in the tests and examples,
but I think this is fine.

4) Replaced the step size and Anasazi operator managers with factories.  The
transition from managers to factories is now basically complete, allowing
the user to change any strategy object without recompiling LOCA.

5) Updated all of the LOCA examples and tests to use NOX::Utils and to
create the LOCA::GlobalData object.  All tests pass with no heap leaks.