Change matrixmarket+ format to list number of processes, and index into list of pins for the start of each process.  This is more like mondriaan's distributed matrix format.
Process new zdrive input parameter "initial pins".  It can be:
  linear or block:    assign first n/p pins to process 0, and so on
  cyclic:  assign first pin in file to first proc, and so on
  file:    use the pin assignments specified in the matrixmarket+ file
  zero:    assign all pins to process zero
  row or rows:  each process gets entire rows, rows are not split up
  col/cols/column: each process gets entire columns

When we say a pin is assigned to a process, we mean that process
returns that pin in they hypergraph query functions.  So
this option helps in testing the hypergraph query functions. It
may also help in stressing the parallel hypergraph code.

Right now "initial pins" is only considered when matrixmarket+ files
are read.  I need to look at matrixmarket and hypergraph and see if
I can add it to them.

I also made the matrixmarket+ code obey the "initial distribution" and
"initial procs" zdrive parameters for distribution of the vertices.