Adding a missing include file, ml_common.h.  The symptom was a really unhelpful message in the link of examples/MLAPI/Blackboard.exe:
Blackboard-Blackboard.o(.text+0x6b5): In function `main':
: undefined reference to `vtable for MLAPI::MultiVectorCombination'

The problem, as I understand it, is that the v-table for class foo gets jammed
in the first .o file that implements one of foo's non-inline member functions.
In this case, it was MLAPI_Expressions.cpp (wrong), not
MLAPI_LinearCombinations.cpp (right).

A big thank you to Michael Gee for helping track this problem down.