Address bug 2061. When AztecOO::Iterate detects a solver status other than AZ_normal or AZ_maxits, it used to simply set a negative error code and then call EPETRA_CHK_ERR which would print the error-code and the file/line location to the screen, and return.
Now, if the solver status is something other than AZ_normal or AZ_maxits,
it determines whether EPETRA_CHK_ERR would print anything (i.e., if
Epetra_Object::GetTracebackMode() > 0), and if so, prints a meaningful
error message to the screen before exiting with the same return-code as
before. So no behavior is changed except the content of the screen output.

AztecOO tests run and pass.