Rewrite of Zoltan_PHG_Fill_Hypergraph: Using hypergraph query functions, processes return pins of a hypergraph, but may not return entire edges. Previous version did all-to-all communication to move entire edges to each process.  New version hashes global edge IDs to processes, and the assigned process is responsible for computing edge global number, edge weight, number of pins in edge.  We also hash global vertex IDs to processes in order to find out which process "owns" our pin vertices.
Added a field Remove_GEsize to the Zoltan_HGraph structure.  The
Remove_Esize field previously was the size (in pins) of the removed
edge, but now it is just a partial edge.  The Remove_GEsize field is
the number of pins in the edge across all processes that returned
part of the edge in the hypergraph query functions.

The new version is slower on small hypergraphs.  It is comparable on
medium hypergraphs (ibm_01, ibm_03).  We should assume it is faster
on large hypergraphs.