Re-commit some changes that were committed the day before the outage of software.sandia.gov. It appears that the repository was restored to a point consistent with a day or two before the outage, which means that some code changes and new files were lost. It isn't a disaster since I had everything on my own workstation, although a couple of newly added files were lost when I typed 'cvs update'.
This commit also includes a new src/zoltan directory, in which I've
pulled over Rob's code from epetraext/zoltan. I decided it is more
better to have that code in isorropia, since it makes it easier
to modify and more convenient to work on.

This commit gets us back to the point just before the outage. Next I'll
implement the changes we discussed in the meeting Thursday.