Renamed LOBPCGSolverManager and SimpleLOBPCGSolverManager to LOBPCGSolMgr and SimpleLOBPCGSolMgr for decreased verbosity. Started work on BlockDavidsonSolMgr; not complete yet.
Modified LOBPCG tests LOBPCG_test and LOBPCGComplex_test to use LOBPCGSolMgr.

Fixed a bug in StatusTestOrderedResNorm; it was calling the sort manager using MagnitudeType instead of ScalarType; this exhibited
only in for complex scalar types.

Went through LOBPCG solver and added correct accounting for timers and counters. Also modified setFullOrtho() so that, instead of
trying to orthonormalize P when necessary due to setFullOrtho(true), it just invalidates P. The user can do this herself if she
wants; full orthogonalization shouldn't be toggled frequently anyway.

Made minor changes to BlockDavidson; more on the way.