Continuing to address bug 2389, performance of matrix-matrix multiply functions.
Changed one of the kernels (A^T*B) to use an outer-product formulation.
Also made some code-level optimizations to a couple of the other kernels.
The performance of A*B and A^T*B are now significantly improved, as the
following timings show:

*********before:*************
size: 30000x30000, C = A*A, time: 2.3408
size: 30000x30000, C = A*A, time: 0.664326 (C already Filled)

size: 30000x30000, C = A^T*A, time: 2.9362
size: 30000x30000, C = A^T*A, time: 1.22234 (C already Filled)

*********after:*************
size: 30000x30000, C = A*A, time: 1.13844
size: 30000x30000, C = A*A, time: 0.637922 (C already Filled)

size: 30000x30000, C = A^T*A, time: 1.15695
size: 30000x30000, C = A^T*A, time: 0.63616 (C already Filled)

EpetraExt tests all run and pass.
Verified that the kernels run clean in purify.