merge from release7-0 branch
i went through the documentation and fixed \name section errors that i saw.

i also changed a typo in the \file command in TypeNameTraits and removed an
extra \brief statement for namespace Teuchos located in
Teuchos_ErrorPolling.hpp

a local compile of the documentation and tests suggests everything is
still in order.