Start of a Zoltan tutorial.
The tutorial is a latex file, but "make all" will convert it to
html and pdf.  tutorial.html is like ug.html and dev.html, but
it includes the Tutorial as a frame instead of as "native" html.

Reading the Tutorial in a frame is not fun, so I added a button
on the left to open the tutorial in it's own browser window.