Adding repartitioning metrics to parmetis code.
Please note that both in phg and parmetis, when calculating move cost
(and hence repart cost) we assume object is moved when its part
assignment is changed; not when its processor assignment is
changed. This needs to be fixed in both codes; but I'll leave it as it
is for now because 1) this should be OK when p=k. 2) we have more
serious problems :) it looks like phg_repart tends to move more
objects after first repartitioning, and this needs to be
investigated. I'm working on it now...