Two new top-level load-balancing methods.
RANDOM: This is a basic random partitioning method that
addresses bug 2877. It simply assigns objects to random
partitions. This may be useful to test application robustness
with weird data distributions. It is also a nice example for
new (and old) Zoltan developers as a short, concise method.

SIMPLE: Intended as a simple load-balance method that
tries to balance objects/weights but does not use geometry or connectivity.
Currently this is just a place holder that does nothing!

Possible bug: The random partition numbers don't look very random.
I'm looking into that. Work in progress.