Bug 2865
I have added a new parameter entry validator class called
AnyNumberParameterEntryValidator.  This class is used to allow a number to be
set in the parameter list as an int, double, or string and then to be read as
an int, double, or string (i.e. 3*3 cases).  I have tested this in the
ParameterList testing program.  This is a feature that was in the NOX
parameter list.  Note that a solver code can just create a single validator
object and then pass it into the 'set' operation for a bunch of different
numeric parameters, and it adds just a single extra argument value (in
addition to the documentation string argument, which should be set anyway)
when setting up the valid parameter list.