Here is what I have done as part of this checkin:
* Added a flag to the GLpApp::AdvDiffReactOptModel clas that removes support
for analytic derivatives.  This was needed to really test the Moocho/Thyra
support software for finite differenceing.

* Created macros to be used in decorator subclasses of
Thyra::ModelEvaluatorDelegatorBase in the evalModel(...) function to
standarize propagation of VerboseObject setup, printing evaluation traces, and
putting in Teuchos::TimeMonitors to globally record timing and number of
calls.

* Created a new decorator subclass call Thyra::DefaultInverseModelEvaluator
that can add a general inverse response function to any model that support a
parameterized state equation f(x,p)=0.

* Updated the MoochoThyraSolver class and the NLP-wrapper classes.

- Added more parameterlist control

- Broke out the finite differences for objective and constraints separately
for more control

- Added parameter to turn off tracing to the NLP/Model evaluator to reduce
clutter in the MoochoJournal.out file.

* I updated the run-varied-tests.pl script to test the new default inverse
objective function and using finite differences for the objective and/or the
constraints separately.

* Replaced a tab with four spaces in
Moocho.opt.DecompositionSystemStateStepBuilderStd

* I rebuilt to automatically created input for the MOOCHO documentation.