Added a function to return the current number of tab indentations.  I was going to use this in the AztecOO/Thyra adatpers to clean up the AztecOO printing but I ended up not using it there (see the comment in the source code).
The main thing I did through was to fix Stratimikos so that it set external preconditioner factories correctly for AztecOO and Belos.  This is only really being tested in Charon right now which is not acceptable.  Some major work is going to be needed to fully test Stratimikos and the Thyra wrappers that it gives access to.