These changes where suggested by Andy Salinger to make the EpetraExt::ModelEvaluator software a little easier to use.
1) I added a toString(...) function to convert
EpetraExt::ModelEvaluator::EInArgMembers to a string value and used it, and
the existing toString(...::EOutArgMembers) function, to implement the
TEST_FOR_EXCEPTION(...) error messages.  Before, if an unsupported argument was
accessed, it would throw an exception that gave the numeric value of the enum.
This was hardly of any use to a user.  This had already been done for the
parallel Thyra::ModelEvaluator software.

2) I added getMultiVector() and getMultiVectorOrientation() to
EpetraExt::ModelEvaluator::Derivative to eliminate the need to call
getDerivativeMultiVector() first.  I also added the multi-vector constructor
from DerivativeMultiVector to the Derivative class to make it easier to
construct Derivative objects that wrap a multi-vector.  The reason that I
called the function getMultiVectorOrientation() instead of just
getOrientation() is that I wanted to make it clear that this orientation only
applied to a wrapped multi-vector and does not apply to a wrapped
Epetra_Operator.