This is a quite big check-in.  This is code supporting the creation of a new forward sensivitiy capability for Rythmos.
In general, as part of this work I have:

1) Updated and added a few utilties in Teuchos

2) Performed a few refactorings in Rythmos including moving some code from
Thyra to Rythmos

Here are the details:

Changes/additions to Teuchos:
-----------------------------

* Add a new templated functions Teuchos::as<T>(...) and Teuchos::asSafe<T>(...).
These will functions call a new traits class that will perform checked and
unchecked conversions.  This is the first step in plugging this hole in C++
development.  Later, I will add runtime checked conversions that we need and
more can be added later.

* Add a new macro called TEUCHOS_ASSERT(...) in the new file Teuchos_Assert.hpp
which just calls TEST_FOR_EXCEPT(...).  Sometimes it is just more natural to
express an assertion.

* Added a new utility function called Teuchos::setVerbosityLevelOption(...) that
sets a Teuchos::EVerbosityLevel option on a Teuchos::CommandLineProcessor
object as a fully validated enum object.

* Added a nonmember template function called Teuchos::get(...) that calls
Teuchos::getParameter(...) to get a parameter from a Teuchos::ParameterList
object.  There is no need for a long name like getParameter(...) when
get<Type>(paramList) is pretty clear I think.

* Added a function called Teuchos::getVerbosityLevelParameterValueName(...) that
converts from a Teuchos::EVerbosityLevel enum value into it's string
representation that is used in the 'VerboseObject' sublist.  This was used
in my Rythmos example/test problem.

Changes to Thyra:
-----------------

* Added optional cloning capability to Thyra::NonlinearSolverBase.  This is
very useful for the forward sensitivity stepper.

* Moved Thyra::TimeStepNewtonNonlinearSolver to Rythmos::TimeStepNonlinearSolver

* Moved Thyra::SingleResidSSDAEModelEvaluator to
Rythmos::SingleResidualModelEvaluator

* Added nonmember functions createLinearSolveStrategy(...) and
createPreconditioningStrategy(...) for Thyra::LinearSolverBuilderBase that
define the default argument value and I removed the default argument value
from the virtual functions.  These functions are not called by many of the
Stratimikos examples.

* Added Thyra:: namespace qualification to helper macros in
Thyra_ModelEvaluatorDelegatorBase.hpp to allow them to be used
outside of the Thyra namespace (i.e. in Rythmos)

* Added missing header include to
Thyra_DefaultClusteredSpmdProductVectorDecl.hpp

* Changed the Epetra/Thyra wrapper functions in Thyra_EpetrraThyraWrappers.hpp
to work with base interfaces Thyra::[Multi]Vector[Space]Base instead of the
SPMD-based interfaces.  This makes the interface a little easier to use for
most (not all) use cases.  I now perform the needed dynamic casts within the
functions Thyra::create_[Multi]Vector[Space](.) to make this work.  I have
also allowed the domain space for multi-vectors in Thyra::create_MultiVector()
to be null and it will be created internally.  Additionally, I have created a
function called Thyra::create_LocallyReplicatedVectorSpace() that creates this
domain space.  My justification for making this change was to make most use
cases simpler and to avoid clients seeing types that they should not have to
see in most cases.  Just see the simple_stratimikos_example.cpp code to see
how this is simplified.  I have already updated all of the code in Trilinos
that I could find in epetra/thyra, epetraext/thyra, stratimikos, meros,
amesos, anasazi, belos, rythmos, and moocho.  However, I don't know if there
is any code outside of Trilinos that uses these function directly that would
be affected by this.  The change is trivial to make in every case that I had
to update (and the compiler helps a lot) but people need to be aware that this
change has been made.  Note: This is an example of where static type checking
has gotten in the way of getting work done.  More and more I am willing to
live with replacing compile-time error checking with run-time error checking
if this sufficiently improves the usability of the code and does not result in
any real performance hit.  However, you have to write the run-time error
checking code with good feedback if you go this way.

* Some small other documentation updates.

Changes/additions to Rythmos:
-----------------------------

* Changed Rythmos::StepperBase::getStepStatus() to a const function.

* Added Rythmos::StepperBase::getModel() function and updated subclasses.

* Added Rythmos::StepperBase::acceptsModel() to allow some StepperBase
subclasses to refuse to accept a model in setModel(...).  This is
the correct behavior for the ForwardSensitivityStepper class.

* Reordered the functions in Rythmos::StepperBase to better approximate the
order in which they are called.

* Created a new concrete class Rythmos::TimeRange and used it in
Rythmos::InterpolationBufferBase::SetRange(...).  I also create
the function Rythmos::InterpolationBufferBase::getTimeRange() to
return a Rythmos::TimeRange object.  This time range is needed
in order to write solid preconditions for many of the functions
in the Rythmos::InterpolationBufferBase interface.  There is
a lot that needs to be pinned down in this interface and there
are ToDos that ask the questions that need to be answered.

* I reordered the functions in Rythmos::InterpolationBufferBase slightly to
group them together a littler better.

* I changed the names of a bunch of functions that started with upper case
letters to now start with lower case letters as per convention.  I have
included the script rythmos/refactoring/lowercase-function-names.2007050515.sh
to make this change automatically in other code.

* Add optional cloning capability to Rythmos::StepperBase.  This is very
useful for the forward sensitivity stepper.

* Add optional cloning capability to Rythmos::InterpolatorBase.  This is very
useful for the forward sensitivity stepper.

* Added override Rythmos::BackwardEuler::setInitialCondition(...) and changed
logic to allow initial condition to be set later.  This function validates the
space for x and gets the initial time t from the InArgs object and does not
just assume that t0 == 0.0.

* Changed behavior of Rythmos::BackwardEuler::setInterpolator(...) to require a
nonnull object on input.  It has been my experience that you should not allow
the input RCP to be null since this can hide bugs that are hard to track down.

* Added the private data member x_dot_ to Rythmos::BackwardEulerStepper and used
this to assign solutionDot in getStepperStatus() and getPoints(...).  I think
that every stepper needs to maintain x_dot.

* Created a new class Rythmos::FowardSensitivityStepper that directs the
integration of a state equation and sensitivities at the same time.

* Created a new class Rytymos::ForwardSensitivityModelEvaluator.  This class is
used by Rythmos::FowardSensitivityStepper to represent just the forward
sensitivity equations.

* Created a new interface Rythmos::SingleResidualModelEvaluatorBase and
derived Rythmos::SingleResidualModelEvaluator from it.  This is
the interoperability interface.

Other changes:
--------------

* I updated the token-replace-list-r script to automatically get the
directory it lives in so that it can call token-replace-list.pl without
needing any help from the path.  I will make this same change
in other scripts as well later.

* Added standard test harness targets and export makefile install features to
sacado/Makefile.am

* Added comment to sacado/test/definition file that tests will only be
built and run if CppUnit is enabled