When RCB_OUTPUT_LEVEL and RIB_OUTPUT_LEVEL are "1" or greater, extra analysis is done and printed out in Zoltan_RB_stats.  I added two lines, which correspond to the "FINAL_OUTPUT=1" statistics in graph and hypergraph partitioning.
 STATS Runs 1  bal  CURRENT 1.000000  MAX 1.000000  MIN 1.000000  AVG 1.000000
 STATS Runs 1  moveCnt CURRENT 22.000000  MAX 22.000000  MIN 22.000000  AVG 22.000000

The first line does the same calculation that FINAL_OUTPUT=1 does to
calculate how balanced the partitioning is.  The second line counts
the objects that must be moved in the new partitioning, calling the
ZOLTAN_OBJ_SIZE_FN (if it has been defined) to weight the counts by
the sizes of the objects.

The other statistics printed out when FINAL_OUTPUT=1 involve object
connectivity (cutn, cutl) which we don't have in the geometric methods,
so they do not appear.