Added Makefile.export support to Pliris.  Now that Epetra has an optional dependence on Teuchos, Pliris tests were failing to compile during the nightly builds because of not being able to find the teuchos library.
For now the Makefile.export support assumes use of the double precision
pliris library because it allows us to follow the macro naming convention
used in Makefile.export.  Support could be expanded to cover the other
pliris libraries as well.

Existing Pliris users should not be affected by this change.  However,
if users were compiling the teuchos library, but not linking to it for
exe's that use pliris, they will run into a link error.  That error is
the result of the new optional epetra dependence on teuchos, not a result
of this commit (and is exactly the same problem that the test harness was
having).  Using Makefile.export would make future similar dependency
changes transparent to the user.

I built pliris successfully with just pliris enabled, and with pliris and
epetra enabled, and with pliris, epetra, and teuchos enabled.