SUMMARY:  Added partial-clean target
BUGS:     3545

CHECKLIST:

1) Did you add any source files? No

2) Did you update or change the build system? Yes
2.a) Did you remember to do a bootstrap and do a 'cvs add' on the
     Makefile.am files and the Makefile.in files? Yes
2.b) [Optional but recommended] Did you blow away the libraries and
     executables and then rebuild the libraries? Not needed

3) Did you rerun the test suite on at least the packages that you
   changed? Yes

COMMENTS:

I added the top-level target:

    partial-clean:
        find . -name "*.a" -exec rm {} \;
        find . -name "*.exe" -exec rm {} \;
        find . -name "*.xml" -exec rm {} \;

What this target does is to allow you to more quickly test the rebuild of
libraries and tests/examples by not deleting *.o files.  To retest a (serial)
build of Trilinos just do:

    make partial-clean
    ./do-configure # Or whatever you call your configure script
    make -j6 everything
    make runtests-serial TRILINOS_TEST_CATEGORY=FRAMEWORK

I my experience, this catches 99.9% of the errors that are caught by blowing
away the build directory and then configuring and building from scratch.