SUMMARY:  Removed build warnings from tests for NaN/Inf detection.
     BUGS:   3562

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

To test the NaN/Inf detection of the NOX_FiniteValue object, we need to generate nans and infs.  Some compilers check for numbers divided by zero and print a compiler warning.  To eliminate this warning, the Teuchos::ScalarTraits object was used.  The scalar traits uses a trick to eliminate the compiler warning - it returns the zero denominator via a function call so the compiler can't figure out it is a divide by zero (it only sees a function).


(checklist version 2007-06-20)