SUMMARY:   Added error checking to the preconditioners in LinearSystemAztecOO.
     BUGS:   729 (that's correct - 729)

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

A long time ago (see bug #) Ross pointed out that we were not checking the error codes returned from Ifpack.  This commit fixes that oversight.  We now check the error codes when building, recomputing, and applying preconditioners.  Checks are in place where appropriate for both ML and Ifpack.  If a non-zero error code is returned, the default handling is to throw an exception.  However, I seem to recall runs in Charon where the preconditioner failed but the linear solve still succeeded.  To allow for this case, you can set a parameter in the linear solver sublist that will allow the run to continue, but only print a warning if warnings are enabled.  The parameter is described below (from the doxygen documentation for LinearSystemAztecOO object):

"Throw Error on Prec Failure" - boolean.  If set to true, an exception will be thrown if the preconditioner fails to initialize or recompute/refactor.  If set to false, a warning will br printed if the NOX::Utils::Warning is enabled in the printing utilities (NOX::Utils).  Defaults to true.

Amazingly, all nox tests are still passing...


(checklist version 2007-06-20)