SUMMARY:   [Teuchos_TypeNameTraits not robust when using __cxa_abi_demangle]
     BUGS:      [3604]

     CHECKLIST:

     1)   Did you add any source files?  [no]

     2)   Did you update or change the build system?  [no]

     3) Did you rerun the test suite on at least the packages that you
        changed?  [yes]

     COMMENTS:
Teuchos_TypeNameTraits wasn't checking the return-value of __cxa_abi_demangle to see if it is NULL, and wasn't checking the output status value. This change checks for non-zero status and also for NULL return-value. If either of those conditions occurs, return the string "<demangle-failed>" rather than attempting to reference the return-value from the demangle function.

I've run the Teuchos tests as well as the moocho tests. The moocho tests were failing due to this bug, and they're now passing. On my linux-g++ platform, the demangle function is working correctly most of the time, but for at least some type-names it is failing to demangle successfully.
I will also make this change to the 8.0 branch.
     [???]

(checklist version 2007-06-20)