SUMMARY: Fix non-debug build for new Teuchos::Array class.
BUGS: None filed

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I only rebuilt
   the one test that depends on this code.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, just this single test

COMMENTS:

I noticed that the Trilinos test harness reported failed builds for the the
test Array_test.cpp.  I realized that I did not actually test with
TEUCHOS_DEBUG not defined.  I was able to reproduce the compilation failure
and this checkin fixes this.  This is the problem with ifdefed code, you have
to actually build with all of the branches of the ifdef!