SUMMARY: More progress on Rythmos time discretization code and cleaned up a few warnings given -Wshadow -Woverloaded-virtual
BUGS: None filed

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, 815 out of 816 passed.
   The one failing anasazi test fails also without these changes
   so I am checking this code in (see bug 3723)

COMMENTS:

The main thing that I did here was to move the backward Euler time
discretization code along.  The code now runs and seems to solve the system
but I have not added a test to verify this yet.  There is still some more work
to be done yet.  I have not checked in a test for this but there is a
prototype test on gabriel.sandia.gov at
/home/rabartl/tests/Rythmos/20071121/test1 that will be cheked in as soon as
we get the verification test put in.

All of the non-rythmos files contain changes to remove shadowed identifiers
that are flagged by the g++ compiler option -Wshadow.  I have a *lot* more
code to clean up before all of these warnings go away in the packages that I
develop!  This was requested by Richard Drake since Alegra is starting to use
this flag now.