SUMMARY:   Removed shadowed variable warnings from Belos code
     BUGS:      3727

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     Using the compiler flags stated in bug 3727, a lot of warnings started
cropping up for Belos.  Most of these were accidental recreation of local
variables, and they have been taken care of.  There was some unintentional
hiding of methods due to the use of the [CG/Gmres]Iteration classes and those
have been taken care of as well.

The Belos tests pass in serial and parallel.