SUMMARY: Fixing Teuchos::Array and Teuchos::ArrayView to compile on Sun compiler on sass9000
BUGS: 3731

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (including adding
   files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes, but there where a few files that where changed.

4) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes and no (see below)

COMMENTS:

Here I have modified the new Teuchos::Array and Teuchos::ArrayView classes to
compile on the Sun compiler (CC: Sun C++ 5.7 Patch 117830-07 2006/03/15) on
sass9000.  This port is more proof that this is still not an ISO 98 C++
complient compiler yet (as if such a thing actually existed but we should
expect better).

Specific changes:
-----------------

Teuchos_Array.hpp: Moved the declaration and defintion of the friend functions
outside of the class (only needed for the Sun).

Tuechos_ArrayView.hpp: Ifdefed out (for __sun) one of the createVector(...)
overloads.  The Sun compiler says that the call in the test is ambiguous but
g++ 3.4.6 disagrees.

ArrayRCP_test.cpp: Ifdefed out test for count.  The Sun compiler does not
delete C++ objects on the stack when it is supposed to (i.e. the reference
count is off).  The issue of in-class definition of member function templates
is no big deal but it does result in a *lot* of extra code that you have to
write (just do a diff to see this).

teuchos/test/definition: Added the test ArrayView_test.exe to the test
definition file.  Amazingly, I had forgotten to do this!

Tests built and run:
--------------------

Here I only built and tested Teuchos on the Sun.  It would take a full day to
build and test all of Trilinos on this machine. (and why are people using this
platform again?)

However, I did do a pretty complete build of Trilinos on the Linux/gcc machine
sais504 and all of the tests passed! (except for the same Anasazi test and a
Rythmos test (see bug 3732))