SUMMARY: Adding files needed to build SIERRA version of Trilinos
BUGS: None filed

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? Yes
1.a) Did you do a 'cvs add'? No, already added
1.b) Did you add them to a Makefile.am file in some way (either as
     a header or soruce file or to EXTRA_DIST)? Yes
1.c) Did you do a new bootstrap? Yes

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? No, I only rebuilt
   and installed thyra and epetraext

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, this does not affect the
   test suite.

COMMENTS:

Here I added the example file sillyModifiedGramSchmidt.hpp to be installed
since it is actually used in one place in library code.  This is a bad
practice that needs to be changed.

I also added the file EpetraExt_Permutation.cpp to the list of header files in
EpetraExt so that they get installed.  Amazingly, this file actually has
include guards and gets included in the file EpetraExt_Permuation.h!  Am I
missing something here?  So be it ...