Bug 2826.
Backing out most of the changes I made to Isorropia for this bug.  Some
errors showed up when compiling PyTrilinos.  The errors have to do with
Isorropia not defining BLAS_LIBS and LAPACK_LIBS anymore.  In the long
run, Isorropia should not have to define these values, but for now I will
back out the change to fix the errors we are seeing.