SUMMARY:   Bug fix for repartitioning (from Karen).
     BUGS:      No number?

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

From Karen:

The second bug is major and could possibly be the cause of the problems Chris saw.  Basically, I had an off-by-one error in the distribution of repartition edges when I augmented the hypergraph.  (I hate off-by-one errors.)  I attach my revisions of phg/phg_build.c, version 1.93.  Unfortunately, I don't know whether or not I was working with the most recent version of phg_build.c, and without CVS or access to software.sandia.gov, I have no way to tell.  But I think the changes should apply no matter the version, as no one else on the team was hunting for this problem.  All the changes are in Zoltan_PHG_Add_Repart_Data and macro ProcForRepart.


(checklist version 2007-06-20)