SUMMARY:   Added Epetra_Comm::Broadcast for character arrays.
     BUGS:      3877

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

Epetra_Comm contains broadcast methods for ints, longs, and doubles.  When
reading a matrix file on processor zero and sending the data to other
processors, it is convenient to have a broadcast for characters read from the
file.  Adding the new method is simple within Epetra, requiring only some
copy/paste and small modifications.

I added the method to Epetra_Comm as well as the MPI, MPISmp, and Serial
implementation.  I added tests of the new method to
epetra/test/Comm/cxx_main.cpp.


(checklist version 2007-06-20)