SUMMARY: Fixed adjoint solves with Stratimikos/AztecOO/Ifpack and added test
BUGS: 3919

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? Yes
1.a) Did you do a 'cvs add'? Yes
1.b) Did you add them to a Makefile.am file in some way (either as
     a header or soruce file or to EXTRA_DIST)? Yes
1.c) Did you do a new bootstrap? Yes

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, of the 889 tests run
   only the one sacado RAD test failed (see below).

COMMENTS:

Here I have added a faily strong test for Stratimikos/AztecOO/Ifpack.  In the
process, I have upgraded the test driver test_single_stratimikos.cpp a good
bit as it now calls Thyra::LinearOpWithSolveTester.  I also upgraded
Thyra::LinearOpWithSolverTester to accept a PL through the
ParameterListAccepter interface.

There was a bug in Thyra::AztecOOLinearOpWithSolveFactory
w.r.t. preconditioners and adjoints.  I changed the behavior so that it just
does not set an adjoint solver if the preconditioner does not support an
ajoint and the test failed gracefully with a night error message.  I then
turned on support in Thyra::IfpackPreconditionerFactory for adjoint support
and the test passed.

The test file _FourByFour.aztecoo.ifpack.adjoint.xml is set up to verify that
the adjoint solve only takes on iteration since it creates a perfect
preconditioner.  This is a pretty good test that the preconditioner is being
used for the adjoint and that it is being applied correctly.  This is a good
start but Stratimikos needs more tests.

Failing tests:
--------------

The same know failing test:

    sacado - trad_dfad_example.exe...                      ! FAILED     <1 second

is failing is due to the problem reported in bug 3922.