SUMMARY:   Commented test of number of removed edges and error message                 if too many removed.
     BUGS:      working on 3965

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  no; this commit un-did a previous one.

     COMMENTS:

I think there are problems with this test.
(1) it doesn't work for Zoltan_LB_Eval, which "removes" all edges
    by forcing esize_threshold=0.
(2) it doesn't work when gnVtx < zz->Num_Proc, even if no edges are
    removed.
(3) I don't understand why it subtracts the global number of removed
    *edges* from the global number of *vertices*.

For now, I'll comment out this test; we can add it back if we decide
it is needed and if we can correct the problems.


(checklist version 2007-06-20)