SUMMARY:   Changed domain maps set by the redistributor
     BUGS:      3962

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes, passed.

     COMMENTS:

We now set the domain map of the balanced graph or matrix more cleverly according to these rules:

a) if the old domain map was same as the old range map, preserve this property,
b) otherwise, set new domain map to be same as the old domain map.

Exception: If the input was a row matrix, we don't know the old range and domain maps, so we create a linear domain map if the matrix is non-square.



(checklist version 2007-06-20)