SUMMARY: Fixing problem with Array<MPI_Request> once and for all (again)
BUGS: 4016

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No

3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes, both serial debug
   and mpi optimized

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? Yes, all 924 serial debug and
   696 mpi optimized tests run passed

COMMENTS:

This is a more substantial fix to the problems with forward declared but
undefined types being used with the Teuchos memory management classes.  Again,
this is a big problem with MPI types with many implementations that forward
declare a struct but don't define it.  This is C's version of encapsulation
but it has some nasty consequences when this is mixed with C++ and RTTI with
the typeid(...) operator.

Here is the deal.  You can not call typeid(T) or typeid(t) where T is some
forward declared but undefined type and p is an object of type T&.

I made a few small tweeks in several files to deal with the fundamental
problem and I added some tests to check the behavior and to document the
issues.

Here is the outline for the solution/changes:

(*) Teuchos::TypeNameTraits::getConcrete(...) has been added to allow a client
to request the type name of a concrete objects t2 of type T.  The default
implementation just uses typeid(...).  This traits function is called by
Teuchos::RCP and Teuchos::ArrayRCP and really it is only neeed to generate the
information needed to track what RCP nodes are currently in play.  I had to
add getConcrete() to a bunch of existing TNT specializations but that is not
hard.

(*) I removed the TNT specialization for std::vector since it required that
the underlying type have a TNT specialization.  This allows you to use
typeid(...) on std::vector<Undefinded*> without it's embedded type Undefined
being defined at all!

(*) For all of the error messages that I could, I replaced
"SomeType<"<<TNT<T>::name()<<>" with typeid(*this).  This allows you to create
and use an Array<Undefined*> object in debug mode with full error checking and
good error messages.  However, you can't create views as ArrayView objects in
debug mode because of the dependance on ArrayRCP internally.

With this, I can keep using Teuchos::Array<MPI_Request> inside of
Teuchos::MpiComm.

I added some test classes to for undefined types and some test cases to show
how you can embedd pointers to undefined types in RCP objects as long as you
define a specialized deallocator object.

Some other changes to mention:

(*) I removed some strange const_cast<...>(...) calls from Teuchos_RCP.hpp.
These many have been needed on some older compiler.  I hope this will not
break any of the ports ...

(*) Updated CMake build system and added UnitTest stuff and new Comm tests