SUMMARY: Finishing port of SIERRA + Trilinos 9.0 to sun 5.7 on sass8000
BUGS: 3978

CHECKLIST:

1) Did you add any source or other files needed for building, testing,
   or documentation? No

2) Did you update or change the build system (other than just adding
   source files)? No
3) Did you do a cvs update *very* recently before rebuilding and
   retesting? Yes

4) If justified, did you type 'make partial-clean' and then rebuild all
   libraries and executables with 'make everything'? Yes and no, I only
   rebuilt Salinas + Trilinos 9.0 (dev without a bunch of packages)
   with the sun 5.7 compiler.

5) Did you rerun the test suite on at least the packages that you
   changed (rerunning the entire Trlinos test suite on all affected
   packages is *strongly* recommended)? No, I just built
   the Salinas + Trilinos 9.0 executable with the sun 5.7 compiler,
   I did not run it

COMMENTS:

Note, this port only includes just enough of Thyra and other software to
support the quadratic eigensolver in Anasazi used by Salinas, nothing more.
However, this will represent actually the first time that thyra will be a
critical capability in a production application code that the APP developers
care about.

The main thing that I had to do where was to put in a chain of using
declarations for Thyra::LinearOpBase::apply(...) and all of its virtual
overloads.  I just did not have time to fix this problem right by getting rid
of the overloaded functions.  I started to this but I realized that with me
going on vacation Saturday, that I just did not have time to fix this right
and update everything.

The other irritating thing that I had to fix was that the sun 5.7 compiler
refusing to perform implicit conversions on argument to templated functions
even when the template arguments are fully specified.  With the sun 5.7
compiler, you have to have an exact match for all of the arguments passed into
a template function call.  Not the end of the world, but it does make it hard
to write portable code when you have a compiler that does (or does not) do
this sort of thing.

The last issue that I had to fix was a problem with the Teuchos LAPACK wrapper
to the function GEES(...).  This function takes a function pointer to a
user-defined function which is just passed in as null.  The sun 5.7 compiler
complained that the function pointer itself must be declared as extern "C"
since the function itself is extern "C".  Therefore, I created a type for the
function pointer in extern "C" and the compiler was happy with this.

The diff tells it all:

---------------------------------------------------------------------------

RCS file: /space/CVS/Trilinos/packages/teuchos/src/Teuchos_LAPACK.hpp,v
retrieving revision 1.55
diff -u -r1.55 Teuchos_LAPACK.hpp
--- packages/teuchos/src/Teuchos_LAPACK.hpp     17 Jul 2008 22:10:05 -0000      1.55
+++ packages/teuchos/src/Teuchos_LAPACK.hpp     7 Aug 2008 19:42:22 -0000
@@ -87,6 +87,16 @@
        </ol>
 */

+
+extern "C" {
+
+
+typedef int (*gees_nullfptr_t)(double*,double*);
+
+
+} // extern "C"
+
+
 namespace Teuchos
 {

@@ -1424,7 +1434,8 @@
   template<typename OrdinalType>
   void LAPACK<OrdinalType, double>::GEES(const char JOBVS, const OrdinalType n, double* A, const OrdinalType lda, OrdinalType* sdim, double* WR, double* WI, double* VS, const OrdinalType ldvs, double* WORK, const OrdinalType lwork, double* RWORK, OrdinalType* BWORK, OrdinalType* info) const
   {
-    OrdinalType (*nullfptr)(double*,double*) = NULL;
+    //OrdinalType (*nullfptr)(double*,double*) = NULL;
+    gees_nullfptr_t nullfptr = 0;
     const char sort = 'N';
     DGEES_F77(CHAR_MACRO(JOBVS), CHAR_MACRO(sort), nullfptr, &n, A, &lda, sdim, WR, WI, VS, &ldvs, WORK, &lwork, BWORK, info);
   }

---------------------------------------------------------------------------

This should effectively close bug 3978 which I will do so tomorrow after I see
all of the nightly tests passing.