Summary:  Adding a file where auxiliary functions operating on serial dense matrices           and vectors can be placed.
Bug #:  None

This check-in includes a new file that contains helper functions that operate
on the serial dense matrix and vector classes.  The first function included in
this file is a matrix triple-product, as requested by Mike Eldred and Bill Bohnhoff.

This helper function has been added to examples and tests, to ensure that its
functionality is correct.