Summary: More enhancements to TRILINOS_ADD_TEST(...) and more
Here is what I did:

(*) Add a global user option VERBOSE_CONFIGURE that is OFF by default.  When
turned on, it results in several MESSAGE(...) statements being executed
showing you what examples and tests are being added.  This is good for
debugging.

(*) Appended the number of processes being used to the name of the MPI test

(*) Avoid appending a counter to the test name if there is only one argument

(*) Added quotes to the lists of command-line arguments so that they got used
together.  I was creating a new test for each command-line argument which is
not what I want.  The BadUnitTest tests show an example of this.

(*) Updated linux-gabriel-gcc-3.3.4-mpi-cmake to allow extra cmake options to
be passed in on the command-line.

I also manually tested that the PASS_REGULAR_EXPRESSION option works correctly
and it seems to.  I am a little worried that we don't have unit tests in place
for TRILINOS_ADD_TEST(...).

I tested both serial and mpi builds and all of the tests pass.