SUMMARY:   Merging back changes to BKS solver manager regarding correct eigenpair returns
     BUGS:      4247

     CHECKLIST:

     1)   Did you add any source files?  No.

     2)   Did you update or change the build system?  No.

     3) Did you rerun the test suite on at least the packages that you
        changed?  Yes.

     COMMENTS:

     It has been documented that BKS will return the wrong eigenpairs because
it assumes the converged ones have to be the principal Ritz pairs.  To be
correct BKS should use the converged indices when determining which eigenpairs
to return to the user.  This behavior has been changed and the test Chris Baker
entered into Anasazi that exibited this behavior now passes.

(checklist version 2007-06-20)