Summary: Adding package name prefix to all executables to CMake system
Here the main change is to add the package name as a prefix to all test and
example executables created by the CMake build system.  There are two drivers
for this.  First, target names must be unique across all of Trilinos and we
already had a few conflicits like VerySimple with Epetra and AztecOO and
Comm_test with Teuchos and Epetra.  Second, in MS Visual C++, all of the
targets for all of Trilinos are listed out in one big flat list sorted
alphabetically.  Therefore, the tests and the examples for all of the packages
were mixed together and you could not tell which tests/examples went to each
package.

However, I did not want to go into all of the autotools Makefile.am files and
add the package prefix.  Therefore, I had to make the addition of the prefix
automatic.

To make this work, I modified the Perl testing script 'runtests' to
automatically append the package name to the name of the test specified in the
PACKAGE/test/definition file if the raw test does not exist.  The package name
is now added as the first line in every PACKAGE/test/defintion file.

When using the autotools system to build the tests and executables, the
'runtests' scripts will look for the unprefixed test name specified in the
PACKAGE/test/definition file.  If it exists (and therefore has been built),
then it just runs it.  Otherwise, it looks for the prefixed test executables
(which will not exist) and does not run anything.

When using the Cmake to build the tests and executables, the 'runtests' script
will not find the unprefixed test executables.  It will then look for the
prefixed test executables and it will find and run any that have been built.

When the DIRS option is used in either system, it just picks up all executable
files and runs them.  When built with autotools, the unprefixed test is found
and run.  When built with CMake, the prefixed test is found and run.

I have tested this with:

* Cmake system with serial debug

* CMake system with mpi opt

* Autools system with serial debug

I have not built all of the packages affected but I am confident that since I
have tested the 'runtests' script for all of the use cases described above,
this will work for all the other Trilinos packages as well.

Misc:

(*) CMakeLists.txt: Added placeholder for framework level support for TPLs.
This is not working yet.

(*) cmake/TrilinosCMakeQuickstart.txt: More use cases, updated some

(*) cmake/TrilinosPackageLibraryHelpers.cmake,
cmake/Trilinos_Add_Executable.cmake, cmake/Trilinos_Add_Test.cmake,
packages/zoltan/*: Updated for new prefix/nonprefix exec system.

(*) packages/teuchos/example/ExplicitInstantiation/four_files/CMakeLists.txt:
Removed explicit Teuchos_ prefix.

(*) sampleScripts/linux-gabriel-gcc-3.3.4-dbg-cmake: Added
CMAKE_INSTALL_PREFIX to test install