SUMMARY:   Merged release branch with main trunk.
     BUGS:      NA

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system? no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

Merged release branch with main trunk.  All tests pass except
scotch partitioning:

     Test problem simple with method graph-ptscotch FAILED

This test also fails in the release branch; it looks as if the partition
is identical between the computed answer and the repository answer, but
part numbers one and two are swapped.  If this result is consistent
(at least on octopi), we should update the repository answers.


(checklist version 2007-06-20)