Summary: Fixing variable handling code with rebuilds and TPLs
Bugs: 4239

Here I think that I have fixed two serious defects which how the CMake
build system for Trilinos was handling varaibles.  One of these
defects was causing complete rebuilds of the source code every time do
do a configure.  The second defect related to how TPL include
directories were handled.

Specific changes:

(*) cmake/TPLDeclareLibraries.cmake: Fixed the behavior of the
function TPL_DELCARE_LIBRARIES(...) so that it deals with cache
varaibles correctly.  The old behavior was actually setting the cache
variable TPL_${TPL_NAME}_INCLUDE_DIRS on one configure and then
nulling it out on the next!  This would go on and on and it was
driving me crazy.  I think this is completely fixed now.

(*) cmake/TrilinosGlobalHelpers.cmake: Added ASSERT_DEFINED(...) calls
for the TPL exported cache variables to make sure that they are
defined correctly.

(*) cmake/TrilinosSetupStrongCompileWarnings.cmake: Changed the
behavior to only modify the CMAKE_[C,CXX]_FLAGS_${BUILD_TYPE}
varaibles as non-global variables in the parent scope and does not
overwrite the cache varaible.  This was causing the
CMAKE_[C,CXX]_FLAGS_${BUILD_TYPE} cache varaibles to get appended
every time resulting in CMake thinking that it needed to rebuild *all*
of the source files after every reconfigure where the file
CMakeCache.txt was not deleted first.  This is why I did not see the
problem because I was deleteing the CMakeCache.txst file every time
usually.

Tested on godel.sandia.gov with:

- CMake serial debug from scratch (all 228 tests passed)

- CMake mpi release from scratch (all 209 tests passed)