Fix and issue with the intrepid cmake files and the amesos umfpack option
The default for using Amesos with the umfpack option was using the value
given in the variable Amesos_ENABLE_UMFPACK, which does not exist if Amesos
is not enabled. I put in a check that if the Amesos enable is on AND exists
then the intrepid enable is on, otherwise the Intrepid enable is off.