Some RCP work, partial addressing Thyra vector space ownership issue, and MOOCHO QPSchur tests
Bugs: 4388

This checkin contains several sets of changes:


1) Teuchos: Adding nonnull(...) function to test for non-null
memory-management objects and adding [Array]RCP::create_strong() function
--------------------------------------------------------------------------------

First, I added nonmember functions nonnull(...) that return bool for RCP, Ptr,
ArrayRCP, ArrayView, and boost::shared_ptr objects.  This is an interm
solution to the problem of testing for non-null in if statements that Scott
Collis requested.

Second, I added the functions [Array]RCP::create_strong().  These are needed
to convert from a weak [Array]RCP object back into a storng [Array]RCP object.
This is needed to implement ownership of self which is now being used in
Thyra.


2) Progress in making VectorBase objects truly own VectorSpaceBase objects
(bug 4388)
--------------------------------------------------------------------------------

Here, I refactored the concrete Thyra::DefaultSpmdVectorSpace class to contain
a weak RCP to itself.  This is created in a new static create() function which
is used by the existing non-member constructors.  The
Thyra::DefaultSpmdVectorSpace object then creates strong RCPs to pass to
created Vector and MultiVector objects.  This solves the problem with
VectorSpace lifetime reported in bug 4388.  A consequence of this change was a
removal of all of the public constructor functions.  You now have to call on
the provided non-member constructor functions.  I had to make some small
changes to the Thyra::VectorSpace handle class to accommodate this.  All of
the test of the changes that you see are replacing direct calls to the
constrctors with the non-member constructor.


3) Adding some QPSchur tests to MOOCHO
--------------------------------------

I added some QPSchur tests because I saw that QPSchur was got getting tested
at all in the existing test suite?  The coverage for MOOCHO is very bad right
now (50%) and this will help some but there is a lot of work to do.


4) Misc:
--------

(*) cmake/python/UnitTests/TrilinosPackageDependencies.gold.xml: Updated for
addition of TPL dependencies on UMFPACK and AMD for EpetraExt.



Testing on gabriel.sandia.gov:
------------------------------

(*) CMake serial debug: All packages that depend on Thyra: All 465 tests ran
passed.

(*) CMake mpi release: All packages that depend on Thyra: All 400 tests ran
passed.