Turning off framework unit tests when development mode is offcmake
This avoids the problem of these unit tests failing when packages are removed in a release that does not contain experimental code but the unit test must to check correct logic.  I don't see a sufficient reason to run test tests in a release of Trilinos.  We are not going to be making lots of changes to a release of Trilinos anymore once the branch is created so we don't need to worry about these tests.