Addressing configure problem with Sacado
Bugs: 4420

Here I have addressed a problem with Sacado that was causing other packages to
fail to configure (see bug 4420).  The problem relates to the behavior of
keeping the CMakeCache.txt file and disable Sacado examples.

To fix this, I just removed the dependence on Sacado_ENABLE_EXAMPLES.  Now
Sacado_ENABLE_FEApp will be turned on when Epetra and EpetraExt are both
enabled but will not have any affect unless Sacado_ENABLE_EXAMPLES so
everything should be fine.

I tested this by configuring first with:

  $ ./do-configure.package -DTrilinos_ENABLE_NOX=ON -DTrilinos_ENABLE_Sacado=ON

and then with:

  $ ./do-configure.package -DTrilinos_ENABLE_NOX=ON -DTrilinos_ENABLE_Sacado=

This demonstrated the problem and then I fixed it.