----------------------------------------------------------------------  Modified Files:  	PDP2010.pdf PDP2010.tex  ----------------------------------------------------------------------
I modified the paper to remove author-related information for blind review.

I also modified the title and text to make it more explicit that we are *not*
in the API business, but only want to have a minimal interface for writing our linear
algebra software.

I will submit this version to PDP2010.