SUMMARY:   When building GRAPH from graph callbacks for PHG,                 got the edge weights correct for non-symmetric graphs.
     BUGS:      4490

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  small tests; suite beginning

     COMMENTS:

In PHG doing GRAPH partitioning...

When no edge weights are specified, symmetric and nonsymmetric graphs
had uniform edge weights; nonsymmetry was not indicated in the edge weights.
I modified the code to make edge weights reflect nonsymmetry.  That is,
if no edge weights are specified and both A->B and B->A exists, the
hyperedge {A,B} receives weight 2; if only edge A->B exists,
the hyperedge {A,B} receives weight 1.

I also changed the default value of PHG_EDGE_WEIGHT_OPERATION to depend
on the model being used.  In the GRAPH model, we want PHG_EDGE_WEIGHT_OPERATION
to default to "sum".  In the HYPERGRAPH model, we want it to default to "max".

(checklist version 2007-06-20)