SUMMARY:   Updated test and answer file
     BUGS:      4490

     CHECKLIST:

     1)   Did you add any source files?  no

     2)   Did you update or change the build system?  no

     3) Did you rerun the test suite on at least the packages that you
        changed?  yes

     COMMENTS:

Changes to the way hyperedge weights are computed from graph callbacks
for GRAPH partitioning in PHG caused this problem to fail; the edge weights
were different relative to the weights on the migration edges.
By setting PHG_REPART_MULTIPLIER to 50, I get back to the same answers as
before.

(checklist version 2007-06-20)