This is a major reorganization of the TSFExtended files to better support explicit instantiation and improve compile times.
We are also in the process of reducing the dependence on Thyra to its
fundamental vector/operator interoperability interfaces only. This is to
simplify the code and facilitate debugging. This is complete for operators.
We will do vector spaces and vectors soon.

All tests pass, with and without explicit instantiation.