Summary: Verify Fortran/C/C++ compiler combination
When Trilinos_ENABLE_Fortran is enabled we now verify that the Fortran
and C++ compilers selected by the user can at least link a simple test
project.  This may save users the trouble of attempting a full build
that will not link at the end due to a bad combination of compiler
versions.

(*) cmake/TrilinosFortranMangling.cmake: If the CMake version is recent
enough, include the FortranCInterface module and use its VERIFY method.