SUMMARY:  Adding new status test support for output classes.
     BUGZILLA BUG #(s):  None.

     CHECKLIST
     ---------

     1) Did you 'cvs add' all new files and directories [Y:N:N/A]?  Y

     2) Did you add new source files to a CMakeLists.txt file [Y:N:N/A]?  Y

     3)  Did you modify or add any source or build system files?  Y
     3a) Did you rebuild the libraries and executables?  Y
     3b) Did you rerun the test suite on the packages that you changed and
         all packages that depend on the modified packages?  Y
     3c) Did you change the required or optional dependencies for any
         package?  N


     DESCRIPTION OF CHANGES:

     While developing a simpler outputting class, I created a new StatusTestOutput
base class for status tests that will manage output from a collection of status
tests.  The current StatusTestOutput class that is overly verbose is renamed
the StatusTestGeneralOutput class.  It can handle any status test class and
requires no knowledge of what status tests are being used.
     All the solver managers have been modified to use the new StatusTestGeneralOutput
class, resulting in the same behavior as before.  The BelosStatusTestSimpleOutput class
is a new class derived from the StatusTestOutput class that will be an option from any
current Belos solver.  It gives much simpler output, but it requires a
StatusTestMaxIters object and a single or combo BelosStatusTestResNorm object.

(checklist version 2009-07-14)