Efficiency improvements in NOX for MatrixFree Mode.
Made changes to save 3 MatVecs per Newton iteration
when running NOX in Matrix-Free Mode:
(1) Added check in MatrixFree::Apply(X,Y) method to
return Y=0 when Norm(X)=0. This is typical for the
first Krylov step.
(2) Added hook to NOX::Epetra::Group to disable the
calculation of the linearResid after every Newton.
This was needed by inexact Newton but now only seems
to be used by Modified Newton.
(3) Added a hook to MatrixFree to get a Ptr to the
NOX Solver, so it can get the Residual vector from
the group as the unperturbed base case for the
directional derivative, instead of calling computeF
in the interface directly.

Only (1) changes the default behavior of NOX.

Added 1DfemLeanMatrixFree.C test problem, and
verified that the above changes save 3 calls
to computeF() for every Newton iter.