Fixing behavior of advanced RCP node tracing tests
These tests were failing on the trilinos-test machine for an MPI_RELEASE
build.  The reason is that the logic in the CMakeLists.txt file did not check
that debug support was enabled and there is no node tracing in this case.

I was able to do a SERIAL_RELEASE build and verify that the tests failed.  I
then changed the CMakeLists.txt file and got the tests to pass (by leaving off
these two extra tests that should only be anbled when boost and debug are
enabled).

I tested tihs manually and also using the checkin-test.py script (PS test
results shown below).