Testing changes to pull from a nightly clone
Made changes to the testing system to pull from a nightly clone. This clone is
pulled near 11pm every night (as the cdash website says) and all of the testing
will pull from that clone. This will ensure that all testing is done off of the
same copy of the code. Since git doesn't support checking out at a specific time
this is the only way to ensure that the nightly testing is using a consistent
repository for all builds.

Also changed a path in the godel cron script for where to get eg. I thought I
had made sure that was the right path, but I guess I missed that before.