Added support for exporting CMake info to client packages
(1) Added export_library_dependencies to Trilinos/CMakeLists.txt. This is
a standard cmake function that lets
client projects refer automatically to the TPLs used by Trilinos.

(2) Added code to TrilinosConfig.cmake.in that will export the names
of the compilers used to build Trilinos. Note that this does not
commit the client to using the same compilers.