Changed PyTrilinos test scripts from optparse module to getopt module
For a long time, I have been seeing the following error pop up:

  python(65587) malloc: *** error for object 0x1715674: Non-aligned pointer being freed

As is typical for a memory allocation bug, it was hard to track down.
Older compilers did not show it (gcc 4.0.1 did not, but most newer
ones did).  Debugging indicated that the error was always raised
within an Epetra Print() method, most probably at a call to

  operator<<(ostream, int).

This did not seem to be the likely culprit, as it *should* be a
dependable built-in function.  Most likely, something was happening
elsewhere, and then getting triggered in operator<<().  As I tried to
isolate the issue in a simpler test script, I came across the
following: I could not recreate the problem in a python script without
doing an "import optparse" and I could eliminate the error from the
larger test scripts by getting rid of "import optparse".

This commit changes the use of module optparse for handling
command-line arguments to the use of the older getopt module.  I am
considering reporting this as a bug to python, but getting them to be
able to recreate it will be a chore.