Clean up refactoring; fixing variable names; replacing inefficient file write code
BUGS: 4462

This is mainly a cleanup refactoring.  This is cleaning up code written by
other people so I will explain why I made these changes (kind of like a code
review):

(*) Changed name from GIT_EXECUTABLE to EG_EXECUTABLE: The name 'GIT' is
misleading because 'eg' is actually used, not 'git'.  And, you can't change
from 'eg' to 'git' because the script calls the 'switch' command which is an
eg command, not a git command.

(*) Changed from the variable name BRANCH_SUCCEEDED to EG_SWITCH_RETURN_VAL:
The convention for variables with a name like 'SUCCEEDED' is that a value of
'0' means false, while a value of not '0' means true.  This variable violated
this and was very confusing.  Also, the convention in the other commands in
this script use 'XXX_RETURN_VAL'.

(*) Changed the name Trilinos_FAILED_PACKAGES to
Trilinos_FAILED_LIB_BUILD_PACKAGES to make it clear what this list is actually
for which is to disable packages going forward.  Later, I will create a new
varaible called Trilinos_FAILED_PACKAGES that will contain the list of
packages that sees any type of failure which is needed for the CI logic.

(*) Replaced the logic for how the file CMakeCache.clean.txt is created.
Someone had used a loop that wrote to the file for every line (hundreds of
lines).  I had similar code that I found was a major performance problem on
Windows.  The fix for this is to create the entire output file as a string
first, and then write the whole thing at once.  I would guess that this will
litterally save 30 minutes of runtime on Windows.  Since there were no
excplicit timers for this operation you would never have seen it.

(*) Changed lingering 'CVS' comments with 'Git' comments.

Testing:
-------

I just ran the script by doing a 'make dashboard' and manually verified that
the cleaned cache file file got created correctly and got posted correctly.

I tried to test the git update logic by runing 'make dashboard' with
CTEST_DO_UPDATES=TRUE.  I got the error message:

    CTEST_UPDATE_COMMAND='/usr/local/bin/eg'
    /home/rabartl/PROJECTS/Trilinos.base/Trilinos/cmake/ctest/../.. exists so skipping the initial checkout.
    Doing GIT update of '/home/rabartl/PROJECTS/Trilinos.base/Trilinos/cmake/ctest/../..' ...
       There are modified or conflicting files in the repository
    CTEST_UPDATE(...) returned '0'
    CTEST_DROP_SITE = 'trilinos.sandia.gov'

When I looked on the dashboard and it complained about the files:

  Conflicting files (3)
      cmake/python/data
          CDashSubprojectDependencies.xml Revision: -1 by Unknown
          TrilinosPackageDependencies.xml Revision: -1 by Unknown
          TrilinosPackageDependenciesTable.html Revision: -1 by Unknown

These files just have their time-stamp updated, they are not actually
modified.  CTest is in error here but it seems to be harmless.  Otherwise, the
command seemed to work.