Developers Guide updated to reflect CMake build system
In its current state, the PyTrilinos Developers Guide is simply a set
of restructeured text files and correspondingly generated HTML files
that describe certain acpects of the information needed to be a
responsible PyTrilinos developer.  In the future, these will be tied
together to form a coherent set of documents (in txt, pdf and html
formats), but for now, I'm just making sure that the content is
up-to-date.