Moved spmdSpace() function back to public
In one of my last commits I accedentally made the spmdSpace() overridde
protected when it should be public.  This broke Teko (which is not in Trilinos
yet but should be soon).

I added a unit test so this will never break again by accident.

Testing:
--------

I only tested Thyra because this was a trivial change.