Changed non-const non-contiguous column views to skip copy-back of data on dangling ref
Here I changed the implementation of non-const non-contiguous column views to
use a weak ArrayRCP for the parent's local data and put in logic for if the
parent has been deleted then the copy back is skipped as a perforamnce
optimization.

I put in a very strong unit test for this for this functionality using a
debug-only sensing variable.  This is a great example of how to writte unit
tests to check such behavior.

=============================
Build/Test Cases Summary
=============================

Sat Jan 16 08:05:38 MST 2010

Enabled Packages: TrilinosFramework, Thyra
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=422,notpassed=0
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=451,notpassed=0

Local commits for this build/test group:
----------------------------------------
1011606 Changed non-const non-contiguous column views to skip copy-back of data on dangling ref
2913b1d Updated documentation on how to get the right diff of files.