Added workaround to PyTrilinos example scripts
Two example scripts, exAztecOO_Operator and exAztecOO_RowMatrix, both
suffered from the same problem.  On certain machines, when the AztecOO
solver would call back to the Apply() method, argument "y" would be
passed as a raw Epetra.Epetra_MultiVector rather than as a hybrid
Epetra.MultiVector.  When y was attempted to be accessed via square
brackets, it raised a python error.  These scripts now check the type
of y, and if it is not Epetra.MultiVector, it converts it.  I also
added comments to explain why I do this.  My hypothesis is that this
occurs because of a bug in GNU C++ 4.1.2.