Adding the REALLOC_BUG compile time flag.
There is a realloc bug that has occurred on more than one platform.  realloc
either fails to allocate memory when there is memory available or it crashes.
In every case, a malloc succeeds.  We spent a long time looking at this on
the Aleph application and were unable to solve it.

ZOLTAN_REALLOC now takes a 4th argument, which is the size of the buffer
before realloc'ing.  This is so ZOLTAN_REALLOC can do a malloc/memcpy/free
instead of a realloc if the compile time flag REALLOC_BUG is defined.

Almost all of the ZOLTAN_REALLOC calls in Zoltan were easy to convert.  It
was easy to see what the former size of the buffer was.  In matrix_distribute.c
and matrix_operations.c it was not clear what the size of the buffers were,
so I added two fields to the Zoltan_matrix structure to save the buffer
sizes.  Cedric may want to look at this and see if there is a smarter way
of handling this.

ctest passes when REALLOC_BUG is defined and when it is not defined.