fix an issue with the install(export...) call.
The way I was calling the install(export...) was invalid when there were no
libraries to install.This case happens when only the TrilinosFramework package
is configured. When only TrilinosFramework was enabled the configure failed
since there were no targets registered to install. We now check that there are
targets to install before calling the install(export...) command.