Added an ArrayView version of broadcast(...) for reference types; removed a use of arrayArg
The recent addition of TEUCHOS_DEPRECATED to arrayArg(...) caused an RTOp test
to not compile because of -Werror (RTOp is CLEANED).

What is scary is that the reference-type versions of the
Teuchos_CommHelper.hpp functions are *not* unit tested in the Teuchos package.
These seems to be only tested in RTOp but not very well I suspect.  There is
just a ton of technical debt that I have created in Trilinos.