Putting in if logic to avoid rewriting configured template header files
I put in if (not exists ...) checks to only write the implicit/explicit
instantiation headers if the file does not already exist.  Before, every time
you reconfigured with cmake it would rewrite all of the headers and result in
all the dependent code to be recompiled.  Now it does not do that.

=============================
Build/Test Cases Summary
=============================

Tue Mar  2 15:06:43 MST 2010

Enabled Packages: Thyra, RTOp

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=60,notpassed=0 (6.89645573298 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=61,notpassed=0 (8.25082219839 min)

Local commits for this build/test group:
----------------------------------------
06dcbe0 Putting in if logic to avoid rewriting configured template header files