Moving an header include in hopes of fixing Thyra compile error on razor (Bug 4709)
Here I have made a blind attempt to fix a compile error on the machie razor
with Intel 9.1.  This code was working on Feb 16 but was not recently.  I
suspect the problem is a bug in the Intel compiler and I think moving the
definition of Thyra::VectorBase above the include for the
Teuchos::PolynomialTraits class will fix the problem.  However, I can't be
sure until I can reproduce the error.

=============================
Build/Test Cases Summary
=============================

Tue Mar  2 16:25:41 MST 2010

Enabled Packages: Thyra

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=40,notpassed=0 (0.723322069645 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=41,notpassed=0 (0.320618633429 min)

Local commits for this build/test group:
----------------------------------------
79d3d10 Moving an header include in hopes of fixing Thyra compile error on razor (Bug 4709)