Updated exNOXEpetraExt_2DSim example script to working state
The main changes are that I allow the problem parameters to be set on
the command line, thus making it easier to play with different
scenarios.  I can easily solve the same problem as in exNOX_2DSim, and
have verified that I get the same answer.  Also, the default problem
had solution [1,1] and initial guess [1,1], which actually fails to
converge due to the relative F norm.  So the default initial guess has
been changed to [0.9,1.1].  Finally, I added more informative output
after the solve() method is called.