Disabling explicit instant for Trilinos on intel compilers
For whatever reason, explicit instantiation is just not working with the Intel
C++ compilers (Intel C++ 11.1 in the case I tried).  You get a bunch of link
errors like:

/scratch1/rabartl/PROJECTS/Sierra/SierraTridevCd/build/SierraBuild.trilinos_dev.devsierra/Trilinos/dev/intel-linux-11.1/release/address-model-64/mpi-openmpi-1.3.4/runtime-link-shared/threading-multi/libthyra.a(Thyra_DefaultSpmdVector.o):(.gnu.linkonce.d._ZTVN5Thyra10VectorBaseIdEE[.gnu.linkonce.d._ZTVN5Thyra10VectorBaseIdEE]+0x108): undefined reference to `virtual thunk to Thyra::MultiVectorBase<double>::clone() const'

My guess is that a different set of compiler options needs to be passed to the
Intel compiler on the compile or link line to make this work.  We can figure
this out later.

Note that the change was to just disale

Testing:
--------

(*) I build encore with Intel 11.1 with bake and it built (with implicit
instant shown by running nm on Thyra_MultiVectorBase.o)

(*) I build encore with g++ 4.2 under bake and it built (with explicit instant
shown by running nm on Thyra_MultiVectorBase.o)