Making Intel 10.1 build on godel a regular Nightly build instead of Experimental
Yesterday I addressed all of the failing builds and failing tests for the
packages that were enabled on the Intel 10.1 build.  This ran again as an
Experimental build and the results are shown here:

  http://trilinos.sandia.gov/cdash/index.php?project=Trilinos&display=project&date=20100320#Experimental

The only remaining failing test is Rythmos_CubicSplineInterpolator_UnitTest
but this is a known bad test (see bug 4679).  Hopefully this test will get
fixed soon.

Other packages can be added to this build as needed.