Global cmake version file for Trilinos
I have added a global version file for Trilinos to make branching a bit easier
The file is cmake/TrilinosVersion.cmake and stores the version in several forms
as well as some information needed for release branch testing. This file is
currently used to generate the Trilinos_version.h file and in our release
testing. This is first pass and could use some polish to make it really good.
For instance this file cannot be used in our cron scripts yet as it is a cmake
file. The places where this can be used and what version information we store in
it will likely change as we better understand everything we need to better
automate branching.

Because of the change that Trilinos_version.h is now a configured file and not a
hard coded header file it's position has changed to be the base build directory
had to be added to the list of include paths. This also required changed a few
packages that had hard coded in the old path to Trilinos_version.h. My guess is
that there are more packages that do require this file, but they don't seem to
be specifying it and just relying on picking it up from teuchos.

=============================
Build/Test Cases Summary
=============================

Wed Mar 24 09:08:48 MDT 2010

Enabled Packages: TrilinosFramework, CTrilinos, ForTrilinos, Teuchos
Enabled all Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=1192,notpassed=0 (22.8661535303 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=1244,notpassed=0 (11.5531094392 min)

Local commits for this build/test group:
----------------------------------------
2d531f5 Global cmake version file for Trilinos