Fixing the list of local commits for checkin-test.py on a branch
The commit I just did for the checkin-test.py script was using 'eg log
--oneline origin..' to give the local commits.  Unfortunately that gives the
diff between HEAD and origin/master which is *not* what you want.  I noticed
this in the commit that I just did on the trilinos-release-10-2-branch.  With
this commit, this is now done using the explicit branch name.  I manually
verified that this does what it is supposed to do and updated the unit tests
to check the correct eg command is being created.

=============================
Build/Test Cases Summary
=============================

Mon Mar 29 16:36:30 MDT 2010

Enabled Packages: TrilinosFramework
Enabled all Forward Packages

Build test results:
-------------------
0) MPI_DEBUG => passed: Trilinos/MPI_DEBUG: passed=51,notpassed=0 (0.0 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect commit/push readiness! (0.0 min)

Local commits for this build/test group:
----------------------------------------
c1fd04f Fixing the list of local commits for checkin-test.py on a branch