Fixed an issue where SGInverseModelEvaluator was picking up SG vectors.
I fixed an issue I discovered in Albany where SGInverseModelEvaluator
was trying to make block maps/vectors out of already block maps/vectors.
This is because it was dealing with all p's and g's, not just the non-sg
ones.  I added arguments that specify which are the non-sg vectors,
which are provided by SGModelEvaluator.  Updated the PCE example for
these changes.