Setup automatic update of version header for Sierra
When we went to a configured header file for the version number a static copy
had to be made for the Sierra build system. Sierra uses bjam to build Trilinos
which doesn't support our configured header files. This sets up a system such
that every time the version changes for Trilinos the static version header for
Sierra will be updated as well.