Changes for 64 bit gids.
Changes to ensure that zdrive uses ZOLTAN_ID_TYPE for global IDs and
ZOLTAN_GNO_TYPE for global counts.  In particular, the readers read
in ints (or whatever the chaco, exodus, etc readers read in) but
conversions are made to the proper type when stored in the zdrive MESH_INFO
structure.

Added a function in dr_random_io.c to create a graph.  It is capable of
creating a graph that globally has more than 2*10^9 vertices and edges.
It is requested in the zdrive input file with

file type = create-a-graph
size = 10000000000          /* global number of vertices */

The other mesh and point generators in dr_random_io.c create the
entire thing on one process and then deal it out.  create_a_graph
builds the simple graph in a distributed fashion.