Changes to support 64 bit gids
Many changes required due to the fact that different sized ints may be
using these functions.  Before they were are 32 bit ints.

Zoltan_Map_Create takes as an argument the number of bytes in the key,
not the number or ZOLTAN_ID_TYPEs in the key.

Copied some of the sorts in zz_sort.c so they work on ZOLTAN_GNO_TYPEs
as well as on ints.

Added Zoltan_set_mpi_types() to zz_util.c.  It figures out the
Zoltan MPI type for ZOLTAN_GNO_TYPE.  This can't be done until runtime.
The MPI type for ZOLTAN_ID_TYPE is figured out in zoltan_types.h since
the ZOLTAN_ID_TYPE is provided on the compile command line.