Added option for projecting integrals in Stieltjes procedure.
This required changing the StieltjesPCEBasis interface to take RCP's,
which in turn required changing several examples and tests.  This seems
to work better than the original approach in that it guarantees that the
basis norms are positive.  I also have getQuadPoints() truncate the the
quadrature order at 2*p if not using the original PCE quad points, since
this is all that can be evaluated anyway (if you don't project the
integrals, you might be able to get higher order quadrature points if
the original quadrature was higher order, but generally it is very
difficult to get these higher order points).  For the sparse-grid
library, I then add extra points with zero weight to fill out the
required number of points.  I then tell SGQuadModelEvaluator not to
evaluate these points since they don't contribute to the integrals.