Reworked 1-D basis methods needed for sparse grids.
Removed the point and weight factor methods as they are no longer used,
and added a growth rule method.  This simplified the sparse grid
interface, and allows for proper specification of the growth rule for
Clenshaw-Curtis points.  The growth rule in this case is determined by
whether the discretization is isotropic, which is indicated by a flag to
the Clenshaw-Curtis basis constructor.