Make code for the complex types SS in development mode, but PS in releases
Added a macro that determines if a feature should be PS or SS based on the
release mode. This macro outputs a variable that is can be used in place of
${PROJECT_NAME}_ENABLE_SECONDARY_STABLE_CODE to determine if something should be
enabled. In development mode it will be ON only if SS code is enabled, in
releases it is always set to ON.

Used the above macro to determine the default state for several package's
complex type functionality based on the agreement of developers. This was
already SS in Tpetra. In Teuchos and Sacado this was PS so this will change how
they build by default in developement mode. Also note that there are several
packages that depended on the Teuchos setting for complex to determine their
complex setting so they will also be affected. For release distributions the
complex code will all be turned on now so users should not see a difference.

Note that the checkin test script will not run SS by default, however the
nightly testing does test SS code. If you change something that you believe may
affect the complex code in the packages you touched it is advisable to
explicitly turn on the complex code.

=============================
Build/Test Cases Summary
=============================

Fri Jun 18 09:53:11 MDT 2010

Enabled Packages: TrilinosFramework, Sacado, Teuchos, Tpetra
Enabled all Packages

Build test results:
-------------------
0) MPI_DEBUG => FAILED: Trilinos/MPI_DEBUG: passed=1175,notpassed=1 => Not ready for final commit/push! (0.0 min)
1) SERIAL_RELEASE => passed: Trilinos/SERIAL_RELEASE: passed=1209,notpassed=0 (0.0 min)

Local commits for this build/test group:
----------------------------------------
4416d32 Make code for the complex types SS in development mode, but PS in releases

***
*** WARNING: The acceptance criteria for doing a commit/push has *not*
*** been met, but a commit/push is being forced anyway by --force-commit-push!
***