Fix to contiguous allocator.
Eric Phipps uncovered a bug in the contiguous allocator when building on 32bit systems.  This commit is a refactor of the padding algorithm to do the right thing on all platforms.  The bug was introduced a month ago when I changed over ctor/dtor duties to new support functions in the ArrayRCP.