Changing the final email to have the amended SHA1 listed
Before this commit, the SHA1 emabedded in the final email that got sent to the
developer and to the checkin archive mail list was the SHA1 before the last
commit was amended.  Now the SHA1 shown is after the amended commit and
therefore now the SHA1 that gets archived is actually in the global repo.
This will help to match up commits.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=52,notpassed=0 (1.06 min)
1) SERIAL_RELEASE => passed: passed=52,notpassed=0 (0.73 min)