Quick fix for an issue with discovering mpi compilers
When the MPI_BASE_DIR variable was set it was possible to find incompatible mpi
compilers because of the order in which cmake looks for compiler names. The name
mpicxx is preferred over mpic++ for instance. This was a problem because even if
MPI_BASE_DIR was set several system paths were also put on the list of paths to
search. I have changed the logic to not put any system paths on the list of
search paths if MPI_BASE_DIR was set. This should ensure that the compilers
specified by the config script are preferred over any other compilers. There is
still a possibility for the system to find compilers outside the MPI_BASE_DIR
tree, but only if there are missing compilers. IE if no c++ compiler exists in
MPI_BASE_DIR then the users path will be searched for one still.

I believe that there is still a potential for our find mpi to do the wrong thing
as if MPI_BASE_DIR is not set standard system paths would be searched before the
users path which doesn't feel right to me. But this quick fix addresses the
known issue of finding incompatible compilers due to an ordering issue. And I
feel that the potential for harm with the current functionality is pretty low as
if there are multiple compilers on a machine it is always a good idea to give
the configure script hints on how to find the compilers you want, which the
MPI_BASE_DIR variable does.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1157,notpassed=0 (324.56 min)
1) SERIAL_RELEASE => passed: passed=1181,notpassed=0 (266.44 min)