Modifications to RefMaxwell to allow direct use of Ifpack (rather than via ML).
This is due to a (disturbing) revelation that calling Ifpack smoother directly
and calling a 1-level ML using that same Ifpack smoother don't do the same thing.
See Bug #4855

I don't know why and at the moment I don't have time to figure it out at the moment.

This temporary fix allows RefMaxwell to bypass that if you ask for IFPACK by name.