Adding hwloc code
I reorganized the network_topology directory into two directories.

MPI has code that uses the special IAA MPI that is available on
glory to detect network topology that Zoltan would use for
hierarchical partitioning.

hwloc is similar, but uses the portable hwloc library from OPEN MPI
and the University of Bordeaux.  In particular a proposed function
Zoltan_Get_Topology is checked in in the file zoltan_get_topology.c.