Updated failing Epetra_CrsGraph test in PyTrilinos...
A PyTrilinos test which exercised duplicate index-insertions in Epetra_CrsGraph
started failing recently. It was because Epetra_CrsGraph keeps indices sorted
and unique now, so the length of the row wasn't growing with the duplicate insertion.
The PyTrilinos test is now updated to verify that the row stays the same length.

Another PyTrilinos test broke because Epetra_CrsGraph::RemoveMyIndices wasn't
working as expected. This was a bug in Epetra_CrsGraph, and that bug has now
been fixed (in another commit).

I verified this fix on the trilinos-test2 machine. I'm doing the checkin from
a different machine, on which I can't build PyTrilinos, so I'm disabling
PyTrilinos for the checkin-test.py run...