More work on the rama_solver branch.
This was mostly cleaning up the changes to the Sparse3Tensor.  In doing
so I completely changed the internal data structures and how you iterate
over the entries in the tensor.  It now uses a much more efficient,
sorted data structure, and you iterate over it using iterators.  There
are now accessors for both i-based and k-based indexing, to support
matrix-vector products and Gauss-Seidel.  The original Najm et al data
structure is still there too, as it is used by the expansion classes.