(From Sierra) STK_MESH: Progress towards cleaning up unit-tests
1) Consolidate use of box mesh fixtures in unit-tests

There were four different implementations of generate_boxes in the unit-test
area. Several of these implementations ignored the aura flag leading to
extremely misleading behavior in some unit-tests where the aura flag was
set to false, but passing execution of the tests depended on the aura being
generated.

The four implementations were similar enough that they were all able to work
with a shared implementation (with some minor modifications and extra flags);
I placed this shared implementation in fixtures/GenerateBoxes.hpp.

UnitTestMesh was converted to BoxFixture and moves to fixtures. BoxFixture
uses GenerateBoxes.

These changes required that some friends be changed in BulkData. The team
feels we should eventually remove this friend and change the few unit tests
that depend on it.

2) Cleanup of UnitTestBucket.cpp

* Reorganize #includes, grouping common include areas
* Split one huge STKUNIT_UNIT_TEST into four
* Improve documentation
* Use STKUNIT_ASSERT_THROW instead of very verbose alternative
* Remove unnecessary test output
* Put FIXMEs to address large commented-out areas

3) Remove unnecessary header files where possible

4) internal_modification_end exposure sorted out

The fixtures that sometimes need the ability to do modification_end without
generating an aura are templated. When you create instances of these classes
with the bool template set to true, you are promising to define
modification_end_wrapper at some point in the program; otherwise you'll get
a link failure. This seems like the only flexible solution for these classes.

5) Lots more general unit-test cleanup; attempts to improve documentation etc.

6) Repaired UnitTestBulkData_ChangeParts

This file had a broken (it was passing, but not working as intended) unit
test that was not being run in parallel but had parallel testing code within
it.

7) Merge of GenerateBoxes with BoxFixture

8) Remove unit-test related code out of RingFixture and BoxFixture

9) Rename RingMeshFixture to RingFixture

10) Add unit tests for BoxFixture and RingFixture

Build/Test Cases Summary
Enabled Packages: STK
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=4,notpassed=0 (1.25 min)
1) SERIAL_RELEASE => passed: passed=4,notpassed=0 (0.91 min)