Changing an error message in phalanx
When the evaluator dependencies fail to be met, phalanx was printing instructions on how to generate a png file from the dot file to locate the missing dependencies.  However when the graphs get really big, dot creates png files that are unreadable.   I changed the message so that it now instructs users to create jpg files from the dot files.  The jpg files are still readable when you have 100+ evaluators.