PyTrilinos tests now use CMake PYTHON_EXECUTABLE
All of the PyTrilinos tests and examples source names were changed
from *.src to *.py.in.  Infrastructure was added to allow me to copy
from the source directory to the binary directory while performing
substitutions, much like configure files.  Functionally, all source
files with the text ${CMAKE_VAR_NAME} will be substituted with the
CMake variable's value.  Specifically, the first line of all of the
tests was changed to

  #! ${PYTHON_EXECUTABLE}

so that the test scripts will designate the CMake-recognized python
executable as the script executable.

CMake variables are obtained from the top binary directory
CMakeCache.txt file using the new python script,
util/CMakeCacheVariables.py.  The top binary directory name is
obtained from the local Makefile using the newly reinstated
util/MakefileVariables.py.

The cmake/PyTrilinosMakeTest.cmake file had to be changed to utilize
this new logic, and the CMakeLists.txt files in the test and example
directories had to be changed also to copy non-test data files and
python modules in the same way.