Add option to remove boundary conditions to make problem symmetric
I did this by reworking the mesh significantly.  Now I think it is a
little clearer what is going on.  Basically I always create the full
mesh and just don't include the nodes corresponding to the boundary in
the Epetra maps/graphs.  Overall this seems to have very little effect
on CG, so it probably isn't really necessary.