Fixed bug 4995: finalization errors when other MPI modules used
If another MPI python module, such as mpi4py, is used, then the Epetra
or Teuchos modules can cause premature MPI finalization.  This fix
prevents that from happening.

The source fix was added to Teuchos_Comm.i and Epetra_Comm.i.  An
example script, named exMPI_Init.py.in, which tests the fixes was
added to the example directory, along with modifications to the
example/CMakeLists.txt file to add the new test to the suite.

Utility python script copyWithCMakeSubstitutions.py was modified to
properly output non-string values.  I needed access to the
TPL_ENABLE_MPI CMake variable, and it was crashing the script because
its value is a Boolean, not a string.