Modified PyTrilinos Developers Guide to include section on handling C arrays
C array arguments are handled in PyTrilinos by using the numpy.i SWIG
interface file.  This new section references the documentation for
that file, and adds the details for how to initialize in PyTrilinos.
The initialization process is different in PyTrilinos than in the
numpy.i file because PyTrilinos is a multi-module package.