Changing the pass grep strings for debug mode and removing a bad FINAL_FAIL_REGULAR_EXPRESSION.
For some strange reason, when you only have -g but not full debuging, the
stacktrace that gets created does not show all of the functions.  My guess is
that with -O3 that some implementations will hid inlined function calls from
the stacktrace.  This is what happened with gcc 3.4.6 on gabriel.sandia.gov I
suspect (where I am fixing this).

To account for this, I changed the test criteria to only look for a single
func_.() call in the stacktrace when only -g is set.  When a full debug build
is performed the test checks for both func_a() and func_b().

Also, I removing the FINAL_FAIL_REGULAR_EXPRESSION from the nostacktrace test
since it disarms the test and causes it to always pass.