Fixed errors in Stokhos expansion classes and overloaded operators.
I added unit tests for the Stokhos Sacado overloaded operators, which
uncovered a number of errors, particularly in the new ET version.  I
also made the constant orthog poly expansion used by the overloaded
operators static, to save a bunch of calls to new/delete.  There are now
corresponding cpp files to initialize the static data for the standard
int/double instantiation.  I also found a problem with the
implementation of * and *= when using a linear Cijk causing the higher
order terms not be be computed correctly.  I put the common
implementation of many of the expansion functions in a base class so
that the code for things like * and *= is just in one place and not
duplicated.  Finally I setup the ET overloaded operators to not always
require a quadrature expansion, although they still need it for any
expression that is more than quadratic.