Add Zoltan_mpi_gno_type()
Instead of defining the MPI_Datatype for a ZOLTAN_GNO_TYPE during
Zoltan_Init, we compute it and return it the first time
Zoltan_mpi_gno_type() is called.  This gets rid of the external
variables that were being used.  It also returns Zoltan_Init to
it's original semantic, in which app didn't need to call it if they
had already called MPI_Init.  It also removes the need for a conditional
expression with the MPI_Datatype which required care since the MPI_Datatype
can be integral or a pointer depending on the implementation.

This fixes bug #5053. ctest for Isorropia and Zoltan passes on my machine.