Minor syntax change to support building Dakota with CMake.
Dakota supports an experimental CMake build, which means it can be built
inside of Trilinos in a much simpler way than before.  Unfortunately
however, a package in Dakota defines an option called MPI, which
conflicts with the string used in the specification of the COMM in the
AddExecutableAndTest() macro.  This was causing the string comparisons
in this macro to fail.  It was easily fix by replacing
      ELSEIF (COMM STREQUAL "mpi")
with
      ELSEIF (${COMM} STREQUAL "mpi")
etc, which when COMM=MPI, apparently prevents the value of MPI from
being expanded.