(From Sierra) STK: Use-case refactor
We were again hit with the problem of false positives coming
from runtest due to weak checks in the test XML files. I switched
all STK use-cases over to using a system similar to the system I
put in place for the unit-tests.

The error that was being hidden was coming from the linsys usecases.
The fix was pretty trivial and should fix the valgrind errors coming
from this test.