partial distance 2 code is completed. Bugs fixed when non-local GIDs and colors given in color and test
When used with graph symmetrization process GIDs that is received by coloring for
partial distance  may not belong to the processors and needs to be handled properly.
Similar situation was occuring in Color_Test. All, incuding documentation should be fixed now.
Note that: if you run coloring with graph symmetrization in zdrive, both distance-1 and distance-2
test will fail (correctly!) because, zdrive only queries colors of objects that it knows, since graph
symmetrization adds new objects, their colored will not be queried by zdrive (and cannot be queried)
hence those color info never returned to zdrive and hence not passed back to Color_Test. This is
expected behaviour due to design.