Add include directories for zoltan tests
Several Zoltan tests were not picking up the correct list of include directories when in installation mode. The tests had been picking up the include directories from the library, but since those directories are not added when in installation testing mode they needed to be added for just the tests.