BUG: 5155
Belos solvers individually modify the block size or numbers of blocks
based on the operator's dimension.  A general statement or modification
of the block size and number of blocks will not always be correct.
The pseudo block GMRES solver, for instance, is "block size" number
of individual GMRES solvers, so it is not trying to build a subspace
that is "block size" * "num blocks" long, only "num blocks" matters
for this solver.

Also, some solvers don't have a block size associated with them, so this
attempt to get the block size from the solvers may cause an error in
those cases.