MueLu Aggregation debugging
Created a standalone test that reproduced an error I have inside of the Aggregation code. It is an import operation that is working for Epetra but not for Tpetra (with 2 procs.).
I have now to figure out if this import is incorrect (== bug in Aggregation) or if it is Tpetra bug.