Slightly modifying problem sizes.
Unit tests almost all pass in serial.  There are a few failures in parallel, but none that are worrisome.

The following tests fail in parallel:

        10. Hierarchy_FillHierarchy_PRFactoryOnly_UnitTest ...
        18. IfpackSmoother_GaussSeidel_UnitTest ...
        20. IfpackSmoother_ILU_UnitTest

test 10 fails because the gold standard is operator complexity, nnz, etc. for a serial (thus differently sized) problem.
test 18 fails because GS isn't actually block Jacobi in parallel.
test 20 fails for the same reason as 18.

The following tests always fail under Epetra because Epetra does not support for long int.

        47. UCAggregationFactory_LongInt_LongInt_Node_LMOLongInt_Build_UnitTest ...
        49. UCAggregationFactory_int_LongInt_Node_LMOint_Build_UnitTest