Updates for installation and backwards compatibility testing
Added an option to select which branch the script that creates a tarball and
installation will work on. This is necessary for backwards compatibility
testing.

Added an option to report the output to stdout when there is an error. This
makes it possible to see what went wrong in the nightly_create_tarball script
from the TDD dashboard instead of having to track down the log files that the
script creates. The log files will still be created with this option.