Allowing Trilinos_EXTRAREPOS_FILE to be set in cmake configure
This change updates the basic Trilinos CMakePackageArch system to
allow the file that defines the list of extra repos with add-on
packages to be given on the CMake configure command-line.  This will
be used by CASL to define a custom set of CASL-only extra repos and
add-on packages in CI and nightly testing.

The actual changes to the specific builds will come next.