Improved way Dakota's definitions are extracted.
Now we can get all of them with one property query (strangely though
CMake removes all of the -D's, which I had to re-add).  I had to
manually put in Dakota's new MPI definition, since they don't seem to be
doing it.  Still get seg faults in the TriKota tests that use Dakota.
This may be related to this MPI definition I had to add.