Fix two TrilinosFramework test assumptions
There were two tests that were checking for a string in their output which was
based on a bad assumption that all Trilinos clones would be in a directory
called "Trilinos". Instead they now use the variable ${PROJECT_HOME_DIR} to
check for in their output which is much safer and is the same variable used in
the configure for these tests so there won't be any further discontinuities
between the configure and the expected test output.