Added option --abort-gracefully-if-no-enables (bug 5227)
Here I added the option --abort-gracefully-if-no-enables that gives now
behavior to the script where if --abort-gracefully-if-no-enables is set and no
enables are set, then checkin-test.py will abort gracefully.  This will be
used to implement a cron job demon that will move commits between repos
(needed for CASL, see 5227).

There was no existing unit test in place that defined the behavior of the
checkin-test.py script when no files are changed, resulting in no enables, and
what the script would do.  Without setting --abort-gracefully-if-no-enables
the script did and still does fail configure because
Trilinos_ALLOW_NO_PACKAGES=OFF is still being passed to the CMake configure.
There still is no unit test for this case but this is not a typical use case
because people should not be running the checkin-test.py script if there are
not changes that would result in no package enalbes.

Note that when a gracefull abort occurs, no emails will get sent out.  Again,
this is to support the target use case where checkin-test.py will get run in a
loop and we don't want hundreds or thousands of emails going out all of the
time for no-ops.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=86,notpassed=0 (1.04 min)
1) SERIAL_RELEASE => passed: passed=86,notpassed=0 (0.87 min)