Fixed some ref count problems in PyTrilinos
I added the SWIG flag -noproxydel to the build system so that __del__
methods will not be added to python shadow classes.  This will improve
the garbage collecting, as the __del__ method was a no-op and the
presence of a __del__ can prevent the deletion of circular references.

I added 'del suite' to several unit test scripts.  This is because
python 2.7 apparently has a __del__ method in the unittest object and
circular references can cause the suite object to not get deleted
correctly.  This seems to bypass the problem.

I bit the bullet and treat Epetra_Object objects with an RCP now.
This required refactoring some printing infrastructure and making sure
that every Epetra class that derives from Epetra_Object is also stored
with an RCP.  This in turn required modifying the SWIG interface files
to several other packages so that their generated code includes the
newly afected classes.

Added unit tests that exercise the Epetra_Object Label() method in
derived classes to make sure that the inheritance is working
correctly.

Edited the developers guide to get rid of the Epetra_Object trick.