Rewrite Zoltan_Hier to make it faster.
By using Zoltan_Map (which uses Zoltan_Hash) we can reduce the
size of the queries to the Zoltan_DD global data directories.  Also because
of Zoltan_Map we don't need to keep the global number system sequential
and increasing across processes.

The hier test answers are all the same except for one of the ch_hammond
tests.  The metrics are almost identical but the answer is not exactly
the same.

Running hierarchical partitioning (2, 2, 6) with 24 processes each
on one core of a dual socket, dual die, 6-core die node, with a
graph with 1 million vertices, using scotch for partitioning at
each level, the old Zoltan_Hier took 38 seconds and the new
Zoltan_Hier took 14 seconds.