Add a script mode to the update copyright script
The script previously only supported c/c++ files but I needed to use it for a
number of cmake and python files so I added a script mode which removed the
begin/end commenting from the copyright notice that is inserted into a file. To
use this use the option --script-mode.

The default behavior of applying the c style comments at the beginning and end
of the copyright notice was not altered.

Build/Test Cases Summary
Enabled Packages: ThreadPool, Teuchos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=1097,notpassed=0 (269.21 min)
1) SERIAL_RELEASE => passed: passed=1063,notpassed=0 (188.37 min)
Other local commits for this build/test group: 92368cb, a9d97fe