Fixing TriKota enable logic
With this change (and the change in commit 3d03451), TriKota should
now be properly enabled given the correct logic.  This means that
Dakota will only now be built as part of the TriKota package and email
will get sent to the TriKota regressions mail list if it fails for
some reason.  Before, email would have gone to the Teuchos mail list!