Fixed 2 bugs in the Amesos.i interface file
The first bug fix was that the Amesos::Create() method returns a
pointer to a new Amesos_BaseSolver object.  SWIG did not know this,
which prevented it from deleting the object propoerly.  This was fixed
with a %newobject directive.

The second bug was that there was a %extend directive for the
Amesos_BaseSolver class that added a __del__() method which was
prematurely deleting instances of Amesos_BaseSolver and causing a
segmentation fault.  I don't know why the %extend was ther, but I have
deleted it and the class works as expected now.