Fixing Chebyshev options and output when using Ifpack2.
Adding header (similar to ML's) for informative printing to screen.
In Ifpack2, the option to turn on Chebyshev is different than in Ifpack.
I've added a check of the Ifpack smoother type.  This was necessary as how a zero initial guess is indicated is different in Chebyshev than it is in SGS.
This is unfortunate, as this check needs to be updated to support other smoothers, instead of just letting everything pass through to Ifpack2.