Automatic enabling of tests requiring other packages, fixing Ifpack2 smoother options.
The appropriate MueLu unit tests are now enabled if Trilinos is configured with Amesos, Ifpack, and/or Ifpack2.
This was previously broken, as we were incorrectly checking for a macro in the *.cpp file.
The check now being done in the CMakeLists.txt file.

Fixing options for SGS and Chebyshev.  They differ from those in Ifpack.

The following tests still fail, however:

    20. Ifpack2Smoother_GaussSeidel_UnitTest ...
    21. Ifpack2Smoother_Chebyshev_UnitTest ...
    22. Ifpack2Smoother_ILU_UnitTest ...
    57. UCAggregationFactory_Scalar_LongInt_LongInt_Node_LMOLongInt_Build_UnitTest ...
    59. UCAggregationFactory_Scalar_LongInt_LongLongInt_Node_LMOLongInt_Build_UnitTest ...
    61. UCAggregationFactory_Scalar_int_LongInt_Node_LMOint_Build_UnitTest ...