Using new power and scaling method.
AP is still fillCompleted.  For reasons I don't understand, A*P w/o fillCompleting
yields a bizarre matrix, and the final smoothProlongator is wrong.
The good news:
   scaling is now done implicitly within the power method
   scaling AP does not required fillCompleting

Bottom line:  two fillCompletes are done, one for AP, and one for the final smoothed prolongator.

Timers are slightly rearranged to reflect all this.