Fixes for change to teuchos time monitor.
MPI_Init must now be called if time monitor is built with mpi enabled as it is now making direct mpi calls even if only on a single process.  So all unit tests that are built for mpi with 1 proc are failing unless we add a GlobaMPISession object.

Build/Test Cases Summary
Enabled Packages: NOX
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=141,notpassed=0 (14.98 min)
1) SERIAL_RELEASE => passed: passed=148,notpassed=0 (6.71 min)