Add subpackages to seacas
Changed SECAS to use the subpackage architecture. Pretty much every library and
application in SECAS is now its own subpackage which will allow those that rely
on specific parts of SECAS to use only those parts they need.

As part of this I did some clean up to make SEACAS use the package architecture
more where it should be. Some of the issues with why it couldn't have been
resolved with the subpackage architecture such as the application aprepro which
previously would mistakenly include the aprepro_lib and end up with multiply
defined symbols. Since the application aprepro doesn't depend on the library
aprepro_lib this no longer happens.

I also changed the condition for how the scripts directory is added. The script
in that directory seemed to depend on nemslice and nemspread since those are now
separate subpackages and not just part of the applications directory I made the
scripts dir only be configured when both nemslice and nemspread are available.