Adjusting stacktrace test for Intel release build.
When stacktracing was enabled for a CASL-related Intel 12.0.x build,
it would found that the test
Teuchos_TestForException_throwTest_stacktrace failed because it was
not printing out any of the functions in the call stack between main()
and the final TEST_FOR_EXCEPTION(...) call.  I make sure this was not
due to an inlining issue, I moved the functions func_a() and func_b()
to a separate file so they could not be inlined.  However, this only
partially fixed the problem.  It would seem that with Intel 12.0.x,
with high optimization (i.e. -O3) that the Intel compiler may not be
constructing a complete call stack.  In any case, I changed the CMake
logic to not look for any of the intermedate functions called in the
call stack.  Now the test passes.  I can't be concerned about testing
all of the behaviors of the stacktracking code based on BinUtils for
every different type of compiler for various levels of optimization.
If it works for g++ with -g -O0 I am happy.

I also added a little more verbose-mode debuging output for compiler
options.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=162,notpassed=0 (6.50 min)
1) SERIAL_RELEASE => passed: passed=159,notpassed=0 (2.46 min)