Replacing '/' with '.' in REGEX directory paths to accomidate Windows.
With my last commit, the test that checks for the correct file name
and line number used unix paths and was changed to run on all
platforms, including Windows.  This failed on Windows because the C++
preprocessor gives file paths with '\' instead '/'.  Replacing '/'
with the REGEX special match character '.' to match either '/' or '\'
should fix the problem.

I have not tested this on Windows but I suspect this will fix the
errors on KALLIKRATES for the test
Teuchos_TestForException_throwTest_nostacktrace shown for the last two
days.

Build/Test Cases Summary
Enabled Packages: Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=75,notpassed=0 (0.51 min)
1) SERIAL_RELEASE => passed: passed=72,notpassed=0 (0.41 min)