Fixing Transpose utility to work with Epetra or Tpetra.
Moving Transpose method from Utils class to Utils2 class.  Transpose is the only method so far that needs
a specialization for the case GlobalOrdinal=int (for Epetra).  However, if it remained in Utils, I would
have had to created specializations of all the other methods, which would be really time-consuming.
Many thanks to Jeremie, who suggested this and helped diagnose the final compilation bugs.