Patch to get insertGlobal working with Epetra.
MakeTentative now working for Epetra in serial and parallel.

The import's target vector for holding row numbers has the same length as the local number of fine DoFs.
The loop was incorrectly trying to insert a global row for every entry in the target.  Since most of the entries
are zero (i.e., no import has occured), Epetra was issuing an error when global row 0 wasn't owned by the current
processor.  This was an easy fix -- just initialize the target vector to -1, then check that the row number is
non-negative before inserting.