MueLu: Implementation of DefaultFactoryHandler The DefaultFactoryHandler will manage default factories. It will also allows to remove the Scalar template parameter in Graph factories.
Implementation notes:
---------------------
There is two classes: one base class and one templated class
The templated class is instanciated by Hierarchy and the base class is used everyelse where.
This design allows to use a templated factory inside of a class who isn't (ex: calling CoalesceFact inside of AggregationFact, even if AggregationFact is not templated with Scalar type)
The templated version of the class is needed to know what the default factories are (ie: Factory<int,double> or Factory<long,complex> for instance).

I added an interface FactoryBase at the top of SingleFactoryBase/TwoFactoryBase to simplify the use of the DefaultFactoryHandler (no need for dynamic_cast when calling Build())