Adding an explicit -DTPL_ENABLE_MPI:BOOL=OFF to SERIAL_RELEASE
I added an explicit -DTPL_ENABLE_MPI:BOOL=OFF to the SERIAL_RELEASE
built-in build just in case there are packages that have a required
dependence on MPI to make sure they get disabled.  I put this in for
when Panzer gets added to the main Trilnos git repo so that it will
get disabled in the SERIAL_RELEASE build.