MueLu: fixed test RAPFactory Correctness
It was a really bad idea to enable the 'factory' parameter of the Set/Get method in Level. Before the ExtendedHashTable, this argument was just ignored, and it was fine. Now, a lot of tests are passing but are just wrong (factory build called several time...).
Keeping the distinction between my NewGet/Set and the old ones until I go through all the factories and add the parameter 'factory' to each Set/Get would have made transition smooths...