Fix last python script with misplaced #! /bin/env python line
I thought I had fixed all these before the initial commit of the copyright
header, but apparently a few still had the shebang lines in the wrong place.
This is the only one left that grep is showing as being broken so this should be
the last change like this needed.