Disable pulling extra repos on release branchs by default
The default was to always pull some set of extra repository when the
CTEST_TEST_TYPE was Nightly. This is a bit of a problem for release builds as
precopyright specifically is not guaranteed to work with a release branch. I am
changing the default for release testing to not try to retrieve any external
repositories by setting the Trilinos_ENABLE_KNOWN_EXTERNAL_REPOS_TYPE_DEFAULT to
"None". This should make it clear that no external repos will be used, but will
still allow them to be used if desired by setting
Trilinos_ENABLE_KNOWN_EXTERNAL_REPOS_TYPE to the desired value either with a -D
or as an environment variable.

Also did a little bit of clean up that was possible now that we have some more
information in the driver core. The Track and branch can now be defaulted to a
more reasonable value which should do the right thing for the branch we are on
instead of expecting the ctest*.cmake files to specify them. The ctest*.cmake
files can still specify them if the default isn't what is needed though.

The default for development testing has not been changed