Fix a few framework tests broken by recent commit
My recent commit changing some of the defaults for dashboard testing caused a
few tests to fail. The first was because the part of the update that checks out
the correct branch was not properly ifd out for testing like the ctest_update
call was. This needs to be refactored a bit to be part of the same macro that
now calls the ctest_update, but that was more than I felt comfortable doing on a
branch. The other was a couple tests were very strict about expecting a certain
external repository set, but weren't actually setting which set that was so it
changed with the change in default for release mode. That was fixed by
explicitly stating the external repos they need.