MueLu: Hierarchy/DefaultFactoryHandler
Instead of storing a defaultFactoryHandler in the Hierarchy, a new one is now created and set at the beginning of FillComplete/SetSmoothers/SetCoarseSolver. At the end of each methods, the defaultfactoryhandlers of Levels are reset to null.
I did this because default factories should only be temporary and should not be kept by Hierarchy.

I also discovered and fixed a lot of problems in unit tests.