Adding a TpetraCrsMatrixInput adapter to TestAdapter.hpp
which uses the high order 4 bytes of an 8 byte ID.
Simplified code in unit test XpetraInputAdapters.cpp and
added a test of the creation and use of the 64-bit adapter.
This compiles, but Tpetra::Map throws an error when I give
it these Ids.  "Please contact Tpetra team" is not a good sign.