Reimplement CommTest to use only Teuchos::Comm interface.
That test should now be capable of running in parallel or serial
regardless of whether or not MPI was compiled in. The input
adapters test still creates a Teuchos::Comm by hand using
MPI_COMM_WORLD rather than calling Teuchos::DefaultComm because
it is using both Epetra_MpiComm and Teuchos::Comm.