Moving CASL builds before Trilinos builds, changing timeout from 4 to 5 hours
With these changes, the CASL add-on packages should be tested first and finish
first followed by the native Trilinos package builds.

I also increased the timeout from 4 to 5 hours because the
MPI_DEBUG_GCC_CASLDEV build timed out at 4 hours and som other builds were
very close to 4 hour.  This is all caused by the single Watts Bar test problem
(that really needs to be moved) to a weekend test.

If this does not work, we will have to disable the Watts Bar test in most
builds because it is just taking way too much time.