Adding an accessor for the expansion object
Also fixing issue (via hack) where the parameter list
is used as an RCP but never saved. A sublist is saved
as an RCP.  If the original parameter list gets
deleted or goes out of scope then the sublist is invalid
and a seg fault results. This change copies the parameterlist
to a new RCP that will be guranteed to exist until the
object is deallocated.