Seacas Mapvar fixes.
The INDEXX function call in Mapvar had the calling
arguments in the wrong order resulting in a seg
fault.

Changed an error message to a warning. Mapvar
checks that the element types of the sending
and receiving meshes are the same. In my case
this is not necessary and seems to run fine
interpolating a nodal field on a tet mesh to
a nodal field on a mesh with spherical elements.
A very ugly warning is still printed so if it
does fail in some cases, it should be apparent
what went wrong.