MueLu: Level::Keep now needs valid factory pointer. The usage of Level::GetFactory() does not work if default factories are needed, that have to be created by a factory manager object that often itself does not exist when Keep is called!
Note: Keep is meant to be used by the MueLu-user for debugging purposes.
So he can call Keep for a specific variable generated by a specific factory before
Hierarchy::Setup is called! Keep is not meant to be used anywhere else by the MueLu
developer! The developer should always use the Request/Release mechanism.

Keep is now just a flag that marks a variable not to be freed after the multigrid hierarchy is built.