cleaning up configure time options
All Zoltan2_config.h macros are "HAVE_ZOLTAN2_*" or "SHOW_ZOLTAN2_*" so
we don't conflict with other package or application defines.

Removing the MALLINFO and Mallinfo definitions.  It seems this is
rarely helpful.  We can do memory profiling on linux with meminfo.

We can't compile at this point, because I am in the process of
reorganizing directories.  I'm committing now so that we don't
lose the history of changed files.