Epetra: Modifying Epetra_CrsGraph to allow user to replace the ColMap if no entries have been inserted yet.
Also modified documentation for Epetra_CrsMatrix to indicate this.

The idea behind this is to allow the user to replace a ColMap if it hasn't
been used for anything meaningful yet.  This change is being made to
support a refactor to the EpetraExt::MatrrixMatrix::Multiply to allow
insertions in local indices, which will require the MMM to specify the
column map.