MueLu: new default strategy in Level
Previously, when factory == NULL, Level was calling directly the factory manager to get the default factory.
Now, if user data is available, it is considered as the default and the factory manager is ignored.
Otherwise, the factory manager is used as usual.

This strategy allows to use the same factory manager on the fine and coarse level without any trouble.

Example :

FineLevel:
----------
A          -> User provided
Nullspace  -> User provided

CoarseLevel:
------------
A         -> RAPFactory
NullSpace -> NullspaceFactory