Framework: put Relation on a diet
Relation had become bloated on the stk_migration branch with extra data members,
increasing the size of the class from 16 to 24 bytes. This commit seeks to slim
Relation back down to the same size it is on master.

This commit has an impact on client code due to get-derived-type functionality
for Relations being moved to a free function.

This commit also ensure thats get_derived_type will be inlined for entities
and relations.

(From Sierra)