Fix STK search to return correct owning element(s) - at least 99%
* Added 2D code to populate the z-direction search box since
search was hard coded for 3D. Thanks to Dan for helping with this.

* Added some "slop" to the bounding box (1% larger than max/min).

* Add check skeleton to see if we can interpolate coords back to halo
node isoPar using found isoPar and owning element.

* Cleaned up search, is_in_element() and interpolate_point() interface.
Now, I can obtain a list of possible elements, iterate them to find the
"best match" via is_in_element() and the halo isoParCoords_. As a
verification step, I made sure that with the owningElement and isoParCoords
in hand, I could interpolate the elemental nodal coords to the halo location
and check that this interpolated value equaled the original halo coordinates.

Now, I am ready to deploy the algorithm... Thus far, this is in a really
nice form to share with all Conchas codes..

* Added check to see if master element was null - indicated no element was found
This is a problem, however, in my case, it was due only to the fact that the
search criteria was not broad enough.

* More debug info to figure out why elements are not beeing found

Notes. 99% of the halo nodes are being dealt with appropriately. However,
I still have some problem elements..

* Added different bounding box max:min for z-comp in 2D sims.
Still missing some elements. I think I need to push this to
James or Dan.

* Added dial for OCTREE-type alg; added output for checked haloCoord via isopar

* Yet more diagnosis when a node is not found well

* Dealt with memory issues with passing a haloNodalCoord nDim as apposed to the hard coded NDIM = 3

(From Sierra)