PyTrilinos now builds properly if EpetraExt or IFPACK are disabled
This is in response to bug #5511.  One night the EpetraExt package
failed to compile and so was disabled.  PyTrilinos failed in turn to
build, which is the subject of the bug.  The fix includes ifdef-s that
check both EpetraExt and IFPACK, applied to the NOX.Epetra and ML
PyTrilinos modules, as well as test and example scripts.