Update excluded packages for s903186 clang builds
Disabling SEACAS since it has several subpackages that are exclusively fortran.
The initial pass of just not building the code with a warning was insufficient
since there are subpackages that depend on the now effectively disabled fortran
subpackages which can't tell the fortran subpackage was disabled. Need to come
up with a better solution than just disabling, but for now I think this is
appropriate.

Re-enabled Zoltan, STK and Sundance since it is likely that what ever issues
were preventing them from building were compiler specific. Even if they aren't
it would be good to know what the issues are with the clang compiler.

Left PyTrilinos disabled since we aren't building shared just yet.