Change PartitioningSolution template parameter from a user object to an input adapter.
From the user point of view, this is more consistent.
Now both problem and solution are templated on the same
object.

If we continued to template on the user object, we would
have to have a second template parameter for the scalar
data type.  (Some user objects like CrsGraph don't have a
native scalar data type.)