Ifpack2: added some unit tests
The two first unit tests demonstrate that Relaxation::apply(x,y) can
be used with x==y. I wrote these tests because I want to modify the
way x==y is tested by apply() to fix a bug (see below).

The third unit test is failing and demonstrates that Ifpack2 is not
working when one node do not have any local element. An exception is
thrown when apply() is checking if x and y are pointing to the same
memory location. In MueLu, we are gathering the matrix to a subset of
processor when the problem becomes small to avoid communication
costs. The communicator is still global so some nodes do not have
local element.