update s903186 testing
added a stable clang build to the nightly track. This build enables all the same
packages as the experimental build, but disables all the tests for the packages
that have failing tests. Bugs have already been filed for these failing tests.

Reordered the tests so that the clang builds are first. This is mostly to make
sure that the nightly clang build is ready at the start of the day.