Add Environment::getDebugLevel().
This is in addition to Environment doStatus(), which returned
true if the debug level was anything other than "no_status".  In
RCB we were created detailed status messages like so:

if (env->doStatus(){
  // tons of work
  env->debug(DETAILED_STATUS, msg);
}

So even if the debug level was just BASIC_STATUS, all of this work
was done.  I knew this and thought I would fix it later, but I didn't
realize how costly one of the messages was to create.  For example,
partitioning 4M values in serial into two halves went from .131 seconds
to over 3 seconds.

Debug message that take time to create should be done this way:

if (env->getDebugLevel() >= DETAILED_STATUS) {
  // tons of work
  env->debug(DETAILED_STATUS, msg);
}