SEACAS: update exodiff cmakelists.txt
The build files in Trilinos were out of date with the changes that had been made
to exodiff in sierra. To fix this I made the exodiff CMakeLists.txt more like
the Sierra Jamfile for exodiff which is to glob everything into the set of
sources to build the exodiff exe.