Update trilinos-test2's boost and netcdf versions
Updated most of the builds on trilinos-test2 to use boost 1.49.0 and netcdf 4.2

The installation build was not updated because it requires more significant
changes to the scripts than I feel comfortable doing. The two packages that this
would affect the most, STK and SEACAS are explicitly disabled for the
installation testing so I don't believe that this will cause too many issues.
The root issue is that the scripts that are needed to do the installation
testing only support a small subset of tpls which does not include hdf5. HDF5 is
now needed by netcdf 4.2 which means an option to support hdf5 is needed.