Update Zan's boost and netcdf version Add HDF5
Updated boost to 1.49.0 and netcdf to 4.2. The addition of netcdf required
adding hdf5 1.8.6 as well.

I did not update the boost for the intel builds because it looked like that was
using a compiled boost specific to the intel compiler. However the package that
needs the updated boost, STK, is explicitly disabled for those builds so I don't
believe this will be an issue. It will be addressed at a later date though.