Kokkos: Disable a couple tests on windows.
The tests that were disabled yesterday where the wrong ones. I mistakenly looked
at the file that was causing issues and assumed it was the test in that
directory that used said file. however the actual test that is failing is in a
different directory and just builds against that file too. This disables the
correct cmake target that is failing to build.