Changing logic so that base build dir only added to includes if repo configures version file.
Before, TriBITS would always add an include path for the base project
directory in order to handle the configured Trilinos_version.h file.  When
moving to repos this broke when Trilinos is not the base project so an include
directory was added in Teuchos.

Therefore, to keep backward compatibility and not insert the include directory
for all TriBITS projects, I added the INCLUDE_DIRECTORIES(...) command in the
CMake code that configures the version header file.  To make this work, I had
to turn the functions into macros so that it would append the include path
everywhere.  It may turn out that you don't want to do this and just let the
packages add the include directory as needed but we will see what people say
(if anyone notices).

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Packages
0) MPI_DEBUG => FAILED: passed=1181,notpassed=2 => Not ready to push! (89.31 min)
1) SERIAL_RELEASE => passed: passed=1158,notpassed=0 (49.08 min)
WARNING: Forced the push!