Removing old Trilinos/TriKota build pulling from SVN repo (VRI Kanban #2643)
Now that there is a VERA-based build of TriKota from within VERA pulling from
a stable snapshot of Dakota updated manually, we don't want or need the SVN
pull of TriKota.  Instead, we someone to manually update Dakota within VERA as
needed for now.

Therefore, this commit turns off the Trilinos-based build of TriKota where
Dakota was being pulled from the SNV repo of Dakota.