Fixed abs() and fabs() for nested AD types.
I use the ternary operator ( conditional ? true_exp : false_exp ) for
the implementation of abs() and fabs().  However both results of the
ternary operator have to be the same type.  This wasn't the case when
the scalar type for CacheFad was itself a Fad type.  This fixes that.
Thanks to Jedidiah Gohlke for bringing this to my attention.