Panzer: Hardening of DOFManager and STK code to deallocation of MPI_Comm
This is a result of an MPI_Comm that is passed in being removed by
the user (MPI_Comm becomes null), and STK still actively using
that communicator. This commit tries to allow Panzer to protect
itself in that case by first building a Teuchos::MpiComm and storing
a pointer to it. Then the RawMpiComm is used to gurantee that the
MPI_Comm handle is still active.