Panzer: Changing the assembly engine to use the workset container
This will facilitate more direct use of the workset container, and
hopefully more flexibility in the application of the AssemblyEngine.
In particular moving it to be use by the response library, thus unifying
assembly.

I've also modified many of the tests to account for this change. To this
point all tests pass and Panzer and Drekar with no Stokhos.