1) Toward support for a non-intrinsic 'value_type'.
Differentiate between the 'value_type' and 'scalar_type' of a View
so that embedded non-intrinsic types can be supported; e.g.,
arrays-as-scalars, automatic-differentiation, or uq-expansion types.

A view type 'KokkosArray::View< ValueType**[N][M] , device >'
is recursively analyzed via the KokkosArrayImpl::AnalyzeShape< T > template.
Dimensions and pointers are extracted to build up a
KokkosArray::Impl::Shape type to describe the shape of the array.

If the 'ValueType' is an embedded non-intrinsic type then the
KokkosArray::Impl::AnalyzeShape< ValueType > should be partial
specialized to analyze the array-shape of that type.
For example, if ValueType == Array<double,N> then the shape
adds a dimension 'N' and defines the scalar_type to be double.

For an ordinary view 'value_type' and 'scalar_type' are identical.

2) Consolidation of the "when CMake is unavailable" build scripts
   and selection of more compiler warnings.