Teuchos: Type conversions: Fixed stringToReal test failures by ensuring that when printing a double resp. long double to a string (for testing conversion from string to a floating-point type of lower precision), the test outputs the correct number of decimal digits.
For example, when printing std::numeric_limits<float>::max() to a string as a
double, it's important to use the full 17 digits that double requires (rather
than the 9 for float), since the 9-digit value printed as a double might round
to a double too big to represent as a float.

Similarly, we need to use the full 36 digits for long double (assuming the
maximum binary128 precision, which it has on some platforms, e.g., IBM).