Panzer: Adding square quad test for new DOFManager
This is a recreation of the previous test, rebase lined to work
with the new DOFManager. I also renamed the FEI version of the
test, hopefully it will be easy to optionally excluded it in
the future.

Also fixed a few issues in the new dof manager as a result
of the testing.