Porting to the Intel-Phi (a.k.a., Intel-MIC) required revising concept of a manycore "node."
The previous definition associated "node" with NUMA region,
which the Intel-Phi does not have.  The revised definition of a "node"
is the hardware's collection of processing units which have memory affinity.
On a NUMA device a "node" is a NUMA region.
On the Intel-Phi device a "node" is a core with its four hyperthreads.

This revised definition allows KokkosArray to flexibly interpret
a manycore device to have a two-level topology based upon memory affinity.