Disable STK for the stable clang build
certain parts of STK appear to need a fortran compiler to build. I recently had
to fix the cmake files to allow STK to properly determine the fortran mangling
and that change is causing STK to fail to configure on the clang build. Since
the underlying issue is that parts of STK need a fortran compiler and the clang
build by definition doesn't have a fortran compiler I am just disabling STK for
right now. This needs to be addressed in a better way, but I won't have time to
look into it and leaving it broken is not good it has already been broken too
long.