Tpetra: CrsMatrix: Implemented nonlocal sumIntoGlobalValues().
One desirable feature for CrsMatrix is for sumIntoGlobalValues() to
work for rows that are not owned by the calling process.  This relates
to Bug 5782, because allowing nonlocal sumIntoGlobalValues() calls is
one way to address performance concerns for global assembly.

This commit includes both a unit test for the desired feature (which
I implemented first, according to the principle of test-driven
development) and the feature itself.  It required only modest
modifications to the implementation of CrsMatrix.