Tpetra: CrsMatrix: Reconciled nonlocal sumIntoGlobalValues semantics.
sumIntoGlobalValues() and replaceGlobalValues() calls for owned rows
silently ignore columns that aren't in the (static) graph.  In order
to reconcile this behavior with the new feature of sumIntoGlobalValues
for nonowned rows, I changed the unit test so that the latter does not
throw an exception if the nonowned rows have column indices that are
not in the (static) graph on the processes that own those rows.

If someone later wants {sumInto, replace}GlobalValues() to throw an
exception in the owned rows case, they should ensure that these
methods behave in the same way for nonowned rows.  (The exception to
throw in that case is Tpetra::Details::InvalidGlobalColumnIndex.)

I have not yet changed the name of the unit test or its documentation.
This will happen in my next commit.