Make calore stop using the 'connect-count' attribute on stk-mesh entities.
Calore was the only user of that attribute. Now we'll be able to remove
that attribute, saving an int for every entity.
Calore now sets up bulk-nodes more similarly to the way Aria does it,
creating a particle-element with a relation to the bulk-node to let
fmwk and stk-mesh know that the bulk-node shouldn't be deleted.
The particle element is inactive and shouldn't be written out to any
results files.

(From Sierra)