Fixed PyTrilinos build error
For some reason (perhaps a cmake upgrade), a semicolon was getting
inserted into the link command for libpytrilinos.dylib.  This was
apparently because the EXTRA_LINK_ARGS list variable had a blank first
element.  I changed the EXTRA_LINK_ARGS variable to a single string
instead of a list, and this fixed the problem.  I don't see any reason
it wouldn't be backwards compatible.