Adding APPEND_STRING_VAR_EXT() to deal with [, ], {, }, and ; (VRI Kanban #2818)
After a lot of annoying debugging, I found that the special chars [, ], {, and
} affect how cmake deals wiht list interpretation with ';' chars.  What a pain
in the rear!  This totally broke the function APPEND_STIRNG_VAR().  I
documented this behavior in a unit test and then I wrote the function
APPEND_STRING_VAR_EXT() to deal with them correctly, but only for a single
argument.

This is needed to implement a cmake -P script printing a python datastructure
using the MESSAGE() command needed to implement the --extra-repos-file option
with the checkin-test.py script (see VRI Kanban #2818).