Fixed Teuchos::ParameterList parsing error.
When passing in the "num parts" parameter to
Isorropia::Epetra::Partitioner I was getting the run-time error:

Error!  An attempt was made to access parameter "num parts" of type
"int" in the parameter (sub)list "ANONYMOUS" using the incorrect
type "string"!

The reason was the code that converts parameters to upper case was
trying to convert all parameters, not just the string parameters.  So to
fix this I had it only try to convert string parameters.