Piro: Added more liberal version of solve driver
The PerformSolve driver is designed to handle response-only model
evaluators.
Accordingly, the function accepts ResponseOnlyModelEvaluatorBase as
input argument in order to enforce as much as possible that constraint
at compile-time (All Thyra-based Piro solvers respect that convention).

However this is an overly restrictive requirement since the function
only uses the ModelEvalutor interface.

Advanced users are therefore given the option to use an 'unsafe' version
which accepts any ModelEvaluator at compile-time.
In particular, this loophole is necessary to handle wrapped Epetra-based
response-only model evaluators.