Got a version of building Cijk from scratch working.
This only works for a total order basis with lexicographic ordering.
There are still some todo's to make it more efficient, but it builds the
sparse Cijk tensor by iterating only over the non-zeros in the 1-D
sparse Cijk tensors and requires no map lookups to map multi-indices to
global indices.