Changing the Graph class.
Motivation:creating a graph via fillComplete for aggregation purposes is very expensive, and I want
an option to do something faster.

This checkin introduces an interface class, GraphBase.  The existing class Graph and the new class
LWGraph inherit from GraphBase.  This allows us to do whatever we like for the innards of the Graph
-- reuse A's graph, build a new amalgamated graph, build a local graph, etc.

LWGraph is a "local information" graph.  Instead of fillComplete, it uses row and index arrays for
the local information, and the original graph (A's graph) for communication (maps, import, etc.).

Creating an LWGraph is an experimental option in CoalesceDropFactory right now, and it only handles
matrices from scalar problems, with and without dropping.  I've hard-coded the class SA dropping
right into the CoalesceDropFactory for the moment.

CoalesceDropFactory's existing functionality is untouched, and is the default.

TODO:

    Implement proper dropping function
    Implement handling for multiple DOFs with and without dropping.
    Implement some unit tests.

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=67,notpassed=0 (6.56 min)