Teuchos: TypeNameTraits: Fix for Bug 5845.
The IBM XL C++ compiler relapsed into a bug that it had fixed before,
so we had to update the set of compiler version numbers to which we
apply the work-around.  Thanks to Jim Willenbring, Mike Glass, and
Paul Lin for pointing this out and offering a patch.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => passed: passed=33,notpassed=0 (6.51 min)
2) SERIAL_DEBUG => passed: passed=33,notpassed=0 (6.46 min)
Other local commits for this build/test group: 17684b3, 303fbd6