Ifpack2: Fixed NaN in Condest for Chebyshev preconditioner test.
Ifpack2's test for all the preconditioners was, for some reason,
producing a condition number estimate of NaN for the Chebyshev
test.  I added a whole lot of debugging code to the Chebyshev
implementation and found out that the residual B - A*X with B and
X both zero had inf-norm NaN.  That indicated either a problem with
Tpetra or a problem with the Matrix Market file reader.  Then, I
found out that the test was using a hand-rolled Matrix Market
reader instead of the one I had carefully implemented in Tpetra
a couple years ago.  I replaced the hand-rolled one with my reader
and the problem went away, yay!

Note: The Chebyshev implementation itself was fine.  This was a
problem with the test, not with the code being tested.

Lesson: Please don't roll your own sparse matrix file reader.
I'm not bragging on mine, but we should agree on one and just
fix the bugs there, rather than try to track down random bugs
here and there in random hackish file readers.  Just because a
file format is ASCII doesn't mean it's trivial to parse.

This commit includes all the debugging code in the Chebyshev
implementation.  If Teuchos_ENABLE_DEBUG is set, the debugging
code will spew forth a vast amount of debugging output to stderr
on all processes in the communicator.  I will disable the debugging
code in my next commit.