Tpetra: RowMatrixTransposer: Deprecated current unsafe constructor.
RowMatrixTransposer's current constructor holds a reference (not
an RCP, just a "const CrsMatrix<...>&") to the input CrsMatrix.
This is unsafe, because the input matrix might fall out of scope
between calling the constructor and calling createTranspose, thus
invalidating the reference.

I've fixed this by
1. adding a new constructor that takes an RCP<const CrsMatrix<...> >,
2. storing an RCP<const CrsMatrix<...> > in the instance, rather than
   a "const CrsMatrix<...>&",
3. deprecating the old constructor.

The commit will cause a large number of build warnings, due to use
of the deprecated constructor.  I will fix this in subsequent
commits.