Teuchos: Fixed strange compile error on Mac with Apple's gcc 4.2.1.
The recent work-around for the bug in Apple's vecLib (in particular,
its BLAS implementation having the wrong ABI) seems to have triggered
a compiler issue.  Bill Spotz discovered this on his Mac laptop.  The
compiler was complaining about the close parenthesis in an
"#if defined(...)" construct, so I replaced it with an #ifdef and
also deleted a comment that appears to have confused the compiler.
This fixed this issue.  I don't feel good about this fix because it's
not clear why it works or why the old code broke the compiler, but at
least it works now.

Build/Test Cases Summary
Enabled Packages: TeuchosNumerics
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => passed: passed=15,notpassed=0 (0.60 min)
2) SERIAL_DEBUG => passed: passed=15,notpassed=0 (0.52 min)