Kokkos: DefaultArithmetic: Specialized Init and Assign for SerialNode.
Matt Bettencourt reported that Default::Arithmetic::Init was taking
quite a bit of time for Node=SerialNode.  This commit preemptively
optimizes this special case, which matters for the ASC L2 Milestone
for Sierra / Trilinos integration.  The issue could be that older
versions of gcc (like gcc 4.4.x which we're using for current
Milestone benchmarks) do a bad job of inlining.  Kokkos' parallel_for
will only be efficient if inlining works correctly.

This specialization bypasses Kokkos' parallel_for for two critical
kernels:

1. Init: initial fill for a Tpetra::MultiVector
2. Assign: deep copy of a Tpetra::MultiVector

The rest of the kernels just copy the old implementation.