Kokkos: DefaultArithmetic: Added macro for 'restrict' keyword.
I'm only using this for the Node=SerialNode partial specialization.
The macro KOKKOSCLASSIC_RESTRICT is defined based on the compiler
type: __restrict for gcc, restrict for Intel, and nothing for any
other compiler.  We should really put this into a CMake test, but
that will for sure break Sierra's bjam for Trilinos and make Matt
Bettencourt's life hard again.