Stokhos:  Split computeTripleProductTensor() into 2 functions.
Previously computeTripleProductTensor() took an arguement specifying the
limit of the k index, which is primarily used for linear problems.  This
was problematic as it made the Cijk entries incorrect for tensor product
and Smolyak bases, since it implicitly implied a total order restriction
on k.  I split the function into computeTripleProductTensor() and
computeLinearTripleProductTensor(), neither taking any arguments.