Ifpack2: Details::Chebyshev: Moved assumeMatrixUnchanged to param.
As promised in my previous commit message, this commit moves the
optional assumeMatrixUnchanged argument of compute() to an optional
Boolean parameter in the ParameterList input of setParameters().
The new parameter is called "chebyshev: assume matrix does not change".
It is safe to set this on first call to setParameters() before calling
compute(), because compute() will always compute any missing
quantities.  ("Missing" means that the user did not provided them in
the ParameterList input of setParameters(), and that they have not
been computed by the compute() method.)

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=21,notpassed=0 (5.55 min)
1) SERIAL_RELEASE => passed: passed=17,notpassed=0 (3.69 min)
Other local commits for this build/test group: cfa1a04