Kokkos: Fixed bugs in triangular solve kernels used by AltSparseOps.
The bugs were only showing up for complex arithmetic.  I've set
Teuchos_ENABLE_COMPLEX:BOOL=ON in my check-in test script, but for
some reason it's not enabling complex arithmetic tests in Kokkos.
Thus, I had to rely on Dashboard test failures to diagnose the
problem, such as the following:

http://testing.sandia.gov/cdash/testDetails.php?test=12184704&build=911771