Adding --abort-gracefully-if-no-changes-to-push and other refactorings (Trilinos #5858)
This commit combines all of changes necessary to add the new option
--abort-gracefully-if-no-changes-to-push to the script checkin-test.py and
update all unit tests (see Trilinos #5858).  This is critical to support the
usage of the checkin-test.py script for Almost Continuous Integration sync
servers with multiple git/TriBITS repos (see
tribits/doc/HOWTO.ALMOST_CONTINUOUS_INTEGRATION).

As you can see from the diffs, this was a fairly small change to the
production code in checkin-test.py and CheckinTest.py.  The big change was
that instead of checking if the various git repos had changes at the very end
before the push, that is checked just after all of the pulls, thereby allowing
the option --abort-gracefully-if-no-changes-to-push to skip the various
builds.

The other significant change was that the unit tests in
CheckinTest_UnitTests.py used to require a seprate 'cat modifiedFiles*out'
command to match the 'eg diff ...' command which could be different.  In fact,
many unit tests had inconsistent pairs of 'eg diff ...' which wrote
modifiedFiles*out files and intercepted commands 'cat modifiedFile*out'
commands which avoided reading them in.  There is no need to intercept these
'cat modifiedFiles*out' commands so I just replaced them with python file
reading code and avoided the system 'cat' command.  Therefore, the tests are
now stronger, more consistent, and slightly easier to maintain than they were
before.

However, making what should have been a very simple change turned out to take
about 4 hours to implement because of the difficulting in a) how I did the
refactoring (I used a bad refactoring path) and b) the nature of the full
system tests in CheckinTest_UnitTests.py.  While it has always been difficult
to change behavior of checkin-test.py due to the strong nature of these tests
(which require listing almost all of the commmands in order before running),
this recent experience was about the worst I have ever had.  I would say that
the ratio of time fixing/modifing tests vs. adding/changing production code
(including adding the new unit tests) was about 20:1 in this case.  As
mentioned above, part of this was due to bad unit tests with inconsistent sets
of 'eg diff ...' and 'cat modifiedFiles*out' comamnds.  If I would have just
gone straight to eliminating the usage of 'cat modifiedFiles*out' first and
then updating the unit tests, I think this refactoring would have taken 1/10th
as long.  It is easy to remove commands from the command chain in these tests
but it is very had to add new commands of the right form.

I need to think long and hard about if we need to refactor the software and
tests to take greater advantage of unit testing instead of full-script testing.