Removing ETI support for these trivial inline exception classes (Trilinos #5861)
There is a use case in DataTransferKit that uses unsigned int for local
ordinal and this ETI code did not use the new OCP ETI system.  I floated this
with Mark Hoemmen and he signed off (see commit in Trilinos #5861).

I gutted the Tpetra_Exceptions_def.hpp and Tpetra_Exceptions.cpp files but did
not remove them so as to maintain what should be 100% backward compatiblity.