Tpetra: ImportExportData: Got rid of "friend" declarations.
ImportExportData is just an implementation detail of Import and Export,
so there is no need for elaborate access controls ("protected" etc.)
with "friend" declarations for those two classes.  Instead, I removed
the "friend" declarations, and made ImportExportData a "struct" (which
in C++ means that everything is "public" by default).  I'm hoping that
this will improve compilation speed a bit, since there are fewer
access controls for the compiler to evaluate.  It also reduces the
number of lines of code.

I also added documentation for the fields of ImportExportData.  This
includes notes on how possibly to optimize by deallocating data after
they are not needed.

Finally, I corrected a grammatical error in the internal documentation
of Import.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=44,notpassed=0 (15.72 min)
1) SERIAL_RELEASE => passed: passed=34,notpassed=0 (14.39 min)
Other local commits for this build/test group: a9884d9