Kokkos: DefaultNode: Changed the default Node type to SerialNode.
We were getting a lot of complaints about this.  We had left it the
way it was for as long as we did, because we wanted users to get
hybrid parallelism by default.  However, people are doing a lot of
performance tests now, and it's confusing that people who let Tpetra
pick the Node type (thus, the default type) are getting different
performance results than people who specify SerialNode as the default
Node type.

From now on, if you want (CPU) threads within an MPI process, you
need to specify TPINode, OpenMPNode, or TBBNode explicitly as the
Node type.  I've left the include files for all of these in
Kokkos_DefaultNode.hpp, so users won't have to add include files.

Build/Test Cases Summary
Enabled Packages: KokkosClassic
0) MPI_DEBUG => passed: passed=20,notpassed=0 (1.28 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (2.26 min)