Added "director" feature to several PyTrilinos.LOCA classes
The director feature enables cross-language polymorphism so that the
C++ LOCA solvers can call back to Python for function evaluations,
etc.

Note that these do not compile yet.  There are conflicts with LOCA
arguments and SWIG director variables both named 'result'.  I fixed
this problem in NOX with a '#define result nox_result', but this trick
is not currently working in LOCA.