Piro: Enabled implicit upcasting
It is both safe and convenient to have RCP<U> to be converted to a
sharing provider of objects of type T as long as the implicit conversion
from RCP<U> to RCP<T> is valid.

The above conversion cannot not be automatically inferred through the
implicit template argument deduction mechanism. An explicit function
template argument specification is therefore necessary.