Tpetra: Distributor: Added optional use of MPI tags (off by default).
When we took out the barrier between receives and sends in Distributor,
we observed on some platforms that this caused MPI to abort.  I suspect
that this has something to do with sends and receives not matching up,
or at least their buffer lengths not matching up.  To help rule out the
first possibility, I've added MPI message tags to Distributor.  This is
disabled at compile time by default.  You can enable it by defining
TPETRA_DISTRIBUTOR_TAG_COUNTER at the top of Tpetra_Distributor.hpp.

Enabling the tags option currently causes the Map unit test to hang on
my machine (Mac, Clang 3.2, OpenMPI 1.6.3) in an MPI build, even with
the barrier on, so there is definitely _something_ wrong with
Distributor.