Fixing segfault with coordinate projection.
If a process has no rows, then the block size cannot be calculated locally by dividing by #local elts.
I've introduced a global reduction (max) to handle this.  An alternative is to stick this on the level.
Fyi, block size came up in the context of the SaPFactory a while ago.  What I don't recall is whether
we decided it was OK to put this on the level.

Build/Test Cases Summary
Enabled Packages: MueLu
Disabled Packages: MOOCHO
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=83,notpassed=0 (9.60 min)