Fixing issue with zero/nonzero initial guess.
When using a single level method (i.e., smoother only), the arguments to the coarse level
pre and postsmoother solves were hardwired for a nonzero initial guess.  This commit fixes that.
This issue could also crop up in the multilevel context, as the coarse grid solve would do
extra operations that could otherwise be skipped for a zero initial guess.

Thanks to Eric Cyr for finding this with careful comparisons to ML, etc.