Fixing nnz estimate for matrix matrix addtion.
During a code review with Andrey, we noticed that the nnz estimate for
the result of a matrix matrix add was hardwired to 5.  There was a comment
that this was wrong -- which is absolutely true.

TwoMatrixAdd now has an additional boolean parameter that can be set to true
if the caller knows the first matrix has a fixed nnz per row.

I've also added some logic that does the following for C = A+B.  The point is
to create C with static profiling, which is faster and more efficient
than dynamic profiling.

1) If either A or B has at most one nonzero per row, or the caller has indicated
   that A is has a fixed nnz per row, then we know the nonzeros
   per row in C within +-nnzPerRow(A), so C is created with static profiling.

   This is the case for Ptent + Dinv*A*Ptent when Ptent is based on a nullspace

2) Otherwise, an estimate E is created for a fixed nnz per row  in C.
   If E > maxNnzInAnyRow(A) + maxNnzInAnyRow(B), then C can be created
   with static profiling.  Otherwise C is created with dynamic profiling.

I've tested this with 2D Poisson and Elasticity, and it seems to do the right thing.

Build/Test Cases Summary
Enabled Packages: MueLu
Disabled Packages: MOOCHO
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=83,notpassed=0 (10.59 min)