MueLu: Fixed compiler warnings.
This is a much less aggressive fix than the one I committed a few
commits ago and then reverted because it caused a test failure.
If this fix fails, then there is definitely a bug in MueLu, since
all I did this time was make explicit the integer casts that were
being done implicitly anyway.