Panzer: fix stk mesh unit tests
We have occasionally seen random failures from the trilinos test harness for the stk line and hex unit tests.  It looks like this is a race condition where both test were trying to write to the same output file at the same time.  I changed the names of the output files for all stk unit tests to make sure they don't collide.  Hopefully this fixes things, but I could not reproduce the failure on my machines, so we will have to watch the nightly testing...