Ifpack2: Relaxation: Fixed apply() for alpha != 1.
The apply() method of Ifpack2::Relaxation only worked for alpha == 1
and beta == 0, and silently computed the wrong answer in other cases.
I fixed the alpha != 1 case, and added a check for beta != 0 (which is
a bit tricker to implement, so I am deferring that task for now).

Build/Test Cases Summary
Enabled Packages: Teuchos, Kokkos, Tpetra, Belos, Ifpack2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=243,notpassed=0 (4.67 min)
3) SERIAL_RELEASE_SS => passed: passed=224,notpassed=0 (7.95 min)