Belos: Specialized MultiVecTraitsExt for Tpetra, to fix 64-bit issue.
Travis Fisher pointed out an issue with Belos' Tpetra specialization
for a linear system with over 2 billion rows, for which the number of
rows cannot fit in a (32-bit) int.  The issue popped up in the DGKS
orthogonalization method, but would also show up in ICGS and presumably
in other orthogonalization methods as well.

The MultiVecTraitsExt solution for the >2B rows problem (devised by
Heidi Thornquist, Chris Baker, and myself a few months ago) is only a
generic solution for backwards compatibility.  Solving the >2B rows
issue still requires a nondefault specialization of MultiVecTraitsExt
for the multivector (MV) type.  We had neglected to write that
specialization for Tpetra::MultiVector.  This meant that we had not
actually fixed the >2B rows problem for Tpetra, we had just enabled a
future fix.

This commit is that future fix.  It adds a specialization of
MultiVecTraitsExt which keeps the global number of rows a 64-bit
integer (on 64-bit machines).

Build/Test Cases Summary
Enabled Packages: Teuchos, Kokkos, Tpetra, Belos, Ifpack2
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=243,notpassed=0 (6.33 min)
3) SERIAL_RELEASE_SS => passed: passed=224,notpassed=0 (4.47 min)