Tpetra: Map: Added removeEmptyProcesses method.
This method returns a new Map with a distinct communicator from the
original Map's communicator.  The new communicator only contains those
processes in the original communicator that own a nonzero number of
elements in the original Map.  The method leaves the original Map and
communicator unchanged.

This commit also fixes syntax errors in the following commit:

bf3d08d924b96e55389fc347add8e59431c37bdf

I am not able to revert that commit because of an intervening merge.
This commit also changes DistObject's new method from
filterEmptyProcesses to removeEmptyProcesses.  Thanks to Chris Siefert
for suggesting a name.