Ifpack2: Relaxation: Added new diagonal entries check.
I had previously added the "relaxation: check diagonal entries"
parameter to Relaxation, which does additional checks on the
diagonal entries of the matrix (for example, counting the number
of entries that are zero, negative, or small).  This commit adds
one more check: that the computed inverse diagonal is not very
different from the "actual" inverse diagonal (i.e., the result of
1/diagonal, without any fancy zero or small entry detection).
The check costs one more all-reduce, but the parameter is false by
default, and the information should be valuable to people who want
to diagnose issues with Jacobi, Gauss-Seidel, or Symmetric Gauss-
Seidel.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=170,notpassed=0 (5.93 min)
3) SERIAL_RELEASE_SS => passed: passed=153,notpassed=0 (1.69 min)