Ifpack2: LocalFilter: Removed excess state, and improved public docs.
I'm revisiting LocalFilter, because for the Milestone review, we should
be able to say that incomplete factorizations work as subdomain solvers.
Ifpack2 uses LocalFilter to restrict a matrix to locally owned rows and
columns, so that it is legitimate to use a linear solver (the subdomain
solver) on that matrix.