Added ${PROJECT_NAME}_DISABLE_ENABLED_FORWARD_DEP_PACKAGES_DEFAULT to allow project override.
With this a project can redefine the default value for
${PROJECT_NAME}_DISABLE_ENABLED_FORWARD_DEP_PACKAGES.  Trilinos has decided
that this should be OFF by default (thus maintained) but VERA needs to have
this ON by default.  That is the driver for this change.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: CTrilinos,ForTrilinos,PyTrilinos,Didasko,Mesquite,Phdmesh,Pliris,Claps
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=104,notpassed=0 (1.81 min)
1) SERIAL_RELEASE => passed: passed=104,notpassed=0 (1.10 min)