Teuchos: CommHelpers: Added specializations for unsigned integer types.
Tpetra Import and Export initialization depend on sending and receiving
data of type size_t.  Other Trilinos packages might also.  Teuchos did
implement sending and receiving size_t and other unsigned integer
types, but did not have specializations of functions like send() and
ireceive() for these types.  The specializations are faster because they
bypass Teuchos' generic serialization code and use built-in MPI_Datatype
directly.  This commit adds specializations for some communication
methods, for unsigned versions of long long, long, int, and short.

I didn't add specializations for size_t, because size_t is a typedef to
some built-in unsigned integer type which depends on the platform.  The
built-in MPI_Datatype MPI_AINT (corresponding to MPI's typedef MPI_Aint)
is for a signed integer type, so it would be good for ptrdiff_t but not
for size_t.  The logic for mapping size_t to its implementation type is
complicated (sizeof(unsigned long) == 4 and sizeof(size_t) == 8 on
Win64, whereas on most other common operating systems, long and size_t
have the same size).  Thus, I omit this logic in this commit, and just
try to get coverage on Linux and other *nix-like platforms (such as Mac)
for now.  The generic implementation will provide coverage on Win64 for
the functionality, if not for maximum performance.