Panzer: Updating Epetra model evaluator to use 2nd gen RL interface
Previously the "Aggregator" interface in the ResponseLibrary was
being used. This change removes that usage and now used the new
interface. A side effect of this is that the model evaluator now
allows you to add responses after construction. Though you must
still explicitly build the responses. The ModelEvaluatorFactory
takes care of all of this. See the main_driver.cpp for example
usage.

A number of small changes fixing input arguments was also made to
the response library (proper). And the functional response was modified
to make it more flexible.

The main driver has been modified to automatically add in the "Responses"
that use the functional evaluator factory. Additionally the input deck
has been changed reflecting this. Finally the IP Coordinates response is
no longer excercised in the driver because it does not support the model
evaluator interface required for model evalutator compatible responses.