Teuchos: CommHelpers: Fixed compile error.
For some reason, in Teuchos_CommHelpers.cpp, my compiler thinks that
HAVE_TEUCHOS_DEBUG is not defined, regardless of whether I've set
Teuchos's debug mode.  This is not true on other machines.  Thus, my
machine let the code through, which was using an undefined variable.
I've fixed the code by removing the undefined variable (which was in
a sanity check anyway, not in semantically meaningful code).

Build/Test Cases Summary
Enabled Packages: Teuchos
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG => passed: passed=103,notpassed=0 (2.06 min)
3) SERIAL_RELEASE => passed: passed=100,notpassed=0 (2.10 min)