Tpetra: MatrixMarket: Fixed CrsMatrix reads for nonzero index base.
The following commit says that it fixes this, but actually does not:

9762002d433ce198cf371e35f4030687cea20933

This commit actually fixes the issue.  It also includes a unit test
which exercises the nonzero index base case.

Users will only encounter this issue when reading in a CrsMatrix using
a row Map that they provide, rather than letting the reader create a
row Map for them.