Panzer: Refining approach to enabling/disabling responses for different evaluation types
Before the lone control was if the builder returned null, that was a compile time option.
Now a particular type can be enabled at run time if the user implements the "typeSupported"
method in the ResponseEvaluatorFactory (its pure virtual so they must implement it!). This
allows a little better control over what is evaluated and what isn't. This change updates all
the panzer responses and also the response library. The evaluation of jacobian types is also
tested (though not yet complete) for the functional response.