Tpetra: Matrix Market I/O: Added a test for values indep of structure.
I added to the CrsMatrix Matrix Market Reader / Writer test an extra
test for equality of the values in the matrix, independent of the graph
structure of the matrix.  It's not fully general, because it still
expects the column Maps to be the same, but if this becomes important,
it could be fixed via an Import.