Tpetra: CrsGraph: Simplified mergeRowIndicesAndValues().
This method is only called by CrsMatrix, and only when the matrix owns
the graph, not when the matrix was constructed with a const graph.
The method used to take two template parameters: an iterator type Iter
(the type of the iterator over the matrix's values in that row) and a
binary function type BinaryFunction (over the type of the matrix's
values).  These template parameters were confusing and redundant for
the following reasons:

  - Neither Iter nor BinaryFunction actually specified the type of the
    matrix's values (Scalar).  This made it hard to understand what
    the method was doing and how it was relevant to CrsMatrix.

  - CrsMatrix was only ever calling the method with Iter = typename
    ArrayRCP<Scalar>::iterator, and BinaryFunction =
    std::plus<Scalar>, where Scalar is the type of entries in the
    matrix.  Thus, the only relevant template parameter is Scalar.

  - The method does not need to accept an ArrayRCP, since it does not
    retain a reference to the row's values.  An ArrayView is perfectly
    sufficient, and equally safe (a debug build will still test for
    out-of-bounds array accesses).

Thus, I simplified the method to take only one template parameter: the
Scalar type (of entries in the matrix).  It also now accepts an
ArrayView<Scalar> for the row's values, instead of an iterator to the
beginning of the row.  This may improve compilation time as well as
performance of the method (since there are no unnecessary ArrayRCPs),
but mainly it makes the method easier to understand.