Tpetra: Matrix Market I/O: Improved "automatic" CrsMatrix I/O test.
The test now specifies an upper bound on the number of MPI processes,
which should prevent failures if tests run with a larger number of
processes that on my workstation.  It also now works harder to compare
matrices' structure and values, and provides more helpful diagnostic
output on test failure.