Tpetra: CrsGraph: Removed template params from private insert methods.
CrsGraph's methods insertIndices and InsertIndicesAndValues were
templated on two enum values, for whether the input indices are global
or local, and whether the method should insert them as global or local
indices.  This templating had no performance benefit, since the methods
do run-time 'if' tests on the values of the two template parameters
anyway.  Furthermore, the methods are not part of the public interface
of CrsGraph, so there is no harm in changing them.

This commit turns the template parameters into run-time arguments.
Hopefully this will help reduce compilation time of CrsGraph and
CrsMatrix (which calls these methods).