Tpetra: CrsGraph: insertIndicesAndValues now only has 1 template param.
In my previous commit, I changed the ELocalGlobal template parameters of
insertIndices and insertIndicesAndValues into run-time arguments.  In
this commit, I've replaced the IterO and IterN ("old values iterator"
and "new values iterator") template parameters of insertIndicesAndValues
with a single template parameter, Scalar (same as the first template
parameter of CrsMatrix).  IterO is now Teuchos::ArrayView<const Scalar>,
and IterN is now Teuchos::ArrayView<Scalar>.  This better reflects how
CrsMatrix uses the methods, and also makes their purpose more clear.
(It is visually apparent that Scalar is associated with CrsMatrix, but
was not at all obvious that IterO and IterN were.  The use of ArrayView
also better documents intent.)