Tpetra: Moved op<(CrsIJV,CrsIJV) out of std into Tpetra namespace.
CrsIJV represents an entry of a sparse matrix as an (i,j,v) triple.
It has a comparison operator< (const CrsIJV&, const CrsIJV&) so that
std::sort works for a random-access sequence of CrsIJV.  Before, this
operator was defined as a nonmember function in the std namespace.
This might have been causing issues on some compilers, for some reason
which I don't understand fully.  It is also bad form to pollute the
std namespace, and entirely unnecessary in this case.  I changed the
operator to be a public instance method (still operator<) of CrsIJV.
std::sort will work fine in that case as well, and this avoids
pollution of the std namespace.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=138,notpassed=0 (4.61 min)
3) SERIAL_RELEASE_SS => passed: passed=123,notpassed=0 (8.70 min)
4) MPI_RELEASE_SS => passed: passed=138,notpassed=0 (10.98 min)
5) SERIAL_DEBUG_SS => passed: passed=123,notpassed=0 (6.72 min)