Ifpack2: Chebyshev: Stop requiring domain and range Map identical.
Andrey Prokopenko pointed out that an exception test in Chebyshev was
(unintentionally) forcing the domain and range Maps of the input
matrix A to be identical objects (pointer equality).  This is
unnecessary; it suffices for the domain and range Maps to be the same
(in the sense of isSameAs).  Furthermore, isSameAs only costs an
all-reduce if the two objects are not identical (pointers not equal),
so removing the extra incorrect test for pointer equality won't affect
performance in that common case.  This commit removes the incorrect
test.

Build/Test Cases Summary
Enabled Packages: Tpetra, Belos, Ifpack2, Xpetra, MueLu
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=173,notpassed=0 (11.39 min)
3) SERIAL_RELEASE_SS => passed: passed=157,notpassed=0 (10.21 min)
4) MPI_RELEASE_SS => passed: passed=173,notpassed=0 (13.86 min)
5) SERIAL_DEBUG_SS => passed: passed=157,notpassed=0 (7.59 min)