Apply patch supplied by user Jim Westfall.
Patch alters the BlockMap2PointMap function in Epetra_VbrMatrix to
make sure no gaps exist in the GIDs of the resulting point-map.

Tried to use checkin-test.py, got rejected due to a failing ML_MatrixFree_MPI_4
test. A) I don't think that test has any dependence on the file that
this patch touches. and B) that test passes when I run it outside of
checkin-test.py.