Panzer: Adding a Tangent scalar type for parameter sensitivities
To implement parameter sensitivities I've added a "Tangent" scalar
type. I had to modify many of the gather/scatter evaluators adding
place holders (that in most cast throw an exception) for this new
scalar type. I also cleaned up the trans class a bit and started
the modification for the epetra model evaluator.