Ifpack2: Changed the type of a parameter and removed a redundant parameter from AdditiveSchwarz
Changed the type of "partitioner: overlap" to size_t universally. There was some ambiguity, it seems. Furthermore, "schwarz: use reordering" is redundant in the presence of "schwarz: reordering type". If the latter parameter is "none", then the former should be false. If not "none", the true. Thus the former is not strictly necessary.