Piro: Fixed orientation of multivector-based DgDx
In the LOCA solver, the DgDx derivative is represented as a multivector
with columns based on the map of the state vector x. That layout is
described as a row-oriented one, i.e. in gradient form, in the context
of a model evaluator.

No orientation for the DgDx multivector-based derivative was explicitly
specified, therefore the default column orientation was used, i.e. the
Jacobian form.