Make it possible to specify locations of blas and lapack on windows
bug 5895

The logic on windows made clapack a requirement and not an option for supplying
blas and lapack. It was possible to override the blas and lapack, but to do so
one had to use the TPL_[BLAS|LAPACK]_LIBRARIES variables exclusively. This still
had the downside of the configure looked for clapack and complained when it
couldn't find it which is a source of confusion.  This change checks to see if
the user has specified their blas/lapack through any of the accepted means and
if so skips the check for clapack and instead uses the conventional tpl
discovery logic.