stk::mesh  connectivity capacity growth policy good for speed and memory
A simple policy of doubling a bucket's dynamic connectivity capacity whenver
push_back would exceed it improved speed significantly during the previous
sprint.  However, under this policy, unfortunately ordered calls to
add_connectivity could cause memory growth quadratic in the total connectivity
of the bucket.  This caused a spike in memory in one of the adagio tests.

We solve this problem by tracking the occupancy of up-to-date data in the
connectivity representation vectors.  A new policy only doubles capacity
only if the ratio of current capacity to the footprint of the defragmented
(compressed) up-to-date data would not meet a desired threshold.  The default
capacity_ratio attempts to preserve the performance gains while removing the
negative impacts on memory use.

(From Sierra)