Revert "KokkosClassic: No longer imported into Kokkos namespace."
This reverts commit 13316062e8959b0429cc51c3b73143ca4e1baf10.

Sierra hasn't had a chance even to get the new KokkosClassic namespace
yet, because of a delay in Trilinos integration.  Thus, it wasn't fair
to get rid of the Kokkos namespace.  Thanks to Brent for pointing this
out, and sorry for any trouble this commit may have caused you.

It looks like we can retain the Kokkos namespace in KokkosClassic for
backwards compatibility.  However, all users of KokkosClassic, Tpetra,
and all downstream packages, please take note: If you currently use
the Kokkos namespace to refer to KokkosClassic things, please switch
to using the KokkosClassic namespace to refer to KokkosClassic things.
The Kokkos namespace will belong exclusively to Kokkos Array in the
future.  Please make this switch as soon as you are able.