Tpetra:  Enable GPU ETI of MultiVector on long & unsigned.
Tpetra::MultiVector was not being explicitly instantiated for scalar
types = long and unsigned on the ThrustGPUNode, which prevents
MatrixMarket::Writer from being instantiated on the GPU, since it
instantiates a MultiVector with scalar=long to save the Map.  This fixes
that, and was done by moving long and unsigned from the general
exclude set to the exclude set including int.  Thus CrsMatrix and
friends are still not instantiated on long and  unsigned on the GPU due
to lack of support by CUSPARSE.  Enabling this required also adding
instantiations for unsigned to the kernel operators for
Kokkos::DefaultArithmetic, and also the Teuchos::ScalarTraits for CUDA.