Allowing float and complex support in SS code (Trilinos #5902)
The commits I just pushed for the Thyra empty process changes (Trilinos #5902)
had some tests that failed to build when complex was enabled.  Here I am
allowing complex and float support and instantiations to be turned back on so
that I can reproduce and fix these failures.  Also, in the future this will
protect complex and float support before I push for the MPI_DEBUG_SS and
SERIAL_RELASE_SS builds.