Ifpack: Fixed Clang compilation failure due to non-POD VLAs
Replaced VLAs with non-POD elements by std::vector.

Variable-length arrays (VLAs) are not part of the C++ standard (yet) but
are supported as an extension by some compilers (in particular GCC and
Clang by mimetism) to improve compatibility with the C99 standard.

Clang does not support VLAs with non-POD (Plain Old Data) types, which
is sensible since these would be meaningless in pure C.

There are other uses of VLAs in the modified file, but these have not
been changed.