Mesquite: Fix usage of CppUnit
Mesquite had a dependence on the tpl CppUnit, however, it wasn't using it
through the tpl system. Likely this was done because Mesquite supports an out of
Trilinos build with cmake. However, when building in Trilinos this caused some
issues. One important issue with this is that the installation testing for
Mesquite wasn't able to get the information necessary to use CppUnit since it
wasn't declared as a dependency. This commit changes the in Trilinos build to
use the CppUnit tpl instead of the previous logic. The out of Trilinos build
should be unaffected.

This change required the nightly builds that previously had been running
Mesquite with CppUnit enabled to be updated to use the correct variables to set
and enable CppUnit.