Tpetra: Added "TieBreak" object and interface in the Directory
This class (which lives in the details namespace) provides an
interface for determing ownership of a global id over multiple
sharing processes. It is used with a directory to make an
appropriate decision.

This commit also incoproates  the TieBreak object
into the directory, in particular the DistributedNoncontinguous one.
A new constructor has been added to the Directory, as well as
the DistributedNoncontiguousDirectory (DNDirectory) that takes a TieBreak. Furthermore
the implementation of the DNDirectory has been moved to an initialize
routine which takes a map and an optional Ptr to a TieBreak object.

Also modified the createOneToOne to take a TieBreak object

Also modified the Map_OneToOne test to excersise the TieBreak capability