Ifpack2: SparseContainer: Simplified & optimized apply & weightedApply.
The optimization isn't fully in place yet.  The idea is that the methods
would detect whether the subset permutation (defined by ID(i)) is equal
to the local indices.  If so, the methods could skip copying and
permuting input and output multivectors.  All of that is in place,
except for detection of the special case.  Commented-out code already
exists in the compute() method to handle this case.

I also took the chance to fix a few different preconditioners' apply()
methods.  They were missing implementations for some cases of nondefault
mode, alpha, and/or beta arguments.  Thus, if called with nondefault
values of these arguments, the methods might throw an exception or
silently return an incorrect answer.  In some cases, I added the missing
implementations, and in others, I added checks for the unimplemented
cases.

Finally, I changed Tpetra::MultiVector::elementWiseMultiply so that it
wouldn't throw an exception in case its arguments have different
communicators or incompatible Maps.  This lets (for example) the scaling
vector D have a globally distributed Map (with MPI_COMM_WORLD, for
example) but the other multivector input to have a local Map (with
MPI_COMM_SELF), as long as they both have the same number of local rows
on each process.  This is useful for SparseContainer's weightedApply
method.

Build/Test Cases Summary
Enabled Packages: Ifpack2
0) MPI_DEBUG => passed: passed=22,notpassed=0 (0.48 min)
1) SERIAL_RELEASE => passed: passed=18,notpassed=0 (5.25 min)