Ifpack2: Container,SparseContainer: Improved docs, & marked apply const.
I improved the public documentation of Container and SparseContainer.
I also marked the apply() and weightedApply() methods const, for future
compatibility with Tpetra::Operator.  Eventually, we want Container to
implement Preconditioner.  Since Preconditioner in turn implements
Tpetra::Operator, and Tpetra::Operator's apply() method is const,
Container's apply() method would also need to be marked const if we
decide to make Container implement Preconditioner.

We want Container to implement Preconditioner in order to make it
possible to nest Ifpack2 preconditioners arbitrarily.  For example,
we could extract a Container and apply it directly as a preconditioner
in a Belos iterative solve.