TrilinosCouplings:  Add GPU clone to IntrepidPoisson example.
This adds the ability to clone the linear system and solve it on the
GPU.  It works in parallel, and there is a new option
--gpu_ranks_per_node to specify the number of GPUs per node.  If there
are more MPI ranks per node than that, the rest run on the CPU.  To
support this, the code queries environment variables to determine the
local MPI rank on each node, which works for OpenMPI and MVAPICH.  For
other MPI libraries the user needs to specify the number of MPI ranks
per node via --ranks_per_node.  Additionally the user can specify an
offset for which CUDA device the rank is associated with via
--device_offset.  If there are n gpu ranks per node, local rank i is
associated with CUDA device i + device_offset for i=0,...,n-1.

With Jonathan's help, an xml file is also supplied suitable for calling
MueLu on the GPU.