Kokkos: Simplify and improve compile-time robustness of View.
A templated function interface can be disable through 'enable_if'.
A non-templated function interface can be disabled by substituting
its input or return type with a unusable error-tag class based upon
a compile time condition.  For example, a View has an device shared
memory constructor if it is unmanaged.  If the View is managed then
the device argument type is substituted for the class
'ViewError::device_shmem_constructor_requires_unmanaged'.