Make tests report not run in mpi when the test exe is not available
The default behavior of ctest is to check the existence of an executable before
running it as a test, however for mpi the executable is mpiexec or some loader
program which is always there. This was causing mpi tests to fail with an error
from mpi instead of the desired not run. This commit makes the executable for
every test a required file to force ctest to check for the real executable
before attempting to run the test. If the executable does not exist a not run
status will be reported instead of a failed.

This change required a small modification to the parameter list of
TRIBITS_ADD_TEST_ADD_TEST, However since this is an internal function in Tribits
the change shouldn't affect backward compatibility.