Panzer: Modifying dirichlet boundary conditions to handle edge basis
The idea here is that the user of panzer inherits from the same
dirichlet bc strategy. However instead of simply setting nodal
values the dirichlet condition is constructed by computing the point
wise residual between the field and the user value and then taking
the tangent on the edge at the point. This gurantees all the scaling
and rotation stuff is correct because everything lives in physical
space. Note that this was rather messy but hopefully it will be easier
for a user to apply. This also requires the implementation of
an extra dirichlet residual evaluator.  Finally the boundary condition
in the curl laplacian example had to be modified so that it excercises
this new capability (though not very well at this point).