MueLu: removed unused coarsestManager in the HierarchyManager
This manager was set through XML file, but was never used. In fact,
there is no way one can use it. The problem is that the number of levels
is not defined a priori, so we cannot make requests in time. The only
scenario where we could use this manager would be if we guarantee that
we achieve numDesiredLevels.