Stokhos:  Experiment with CrsProductTensor and read-only cache.
I tried implementing a version of the CrsProductTensor multiply
algorithm that uses the read-only data cache via the __ldg() intrinsic
for sm_35.  So far I can't get it to be any faster than the pure
shared-memory version.

Also tweaked the shared-memory version to simplify the block-column
loop (and saves a register).