MueLu: disabling easy parameter test when demangling is set incorrectly
The tests for MueLu are failing because of not demangling symbols. The
problem is with the variable HAVE_GCC_ABI_DEMANGLE in Teuchos. This
variable is set as a result of a ABI test in
teuchos/cmake/CheckForAbiDemangle.cmake.

The current trilinos test framework on software results in exit code 127
for this test due to unresolved libraries. Specifically, the test itself
is compiled using mpicxx with comes with extra string of libraries such
as "-L<path_to_mpi>-lmpi_css -lopen-rte ...". The run of the test,
however, is done with cmake, which does not have those libraries in the
library path. So it is not run, produces a return 127 value, and results
in the entry:
  "HAVE_GCC_ABI_DEMUNGLE: Internal="
in the cache. This results in Teuchos not demangling C++ names, and
failing tests comparing two outputs, as one is with demangled, one is
without.

For temporary fix for this is disabling the tests when that variable has
incorrect value.