Update environment on trilinos-test(2) for nightly testing
A configure time test is failing due to not finding a shared lib while running
because the needed library is not in a normal system path and rpath isn't used
while building the test. Adding the library's path to LD_LIBRARY_PATH should fix
this though this is a bit of a hack. The reason why the library is needed at all
is because the configure time test is being built with the mpi compilers even
though it isn't strictly needed.