Ifpack2: Added mix-in for preconditioners that can change their matrix.
This commit is part of a sequence of changes intended to make the type
of inner solver used by AdditiveSchwarz a run-time parameter, instead of
a compile-time template parameter.

Our solution involves letting clients change AdditiveSchwarz's inner
solver at run time.  This is the reason for the new
Details::NestedPreconditioner mix-in interface: classes that have an
inner solver and that let clients change that inner solver implement
NestedPreconditioner.  This lets Factory test at run time whether a
Preconditioner subclass can accept an inner solver.

If one sets a NestedPreconditioner's inner solver, the
NestedPreconditioner in turn needs to be able to pass its "local matrix"
to the inner solver.  (For AdditiveSchwarz, the local matrix is the
result of applying LocalFilter to the overlap matrix.)  This requires
any valid inner solver (that is, any Ifpack2::Preconditioner subclass)
to have a setMatrix() method, that accepts a matrix possibly different
from that provided to the inner solver's constructor.  Ifpack2
preconditioners have never supported this before; they did not give
users a way to change the matrix after construction of the
preconditioner.  This is the reason for the new Details::CanChangeMatrix
mix-in.  Rather than trying to change all preconditioners at once to
support the new feature, we can change the preconditioners one at a
time.  Inheriting from CanChangeMatrix marks a preconditioner as
implementing the new feature.

This commit also includes the (minor) changes to ILUT required to make
it implement CanChangeMatrix.  We began with ILUT because it is the only
inner solver currently supported for explicit template instantiation of
AdditiveSchwarz.