Ifpack2: subdomain solver for testing only
This commit enables a subdomain solver in Factory to testing purposes.

The test solver simply copies the values of the right-hand side to
the solution vector.  It's only intended to be used in unit testing
of the Additive Schwarz method.

It's not guarded by ifdefs because I want it to be tested nightly.