Tpetra: New fn to set CombineMode param (w/ validator) in ParameterList.
This commit adds a function, setCombineModeParameter(), that sets a
CombineMode parameter (with validator) in a Teuchos::ParameterList.  The
validator uses Teuchos::StringToIntegralParameterEntryValidator, which
means that it knows how to convert from a string value to a CombineMode
enum value.  The validator uses the new case insensitivity feature of
that class, so that users don't have to worry about capitalization.
This will make the new function compatible with Ifpack2, whose
AdditiveSchwarz class currently rolls its own validation for that
parameter.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=65,notpassed=0 (1.31 min)
1) SERIAL_RELEASE => passed: passed=50,notpassed=0 (5.59 min)