KokkosClassic: TBBNode: Fixed minor syntax error.
KokkosClassic::TBBNode was calling the parallel_reduce functor's (WDP's)
identity() method as a class method, rather than as an instance method
(as it should).  This commit fixes that.

This is related to previous commits that make KokkosClassic more
consistent in how it uses functors.  Those commits, in turn, are related
to the new Kokkos wrapper Nodes that were recently added to facilitate
porting Tpetra to use new Kokkos.  We hadn't fixed TBBNode to make it
consistent with TPINode, SerialNode, OpenMPNode, and ThrustGPUNode,
because TBBNode is disabled by default (unless one enables the TBB TPL).
This commit remedies that omission.

For some reason, when I enable this Node type, many (but not all)
KokkosClassic and Tpetra tests fail with a malloc() issue.  I'm not sure
why, but I know that the build error needs to be fixed, so I'm forcing
the push.

Build/Test Cases Summary
Enabled Packages: TeuchosNumerics, KokkosClassic, Tpetra
0) MPI_DEBUG => FAILED: passed=91,notpassed=15 => Not ready to push! (1.97 min)
1) SERIAL_RELEASE => FAILED: passed=77,notpassed=12 => Not ready to push! (0.12 min)
Other local commits for this build/test group: 66e477e
WARNING: Forced the push!