Add option to allow tests to be run in serial
An issue came up with cuda tests when running ctest -j n where n > 1 that was
allowing multiple tests to attempt to run on a card that was in exclusive mode
which causes the tests to fail. Cmake supports the ability to flag a test as to
be run in serial, but there wasn't an interface to that through tribits_add_test
or tribits_add_advanced_tests. This provides that interface with the option
"RUN_SERIAL". When RUN_SERIAL is set the test will be the only one allowed to
run while it is executing.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1554,notpassed=0 (37.21 min)
1) SERIAL_RELEASE => passed: passed=1518,notpassed=0 (11.89 min)