Stokhos:  Changed static storage size to match view column size.
Previously the static storage size was "size per thread", not "size
across all threads".  That makes it virtually impossible to use the view
in "MP::Vector mode" with a different number of threads (e.g., between
host and device).  I changed this so they always match, and do a
run-time dispatch in the kernels based on the chosen number of threads.
Thus not all values of "size per thread" are supported, but most of the
ones we should need are.