Tpetra::Vector: Added offsetView(NonConst) shadows in Vector.
offsetView and offsetViewNonConst are MultiVector methods that return a
view of a subset of rows of a MultiVector.  They are useful for
constructing the input or output of block operators.  Recently, a Sierra
developer complained that when calling these methods on a Vector, they
return a MultiVector rather than a Vector.  What's worse is that the
resulting MultiVector can't even be dynamic_cast back to a Vector.  The
result is that if users ever call offsetView or offsetViewNonConst, they
are committed to using MultiVector, even if they only ever want to use
Vector.

I decided to fix this by reimplementing the two methods in Vector, but
making them return Vector instead of MultiVector.  Reimplementation was
necessary so that the return values could have a different type.  This
means that the methods in Vector shadow those in MultiVector, but that's
allowed in C++.  Clang 3.2 doesn't give compiler warnings in that case.
If you get compiler warnings as a result, please let me know.

The result of this change will be that users can choose whether they
want to handle vectors as Vector or as MultiVector instances.

I also took the opportunity to fix description() and describe() so they
generate valid YAML output.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=65,notpassed=0 (1.60 min)
1) SERIAL_RELEASE => passed: passed=50,notpassed=0 (0.09 min)