Stokhos:  Make Epetra(Ext) an optional dependency for Stokhos.
Previously Epetra and EpetraExt were always required to build Stokhos.
Now they are optional since all of the new development is based on
Kokkos and Tpetra.

Also took this opportunity to reorganize the source code and examples,
and removed some examples that aren't relevant.

Buried in this commit is the beginnings of a Tpetra-MP::Vector test that
is currently disabled.