Kokkos: LinAlg: Fixed device tests for ArithTraits.
The device tests now build and pass.  I have not yet enabled the test
for the Kokkos::Cuda device; when I do, I get build errors that the
ArithTraitsTester<ScalarType, Kokkos::Cuda> argument to parallel_reduce
is an incomplete type.  I don't get these errors for the other partial
specializations of ArithTraitsTester for the other Kokkos devices, so
this must be a CUDA-specific issue.