Kokkos WrapperNode: fixed two issues when using the CudaWrapperNode
Added define for KERNEL_PREFIX from classic kokkos, so that legacy functors
can be used with the new CudaNode.

Fixed some issues with creation and destruction of nodes when both
a HostWrapperNode and a CudaWrapperNode are in use. The CudaWrapperNode needs
to initialize/finalize the HostDevice (its mirror_device_type) as well, but needs
to be aware that a HostWrapperNode might have intialized it already or is still in use.
The same is true the other way around.