Teuchos: TpetraRefactor related needed to add some functions to CUDA ScalarTraits
When compiling with the Kokkos CudaWrapperNode everything is compiled with NVCC,
thus all code is also parsed as device code, even if run only on the host. For that
reason a number of functions needed to be added to the cuda specialisation of
ScalarTraits. Note the random and seedrandom functions are dummies only.
For the same reason the Cuda Variant of the ScalarTraits has to be included by the
non-Cuda variant if compiling with NVCC.