Kokkos: Core: Added public documentation to parallel_scan.
Thanks to Carter for explaining to me how to use parallel_scan to
implement either an inclusive scan or an exclusive scan, depending on
the order of operations in the scan functor's operator().  I added both
examples to parallel_scan's public documentation.  I also added Carter's
third example of using an exclusive scan to build up an offsets array
from a counts array.

This commit ONLY affects the public documentation of Kokkos, so it is
not necessary to run the check-in test script.

Build/Test Cases Summary
Enabled Packages: KokkosCore
Enabled all Forward Packages
0) MPI_DEBUG => passed: configure-only passed => Not ready to push! (0.72 min)
1) SERIAL_RELEASE => passed: configure-only passed => Not ready to push! (0.72 min)
WARNING: Forced the push!