Fixing proper override of MPI_EXEC_MAX_NUMPROCS in TribitsCTestDriverCore.cmake (Trilinos #5961)
With this change, MPI_EXEC_MAX_NUMPROCS will only be overriden through the env
or in CMake code in TribitsCTestDriverCore.cmake if it is explicitly
overridden.  Before, this was always set to a default of 4 reguardless what
the internal project CMake files said.  The new behavior is to only override
MPI_EXEC_MAX_NUMPROCS when it is explicitly set to != 0 (which is a flag to
not override it).

I have added a new unit test Override_MPI_EXEC_MAX_NUMPROCSS to ensure that
the override still works (in case anyone is using this already) and other
(modified) unit tests show that MPI_EXEC_MAX_NUMPROCS is not set at all in the
configure if it is not overridden.

Build/Test Cases Summary
Enabled Packages: COBRA_TF, VRIPSSTiamat
0) MPI_DEBUG => passed: passed=134,notpassed=0 (31.58 min)
1) SERIAL_RELEASE => passed: passed=125,notpassed=0 (3.24 min)