Amesos2: Added {Im,Ex}port caching to Tpetra MultiVecAdapter (Bug 6011).
Amesos2's adapter for Tpetra::MultiVector now caches the Import and
Export objects used to redistribute vectors between the users'
distributions, and the solvers' distributions.  Amesos2 used to create
these objects on every solve() call, which was expensive.  The MueLu
developers noticed that this was happening (because they invoke Amesos2
for the coarsest-grid solve) and reported it as Bug 6011.

This fix only addresses Amesos2's Tpetra adapter.  Presumably, its
Epetra adapter has the same problem, and needs the same fix.  (Don't be
fooled by the Epetra_Import and Epetra_Export in the Epetra adapter; the
Tpetra adapter had the analogous objects, but was not using them.)

Note also that this fix could be optimized for the common case of the
users' input and output MultiVectors having the same Map.  In that case,
the adapter would only need to build the Export object, for bringing in
the user's input MultiVector to the solver's distribution.  It could
then use the same Export object in reverse mode to redistribute from the
solver's output MultiVector to the user's output MultiVector.  We have
not yet implemented this fix.