Belos: MultiVecTraits: Cleaned up Tpetra adapter.
I removed the KOKKOS_NODE_TRACE macros.  They were only being used in
a debug build with Node = Kokkos::ThrustGPUNode.  It's possible that
they might still be useful to CASL, but we plan to deprecate
ThrustGPUNode in favor of the new CUDA wrapper Node type.  The macros
just print a label to std::cerr, so it is unlikely that they would be
useful or even desirable at scale.

NOTE: This commit does NOT make the analogous changes to the Tpetra
specialization of Anasazi::MultiVecTraits.