Kokkos/Tpetra: Silencing lots of compiler warnings
This removes unused variables, as well as solving the unknown pragma issue.
The prgama issue is resolved by defining macros which specify whether a
pragma can be used or not. For example one would write

    #pragma unroll
    #pragma ivdep
    #pragma vector always
    #pragma loop count(128)
    #pragma simd
    for(int j=0;j<length;j++)
      a(i,j) += b(i,j);

While ugly it gets rid of the warnings. Its hard to find a more elegant solution,
(at least before C++11) we can't really mix pragmas with macros.

A unit test was added to Kokkos to test this.