Tpetra: Added KokkosRefactor version of Map.
This version of Map implements the following semantics:

  - View semantics (shallow copy, exposed copy ctor and op=)
  - Most methods are thread-safe and thread-scalable on host
  - Internal device object, most of whose methods are thread-safe
    and thread-scalable on device

Once it passes all the tests, I'll add the getDeviceMap() method, which
lets users access the internal device object if they want to use it in
their kernels.

The implementation compiles (for the Kokkos::Threads device) and passes
some, but not all, Tpetra tests.  Thus, it is disabled by default.  It
will only build if the TPETRA_USE_KOKKOS_REFACTOR_MAP macro is defined.
I've tested the build both with and without this macro defined.  I will
work on getting it to pass all the tests; I just want to get this
checked in so I can work on other things.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosContainers, KokkosLinAlg, Tpetra
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_SS => passed: passed=69,notpassed=0 (1.91 min)
3) SERIAL_RELEASE_SS => passed: passed=54,notpassed=0 (0.18 min)
Other local commits for this build/test group: 80c7c31