MueLu: new PerfUtils and new communication pattern routine
Created new PerfUtils file. The goal is to move all statistics related
stuff there, so that we don't put unnecessary stuff in original Utils
and Utils2.

Also added a function to plot communication graph of a matrix. The
pattern is printed as a seris of lines with the pattern:

<rank>: <neigh1>(<weight1>) <neigh2>(<weight2>) ...

where neigh* are neihboring processors, and weight* are the weights. The
weights are currently set to be the number of imported degrees of
freedom.