Tpetra: CrsGraph (Kokkos refactor): Fixed silly build error.
Some Tpetra developer in the past had the habit of giving method
inputs different names in the _decl file than in the _def file.
This caused a build error when I moved the implementation of a
private template method of CrsGraph to the _decl file from the
_def file.

The only way to test this particular build thing was to give it
to perseus' nightly tests.  That's why we didn't see it before
we pushed.