Tpetra: RTI: Fixed build error in test.
<:: is not a valid way to begin a template argument list.  One must
leave a space between the less-than sign and the double colon.  The
specific build error was reported here:

http://testing.sandia.gov/cdash/viewBuildError.php?buildid=1464649

as follows:

error: `<::' cannot begin a template-argument list [-fpermissive]
note: `<:' is an alternate spelling for `['. Insert whitespace between
`<' and `::' (if you use `-fpermissive' G++ will accept your code)

This commit adds a space between the `<' and the `::', as directed
by the compiler.