MueLu: reducing "number of partitions" to being non-generated data
From now on, I will refer to "number of partitions" as NOP.
Originally, we provided a generating factory for NOP which was always
set to be RepartitionFactory. In fact, in RepartitionFactory we both
generated and requested NOP.

This unusual approach resulted in subtle cyclic dependency which has not
been found for years. The dependency goes like this:
 - some partitioner (like ZoltanInterface) requests NOP,
   generated by RepartitionFactory
 - RepartitionFactory requests "Partition",
   generated by ZoltanInterface
When one construct the hierarchy, in most cases we construct fewer
levels than the provided maximum, which requires the release of all
dependencies from the coarsest level. So TopRAPFactory call
RepartitionFactory::DeclarInput() in the RELEASE mode.
RepartitionFactory releases "Partition", but cannot call the full
release on ZoltanInterface as that still has outstanding
dependencies (NOP).

This patch make NOP act as a simple level data. We don't provide any
generating factory for it. Any factory that needs NOP, simply does

  level.Get<GO>("number of partitions")

(NOTE: we cannot use Get<GO>(level, "number of partitions") as that goes
through valid parameter lists, which no longer contain that).

If NOP is not found, we throw a typical exception. With the proper order
of calls, NOP is constructed by RepartitionFactory before anybody else
requires it, and Set on the level with NoFactory. This is similar to the
user data, but as we do not set a keep flag to it, it is removed after
construction.

The only annoying consequence of this patch is the addition of NOP data
on each level after a hierarchy has been constructed.