I fixed a bug that occurred when (1) multiple object weights were used and (2) ParMETIS 4.x was compiled with 64-bit integers and reals. In this case, Zoltan's float part sizes were copied into an array of doubles for ParMETIS.  Zoltan was scaling the part sizes so they would sum to one, but the scaling was incorrect when multiple object weights were used.
I removed the scaling because part sizes are alread scaled in
Zoltan_LB_Get_Part_Sizes.  They should not need to be rescaled when they
are copied.

I left a comment in the code, however, in case the lack of re-scaling
becomes a problem.

Bug reported by Johnny Collins, LANL.