Tpetra:  Fix warning about set but unused variable.
This warning was actually significant in that the way the test was
written, the final pass/fail of the test was only determined by the last
test because the "success" variable wasn't updated as the test went
along.  I reworked the logic to compute the final pass/fail from each of
the sub-tests.