Tpetra:  Fix incessantly failing HybridPlatform tests.
The Tpetra HybridPlatform tests have ALWAYS failed on my machine with
seg faults.  Turns out the reason was a mismatch in how the class was
compiled in the tpetra library versus in the tests.  The HybridPlatform
class has a method for each enabled KokkosClassic node.  The test
requires double, so if cuda was enabled but cuda-double was not enabled,
the test was attempting to disable the cuda node, presumably to
eliminate a compiler error.  However, the macro used to determine if
cuda-double was enabled was incorrect, and therefore always disabling
the cuda node.  That means the class had a different number of methods
when parsed in the test, versus when it was compiled in the library (if
cuda was enabled), causing a mismatch in the layout of the
HybridPlatform object in memory, resulting in seg faults.  Fixing the
macro isn't really the right thing to do because this same behavior
would have happened if you disabled cuda-double.  So instead, I removed
all of this logic and just disable the tests altogether if cuda is
enabled but cuda-double is not.  Now the tests work.