SEACAS: nem_slice - turn on compresssion for nemesis file.
Lowest level of compression seems to reduce the nemesis file size by 97% or more
for most problems with a slight increase in time required to write the file.
As an example, the file size for a 200 Million element model was reduced
from 3,737,425,956 bytes to 164,914,451 bytes a reduction of 96%.

Current setting is compression level = 1 with shuffle.

(From Sierra)