Add a documentation test demonstrating simple usage of a custom connectivity map.
Also fixed a bug where the const getter for an entry in the connectivity-map table
was returning by value instead of by const reference.

Also fixed a bug where get_connectivity (the function that computes non-stored
connectivity if the user tries to access disabled connectivity) was getting stuck
in an infinite recursion if it needed node -> element connectivity but that was
disabled. Added a new unit-test to verify that get_connectivity will now throw in
that case instead of getting into an infinite recursion.

(From Sierra)