MueLu: arcadia can now calculate statistics for multiple runs
When we run a program several times, we produce output files named
"cmd1_run1.epetra", "cmd1_run2.epetra", and so on. Now a user is able to
run arcadia as
  $ arcadia.py -a -o cmd1
which would get the times from all times, and combine them. By default,
the combination is the minimum number. User can change that in the
function stat_time(). In the future, we may want to add a cmd line
switch to compute avg, avg without outliers, max, etc.