Tpetra: Directory: Simplified use of pImpl_ idiom.
It uses a raw pointer now instead of Teuchos::RCP.  Maybe that's a step
backward, but I just felt like it wasn't necessary to invoke heavyweight
reference-counting machinery for a pointer that never escapes its class,
and that doesn't need a custom destructor.  std::unique_ptr (C++11)
would be better for this use case, but we don't want to require use of
C++11 yet.