Fixing TRIBITS_ADD_TEST() for MSVS generator (PHI Kanban #3314)
So MSVS projects have a silly "configuration type" that the CMake generator
has to abide by and it puts binaries under that directory (what a pain).  You
need to know this if you are going to give the directory for the executable
with TRIBITS_ADD[_ADVANCED]_TEST().  What I have done is to override it to be
just one configuration type based on CMAKE_BUILD_TYPE.

TODO: Add some unit tests for this.