Tpetra: Map (classic & Kokkos ref.): Added isOneToOne predicate.
isOneToOne is a collective which tests whether the Map is one to one
(that is, whether every global index is owned by at most one process
in the Map's communicator).

This commit also includes a test for many different kinds of Maps.
This is the first test which assumes view semantics of Map -- in
particular, it assumes that empty construction and operator= work
correctly.  To make this work, I had to modify the classic ("not
Kokkos refactor") version of Map to expose the default constructor.

Build/Test Cases Summary
Enabled Packages: Tpetra, Ifpack2, Belos, Xpetra, Amesos2, Zoltan2
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=603,notpassed=0 (17.63 min)
1) SERIAL_RELEASE => passed: passed=525,notpassed=0 (2.11 min)