TrilinosCouplings: Made MueLu optional in Tpetra scaling example.
It's handy to be able to build this example without building MueLu.
Thus, I made the MueLu package optional.  In order to support this, I
added the HAVE_TRILINOSCOUPLINGS_MUELU macro to
trilinoscouplings/cmake/TrilinosCouplings_config.h.in.  Trilinos'
build system defines this macro automatically, if and only if MueLu
and TrilinosCouplings are enabled, but one must add the macro to the
package's .h.in file in order to use it.

I also took the opportunity to wipe out some old unused files in
TrilinosCouplings, that relate to the old build system.  We should
perhaps readd TrilinosCouplings_ConfigDefs.h(pp) at some point, but
the old one was totally wrong and had a bunch of "using std::*" in the
global namespace that would mess things up if used.  I thought it
better to get rid of a broken file, than to leave it in place and hope
that Future Self will remember what's broken.

Build/Test Cases Summary
Enabled Packages: TrilinosCouplings
Enabled all Forward Packages
0) MPI_DEBUG => passed: configure-only passed => Not ready to push! (1.11 min)
1) SERIAL_RELEASE => Error, The build/test was never completed! (the file 'SERIAL_RELEASE/email.out' does not exist.) => Not ready to push! (-1.00 min)
WARNING: Forced the push!