Ifpack2: AdditiveSchwarz: Got rid of SerialMap_ and SerialImporter_.
AdditiveSchwarz was creating SerialMap_ (a Map) during setup and
SerialImporter_ (an Import) on demand in apply().  It was using them
in apply() to import into a MultiVector called "Serial".  However, it
wasn't actually using that MultiVector!  Furthermore, the so-called
"SerialMap_" was just a recomputed copy of the input Map, recomputed
using the slowest Map constructor (that takes a list of global
indices).  This commit removes these objects.  It should hopefully
make both setup and apply() faster.  I'm looking at AdditiveSchwarz to
help me fix Bug 5992.