Belos: GCRODRSolMgr: Fixed build warning (cast from double to int).
Build warnings in library header files must be treated as build
errors, for the sake of internal customers who build with warnings as
errors.  This particular double->int cast is harmless, because the
input of the cast comes from an LAPACK workspace query (and is
therefore guaranteed to fit in int).  However, it's generally a good
idea to mark such casts using static_cast.  That will forestall any
compiler warnings, and make debugging easier, should the cast actually
be incorrect.