Revert "Kokkos: Make most subpackages secondary stable"
This reverts commit 369ea236d899382641a90c24cbd886688c3c36da.

I am reverting this due to a compile error on the gcc 4.1.2 compilers that
appears to be bogus. This issue has been looked into by Carter Edwards and it
seems the compiler is erroneously complaining about unused template parameters
in a partial specialization. Since the 4.1.2 compilers are on most of our build
machines and the solution is likely going to require moving to new compilers I
am reverting this change temporarily so that we don't constantly break builds
and all down stream packages while we are getting new compilers on these
machines.