Kokkos: OpenMP optimizations for the device interface
When not using hwloc, the OpenMP backend treated threads as belonging
to the an architecture with just one core and one numa region. This
lead to kernels using the device (team) interface to run with single
thread only. This was fixed by changing the settings so that without
hwloc OpenMP sees a hardware with N cores in 1 numa region when N threads
are requested.

Added an early termination of the team barrier when team_size==1 for
the OpenMP backend.