Tpetra: MultiVector (classic only): Simplified scale implementation.
This commit simplifies the implementation of scale(ArrayView<const S>)
by using KokkosClassic::MultiVector::offsetViewNonConst.  It's only
really appropriate for the "classic" version of Tpetra, so I didn't
port this change to the Kokkos Refactor version of Tpetra.  I made
this change to help me diagnose a failing Xpetra test.