Tpetra: MV (classic & Kokkos ref): Added "copyOrView" feature.
This option lets the "classic" version of MultiVector have either copy
or view semantics.  This choice is set in the two-argument copy
constructor, and is transmitted to the results of any subsequent
one-argument copy constructor or operator= call.  That is, for the
classic version of MultiVector, the choice of semantics is
"contagious."

The Kokkos refactor version of MultiVector only ever has view
semantics.  However, if the two-argument copy constructor is given
Teuchos::Copy as its second argument, it does a deep copy rather than
a shallow copy.  The resulting MultiVector does _not_ pass copy
semantics to the results of any other copy constructor or operator=
call, though.

For both the classic and Kokkos refactor versions of MultiVector, you
may query the "copy or view" state.  For the classic version, you may
set this state.  (This is for expert users only; the feature may
change or disappear at any time.)  The Kokkos refactor version has the
interface to set the state, but throws an exception if the user tries
to set any state other than Teuchos::View (view semantics).  "Any
customer can have a car painted any colour that he wants so long as it
is black" (Henry Ford).

This new feature makes it easier to implement new classes that assume
view semantics of MultiVector.  In particular, I was motivated to add
this feature when implementing the new
Tpetra::Experimental::BlockMultiVector class.  This feature is also
helpful for avoiding a potential double copy in the Tpetra
specialization of Belos, when "cloning" a MultiVector.