Tpetra: MV: Deprecated getLocalMVNonConst() in Kokkos refactor version.
In Tpetra::MultiVector, getLocalMV() returns "const
KokkosClassic::MultiVector&", and getLocalMVNonConst() returns
"KokkosClassic::MultiVector&".  Neither return type is a good idea,
and the latter is a particularly bad idea.
(KokkosClassic::MultiVector is a view of the MultiVector's data, so
returning it by nonconst reference is like returning a raw pointer by
nonconst reference.)

Furthermore, the const markings of these two methods don't make sense.
getLocalMV() is marked const and getLocalMVNonConst() is marked
nonconst, even though both methods should really be const.  (They
would only need to be nonconst methods if they reallocated or
otherwise changed the pointer to the data.  Changing the pointer is
different than changing the values to which the pointer points.)

Most importantly, for correctness' sake (so that the methods aren't
returning references to temporaries), both methods require that the
MultiVector keep around a KokkosClassic::MultiVector inside.  This
requirement hinders porting MultiVector to use new Kokkos.

The last point alone motivates us to do the following:

  1. Deprecate getLocalMVNonConst().  This ensures continued correct
     behavior for any code that (foolishly) relies on the
     KokkosClassic::MultiVector being returned by nonconst reference,
     while immediately notifying us of that code (hence the changes to
     CrsMatrix and VbrMatrix).

  2. Change getLocalMV() to return KokkosClassic::MultiVector by
     value, not be reference.  "By value" still means a shallow copy;
     the result is still a view of the MultiVector's data.

We also changed the implementation of getLocalMV() to create and
return a new shallow copy of the KokkosClassic::MultiVector, rather
than relying on lclMV_ (the internal KokkosClassic::MultiVector).  It
still needs to accept lclMV_ for origNumRows and origNumCols, but we
will remove that requirement soon too.

This commit only affects the Kokkos refactor version of Tpetra, except
for a minor change to the implementation of VbrMatrix (only to make
the thing compile and pass tests; there is considerable uncertainty
about the correctness of VbrMatrix).