Tpetra: MV (Kokkos refactor): Partial fix for Bug 6127.
The Gauss-Seidel tests are still failing, but at least they are making
more progress.  It looks like I fixed one of the issues, but one
remains.

See discussion of the origView_ field in MultiVector for an
explanation of Bug 6127 and the partial fix of this commit.

Build/Test Cases Summary
Enabled Packages: Tpetra, KokkosCore, KokkosCompat, KokkosLinAlg, KokkosContainers, KokkosMpiComm
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_DEBUG_KOKKOS_REFACTOR => FAILED: passed=70,notpassed=1 => Not ready to push! (1.96 min)
3) SERIAL_RELEASE_KOKKOS_REFACTOR => passed: passed=52,notpassed=0 (13.79 min)
WARNING: Forced the push!