Update Trilinos default C standard to C99.
Previously the strong warnings in Trilinos were set to -ansi which enforces C89
and C++98. This is an issue for anyone wanting to write C99 compliant code. This
updates the C standard to C99, however it leaves the C++ standard at C++98.
Since the declaration of the expected standards for these two languages are now
separate it will allow us to move to a newer C++ standard later without
affecting the C standard.