Remove the name-mangling in framework register_fields.
Currently have not yet removed the stk-mesh get_field method that takes a name but does
not specify a rank. It is still there and will behave the same as before, meaning
that if fields on different ranks have the same name, it will return the one
with the lowest rank.

(From Sierra)