MueLu: changing calls to TimeMonitor::summarize()
Changing all MueLu tests/examples to set ignoreZeroTimers=true in calls
to Teuchos::TimeMonitor::summarize().  Related to commit e745b05d.

The effect is that processes that either have a hard zero for a
timer (unlikely) or don't have a timer won't be used in calculating the
global stats.  This will make the min, averages, and averages per
iteration more meaningful for us.