Kokkos: removed explicit KOKKOS_ATOMIC_USE_OMP31 from unit test
this define can only be set if OpenMP 3.1 is available. Hence the build
errors with some compilers (i.e. GCC <4.7). This only recently became
a problem when atomics with OpenMP were finally switched to use capture
an OpenMP 3.1 feature. Before they were handled with critical sections.
Using the capture mode was always the intend, hence the name of the define.

Note that if available GCC intrinsics for atomics are always preferred.