Getting C++11 flag determination to work (Trilinos #6136)
At least with GCC 4.6.1 and CMake 2.8.4 what was provided by the orginal patch
was totally broken.  Basically, the first try-compile with /Qstd=c++11 failed
(correctly) but then the try-compile variable CXX11_FLAGS_COMPILE_RESULT was
set to on and the the second try-compile automatically passed :-) This means
that the second set of compiler flags were always accepted :-)

This was fixed by by giving each try-compile its own unique try-compile cache
variable.  Also, to get this to work correctly for GCC 4.6.1, I had to try
-std=c++0x (which passes).  Also, the source was wrong and was missing the
include of <algorithm>.  Also, you have teo pass in the found C++11 compiler
options to the second set of existing try-compiles.

The other changes that I made as part of this commit are as follows.

I changed from try-run to try-compile for the existing C++11 tseting commands.
You can't do try-run when cross compiling.  We will need to reexamine how
TriBITS should deal with C++11 features.

I changed the name of TRIBITS_CXX11_FLAGS to
${PROJECT_NAME}_TRIBITS_CXX11_FLAGS and properly made it a user cache
varaible.

CMAKE_CXX_FLAGS is no longer overridded as a cache variable since that would
result in additional appearances of the C++11 options every time you configure
:-) Now CMAKE_CXX_FLAGS is only appended locally at top level (just like is
done for OpenMP options).

There is more cleanup and testing to be done but at least this technically
works now.