MueLu: adding parameter checking in Easy Interface
This is a partial implementation of parameter validation for the
EasyInterface.  It is off by default.  It's partial b/c MueLu sublists
aren't checked.

Assumptions: any non-list parameters belong to MueLu.
             MueLu sublists  are named "level0", "level1", etc.

It is enabled with the option

   <Parameter name="strict parameter checking" type="bool" value="true"/> ,

in which case the presence of "unused" parameters triggers an exception.

   terminate called after throwing an instance of 'Teuchos::Exceptions::InvalidParameter'
     what():  WARNING: Unused parameters were detected.  Please check spelling and type.
         bogus param = bogus value   [unused]
   Throw number = 1

MueLu sublists ("level0", "level1", etc.) are not checked. The code is there but commented out.
For reasons I don't understand, sublist parameters are always marked unused.

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => passed: passed=143,notpassed=0 (22.55 min)