Tpetra: Fixed makePointMap for constant-block-size data structures.
Tpetra::Experimental::BlockMultiVector::makePointMap makes a point
Map, given a mesh Map and a block size (same for all blocks).  This
commit fixes makePointMap so that it correctly makes an overlapping
Map, in case the input mesh Map is overlapping.  This will let users
use Import and Export correctly on non-block views of the
BlockMultiVector, e.g., to do finite-element assembly for the
right-hand side.

The fix uses a trick analogous to the "Hilbert's Hotel" example that
explains the bijection between the positive integers and the positive
even integers.