Panzer: Fix panzer_stk namespaces that were missed by the conversion script
Panzer had a namespace with stk in it "panzer_stk" that was mostly converted to
the stk_classic convention. However, this was mostly unintentional, though the
prefered solution by the Panzer developers. However, it wasn't complete due to
the conversion script only converting "namespace<whitespace>stk" lines to
namespace stk_classic. To address this a manual run of sed to convert the
namespace panzer_stk lines to namespace panzer_stk_classic.

This was missed previously due to an option needing to be set to enable the
portion of STK necessary for Panzer.