Tpetra: MatrixMarket: Fix for Bug 6139.
writeMap (and writeMapFile) can now handle an overlapping Map.  (This
commit includes a unit test to verify this.)  As a side effect of this
fix, writeMap is now memory scalable.  It used to gather all the GIDs
from all processes.  Now it only stores O(1) process' GIDs at a time.
The implementation pipelines nonblocking messages with file output in
order to overlap communication with local work (in this case, the file
output).

In order to fix this issue, I had to change the MatrixMarket file format
used for Map.  The previous format (which I'll call version 1.0) stored
a Map as a two-column MultiVector, with all the GIDs in the first
column, and their corresponding PIDs in the second columm.  The problem
with this format is that MatrixMarket stores dense matrices in
column-major order.  As a result, readMap had to read _all_ the GIDs
before figuring out which PIDs own them, and writeMap had to know all
the GIDs and write them, before it could write their corresponding PIDs.
The new file format (which I'll call version 2.0) stores GIDs and PIDs
as consecutive pairs, so that the MultiVector has only a single column,
with two times the global number of GIDs.

This commit also fixes readMap accordingly, so that it can accept
_either_ format version.  It distinguishes between the formats by the
number of columns in the MultiVector.  Format 2.0 also has a string in
its header, "Format: Version 2.0", which we could use in the future to
detect the Map format version.

readMap is _not_ currently memory scalable.  I will change this at some
point.

Build/Test Cases Summary
Enabled Packages: Tpetra
0) MPI_DEBUG => passed: passed=73,notpassed=0 (1.99 min)
1) SERIAL_RELEASE => passed: passed=54,notpassed=0 (0.44 min)