Panzer: Enabling "Smart" DOFs that set the derivative value automatically
This is a performance tunning capability (initial at this stage) that
reduces the cost of the DOF evaluator calls by being smart about which
derivative values are passed through. Note that this also is only enabled
for Jacobian AD types.

This commit was a little more complex because the global indexer is needed
in the default equation set. It needed to be passed through a number of
layers to make this work properly.