MueLu: fixing SetParameterList in ParameterListAcceptor
The old behaviour was buggy. If you provided a single parameter, and
call SetParameterList on a factory, it reset all other parameters to
their default values. That meant, for instance, that calls to
SetParameterList and SetFactory were not commutative.

Instead of doing that, we now simply update the parameters with provided
values, not changing others in the list. The easy interpreter tests look
much more sane now. Particularly, calling interpreter with an empty list
now produces all [default] output, which is what we expect.

The issue was in my mind for a few months, since I found out that you
had to call SetFactory after setting all other parameters in the
EasyParameterListInterpreter.