Tpetra::Experimental::BlockCrsMatrix: Added setAllToScalar method.
This is a fix for Bug 6175.

The new setAllToScalar method does the same thing that it does in
CrsMatrix: it sets all entries to a given Scalar value.  I added a
test for it, and also improved the other BlockCrsMatrix tests by
putting an all-reduce at the end of each of them, to ensure
consistency of the success condition across all processes.  Finally, I
fixed the documentation of Tpetra::Experimental::BlockMultiVector.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosLinAlg, KokkosContainers, KokkosMpiComm, Tpetra, Belos, Ifpack2
0) MPI_DEBUG => passed: passed=151,notpassed=0 (4.19 min)
1) SERIAL_RELEASE => passed: passed=126,notpassed=0 (0.41 min)
2) MPI_RELEASE_KOKKOS_REFACTOR => passed: passed=163,notpassed=0 (3.21 min)