Kokkos: fixed possible deadlock of RandomNumber generator on CUDA
This change makes the get_state operation of the random number generators
a thread independent operation on cuda (i.e. there is no block synchronization).
This makes it more costly but eliminates a potential deadlock issue, when
running with non fully populated blocks.

Updated the unit test for the random generator.