Updated  Kokkos Tutorial Examples
The simple_reduce, simple_view, simple_memoryspaces, and data_layouts examples
all required the addition of a typedef <type> value_type where <type> is the
type for the target for the reduction in the functor for parallel reduction.

The Algorithms/01_random_numbers file yields a seg fault if the user
 compiles and runs without passing two integers at the command line. It has
 been changed now to inform the user to submit 2 integers as command line
 arguments and prevents the segfault.

For each of the thread_team examples in /Hierarchical_Parallelism,
the typename needed to be removed in the typedef typename in the line
typedef typename team_policy::member_type team_member;

For each of the Hierarchical_Parallelism files that used shared memory,
dev.team_shmem() replaces dev when defining the shared memory Kokkos::Views.

For 04_team_scan, there was no makefile included in the folder, so
that has been added for user convenience.