Tpetra: MultiVector: Opt'd subView(NonConst) for contiguous index range.
This commit affects both the "classic" and Kokkos refactor versions of
Tpetra.

I optimized subView and subViewNonConst in MultiVector to check for a
contiguous input column index range.  If it is contiguous, the methods
call their contiguous-range counterparts.  This allows simplifications
in the Tpetra specializations of Anasazi::MultiVecTraits and
Belos::MultiVecTraits.