Kokkos: Linalg: updated Matvec kernels to work with HWLOC on MIC
After switching to the team interface the matvec kernel was running
with a single thread per core on MIC, because teams are not sharing
cores right now when using HWLOC and the kernel was using a single
thread per team.

Also updated the perf test to use the non-member intialization function
of Kokkos.