Tpetra:  Make k_rowPtrs_ const.
This fixes an incompatibility between k_rowPtrs_, which was a non-const
view, with the Kokkos::StaticCrsGraph, whose row offsets array is const.
I made k_rowPtrs_ const and added a second typedef, t_RowPtrsNC, for the
non-const version which is used internally in a few CrsGraph and
CrsMatrix methods.

This gets TrilinosCouplings/FENL with MueLu working again.