Kokkos: fix an issue with ExecutionSpace -> MemorySpace check with CUDA
when compiling with NVCC host code can get parsed as if it is executed
on the host (i.e. a KOKKOS_INLINE_FUNCTION operator of a functor
which is instantiated for Threads or Serial). Thus only a runtime but not
a compile time abort is acceptable for the check whether the hostspace
can be accessed.