Tpetra::CrsGraph: Removed Node public typedef.
This commit only affects the Kokkos refactor version of Tpetra.

The Kokkos refactor version of CrsGraph now no longer has the Node
public typedef.  When referring to the Kokkos Node type, use the
standard node_type public typedef instead.  This typedef did not exist
in the "classic" version of CrsGraph, and it was causing shadowing
warnings when building Tpetra.