Tpetra: MultiVector: Fixed normWeighted Heisenbug.
This commit only affects the "classic" version of Tpetra.

The MultiVector unit test fails occasionally ("Heisenbug"), even for
Node=KokkosClassic::SerialNode (so it's not related to threads).  This
commit fixes that, by replacing the KokkosClassic implementation with
a hand-rolled sequential implementation.  normWeighted is not commonly
used and we will deprecate the "classic" version of Tpetra soon
anyway, so it's OK to have a correct but possibly slow implementation.

Build/Test Cases Summary
Enabled Packages: KokkosCore, KokkosCompat, KokkosContainers, KokkosLinAlg, KokkosMpiComm, KokkosAlgorithms, Tpetra
0) MPI_DEBUG => passed: passed=75,notpassed=0 (5.37 min)
1) SERIAL_RELEASE => passed: passed=54,notpassed=0 (4.88 min)
2) MPI_RELEASE_KOKKOS_REFACTOR => passed: passed=77,notpassed=0 (9.44 min)
Other local commits for this build/test group: 071d518