Tpetra: CrsGraph: Update to fix for Bug 6215.
This commit affects both the "classic" and Kokkos refactor versions of
Tpetra.

This relates to Paul Lin's commit on Mon 01 Sep 2014:

876072ef22cf8b50c88c9e1bc5f20a9b82749d2e

which was a fix for Bug 6215.  This commit updates that bug fix by
reintroducing the original intended optimization of only keeping the
original GID ordering if requested by the user.  It also adds the bug
fix to the Kokkos refactor version of Tpetra (Paul's commit only
changed the "classic" version of Tpetra).