Amesos2: Fixed explicit instantiation macro.
In general, there is no need for explicit instantiation macros to
refer explicitly to template parameters that have default values.
Those values should ALWAYS be left implicit, so that if template
parameters' default values change in (say) Tpetra, I don't have to
edit ALL of the downstream packages and hope for the best.