stk::mesh: avoid unnecessary topology walks in markEntitiesForResolvingSharingInfoUsingNodes(..).
If in the current modification cycle (1) no nodes are newly shared and
(2) a given entity was not created or modified, then the function does
not need to check whether that entity is shared.  In the future, we will
consider whether we can reduce the number of buckets being entered in the
outer loop.

(From Sierra)