[NOX] unchecked dynamic casts
There are numerous unchecked dynamic casts to pointer in Trilinos, and quite a
few of them in NOX. Casting to pointers only makes sense if NULL is something
to work with. In Trilinos, this is typically not the case. The recommended
practise is here to cast into a reference instead of a pointer; this makes the
cast properly throw a bad_cast exception if it fails.

Build/Test Cases Summary
Enabled Packages: NOX
Disabled Packages: Pamgen,Intrepid
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=182,notpassed=0 (20.15 min)
1) SERIAL_RELEASE => passed: passed=182,notpassed=0 (2.96 min)