Ifpack2: Fixed assumption of copy semantics in SupportGraph.
SupportGraph's implementation was assuming that Tpetra::MultiVector's
copy constructor and assignment operator (operator=) both do a deep
copy.  This is only true for the "classic" version of Tpetra, not the
Kokkos refactor version.  The latter has "view" semantics, which means
that the copy constructor and assignment operator only do a shallow
copy.  Furthermore, in the classic version of Tpetra, it is possible
to set Vector to have view semantics.  This commit fixes SupportGraph.