MueLu: mark factory parameters as default when GetParameterList() is first called
Previosly, the Factory::paramList_ was simply assigned to valid
parameter list. The problem (or a nitpick, if you like) is that the
parameters were not marked as default. Now, instead of copying a list,
we create a new empty one, and call "validate and set defaults".

This also helps the interpreter output test, as in some scenarios easy
interpreter called SetFactory which set the defaults, and factory
interpreter did not which resulted in different output.