MueLu: minor change to the easy interpreter
We have a slight ambiguity in MueLu: pre- and post- smoother can be
set up in two different ways:

1. Two SmootherFactory classes:
   preSmoother  = SmootherFactory(proto1);
   postSmoother = SmootherFactory(proto2);
   manager.SetFactory("PreSmoother",  preSmoother);
   manager.SetFactory("PostSmoohter", postSmoother);

2. One SmootherFactory class:
   smoother = SmootherFactory(proto1, proto2);
   manager.SetFactory("Smoother", smoother).

The factory interpreter can only use the 1. The easy interpreter can use
both, but prior to this patch implemented 2. The issue is that they
produce slightly different output. This patch converts easy interpreter
to 1 to produce output consistent with the factory interpreter.