Kokkos: Fixed LayoutStride constructor was not working as intended.
LayoutStride construction with an order and dim array was not assigning
dimension correctly.

Here is an example of how it works and is intended too:
    const int rank = 4;
    const int order[] = { 1 , 3 , 2 , 0 };
    const unsigned dim[] = { dim0 , dim1 , dim2 , 3 };
    Kokkos::LayoutStride layout_stride = Kokkos::LayoutStride::order_dimensions( rank , order , dim );
    Kokkos::View<int****, Kokkos::LayoutStride> A("A",layout_stride);

"order" gives the order of dimensions from smallest stride to largest stride.
"dim" gives the logical dimensions of the view.

The following statements for example are all true:

 A.dimension_1()==dim1
 A.dimension_3()==3
 &A(k,l+1,m,n)-&A(k,l,m,n)==1
 &A(k,l+1,m+1,n)-&A(k,l,m,n)==1 + dim1*3
 &A(k+1,l+1,m+1,n)-&A(k,l,m,n)==1 + dim1*3 + 3*dim1*dim2
 &A(k,l,m,n+1)-&A(k,l,m,n)==dim1