KokkosClassic: Use of SerialNode now protected with macro.
We would like it to be possible to disable SerialNode completely (for
example, in favor of the new Kokkos::Serial wrapper Node,
Kokkos::Compat::KokkosSerialWrapperNode).  This commit changes
Kokkos_DefaultNode.hpp to protect includes of Kokkos_SerialNode.hpp,
and references to SerialNode, with the HAVE_KOKKOSCLASSIC_SERIAL
macro.

This changes selection of the default Node type.  Before, SerialNode
was always the default if no other Node type was selected as the
default.  Now, if no default Node type is specified, the header file
emits an #error with an informative message telling users to set
KokkosClassic_DefaultNode.  This is OK because KokkosClassic's CMake
defines a default Node type on its own, if users don't specify one.  I
had to make this change so that it would be possible to build with
SerialNode completely disabled.  This is necessary for the eventual
deprecation and removal of the KokkosClassic subpackage.