Ifpack2: Now does ETI for Kokkos::Serial wrapper Node.
This commit adds ETI (explicit template instantiation) for the
Kokkos::Serial wrapper Node (Kokkos::Compat::KokkosSerialWrapperNode).

Ifpack2 uses CMake logic to generate the .cpp files for ETI
automatically.  It does so based on .tmpl files in ifpack2/src.  There
is one set of .tmpl files for Ifpack2 classes templated on <LO, GO,
Node> (that use / are graphs only) and another set for Ifpack2 classes
templated on <Scalar, LO, GO, Node> (that use / are matrices).  There
is one .tmpl file for each new Kokkos Device type, for which we want
to do ETI.  (Eric Phipps, who wrote this system, did it that way
because the mapping from Kokkos Device to its corresponding existence
test macro isn't an obvious string substitution.  At some point, we
might want to change that.)  Furthermore, there is a chunk of CMake
code in ifpack2/src/CMakeLists.txt for each Kokkos Device type.

FIXME: It would make more sense to automate the mapping over Kokkos
Device types, so that we don't need explicit dependence on the whole
list of Device types in every downstream package.  What if we want to
add or remove a Device?  Then we have to change a lot of downstream
packages.  Kokkos' CMake should export a variable with the list of
enabled Devices.  Does it do that already?