Kokkos,Tpetra: Moved TSQR implementation into TSQR subpackage.
The implementation of TSQR (Tall Skinny QR factorization) used to live
in the KokkosClassic subpackage.  Since we plan on deprecating
KokkosClassic but want to keep TSQR, this commit moves the TSQR
implementation into the new TSQR subpackage of Kokkos.  TSQR builds
and passes tests successfully.

NOTE: This currently only works if KokkosTSQR and Tpetra are enabled
and forward package dependencies are DISABLED.  I have not yet fixed
use of TSQR in packages downstream of Tpetra, like Anasazi, Belos, and
Sacado.

In the previous version of TSQR in KokkosClassic, TSQR was split into
two libraries, kokkosnodetsqr and kokkosdisttsqr.  This complicated
the implementation unnecessarily, especially since DistTSQR (the
distributed-memory part of TSQR) only depends on MPI abstractly,
through TeuchosComm.  In the new TSQR subpackage, both NodeTSQR and
DistTSQR live in the src/ subdirectory and are built in a single
library.

This commit also changes Tpetra to reflect the different CMake
configuration options.  KokkosTSQR is an optional dependency of
Tpetra.  This is a bit different than before (where TSQR was a CMake
option of KokkosClassic, a required dependency of Tpetra), but it
gives us more flexibility in deploying the Kokkos refactor version of
Tpetra.  The current implementation of TSQR depends on KokkosClassic;
our plan is to rewrite it to use new Kokkos instead.