PyTrilinos: Got wrappers for Tpetra MultiVector and Vector to work
SWIG was getting confused and thought that MultiVector and Vector were
abstract classes, and so it was not wrapping their constructors.  This
was fixed by applying the SWIG directive %feature('notabstract').
Once this was working, I added a set of SWIG macros to make it easier
to instantiate multiple Tpetra classes for multiple Scalar types.
Finally, I added Python factories 'MultiVector' and 'Vector' for
generating typed MultiVector and Vector objects, respectively.

Added a typemap for Teuchos::ArrayView when it is used for seed values
for a Tpetra MultiVector constructor.  The Python version takes a
NumPy array instead.

Added a typedef for KokkosDefaultNode, and use it throughout the
interface file.  Unfortunately, SWIG can get confused if you use a
default template parameter, so we have to specify it explicitly.  This
typedef is a little shorter than the actual default parameter, and is
thus a little more readable.