Added state-checking fixture to aid unit testing of change_entity_owner()
This fixture is only partially used in one test.  The set of tests for
change_entity_owner() will be modified and pared down to a set that
covers all possible entity state transitions.

Also cleaned up naming of existing tests in UnitTestBulkData.cpp so that
the group name is the class ("BulkData") and moved some BulkData tests
from UnitTestChangeOwner.cpp.

(From Sierra)