Adding test for AztecOO::recursiveIterate() that should show memory leak (Trilinos #6252)
This is a very slightly modified version of the file aztec_leak.cc that Steven
Hamilton submitted back in October 2013 reporting a memory leak in AztecOO.

I have not confirmed that this produces a memory leak yet.  This test is too
expensive to add to the test suite yet but I don't want to change it until we
can reproduce the memory leak.