Sacado:  Fix Fad view specializations to hide derivative dimension.
The Kokkos view specializations for Fad were making the derivative array
dimension through methods such as dimension() and size().  This fixes
that so they act like view's of the corresponding rank of Fad types.  I
fixed this by borrowing some technology from the Stokhos UQ
specializations.  I also added unit tests to Sacado and Stokhos to check
that the size of the view is what is expected.