[teuchos, ifpack2] Change TEUCHOS_HAVE_* to HAVE_TEUCHOS_*
Previously, there were several spots where TEUCHOS_HAVE_X was
used instead of HAVE_TEUCHOS_X, where X is LONG_LONG_INT,
COMPLEX, or QT. The result was that sections of code were not
compiling when intended. This commit fixes these. The motivation
for this fix is needing to use
    Teuchos::reduceAll<int, long long int>.

The problem is that the following tests now fail:

  99% tests passed, 3 tests failed out of 1440

  786 - Komplex_simple_MPI_4 (Failed)
  1242 - NOX_LOCA_HouseholderBorderedSolve_MPI_2 (Failed)
  1243 - NOX_LOCA_HouseholderTransposeBorderedSolve_MPI_2 (Failed)

I ran checkin-test.py in order to test the commit and thereby
found the failing tests reported above. Given the nature of this
commit, those failures probably need to be debugged as a separate
issue. I'm forcing this commit after consulting with Mark H.