Sacado & Stokhos:  Add specializations of Promote for expressions.
In the new way of doing Promote<> specializations, I have to manually
add specializations for expression templates.  I added these for all of
the AD/UQ types.  I also added tests for these specializations.  Note
that I had to disable testing of the new Kokkos UQ types as they do
expression templates in a way that make doing a general partial
specialization of Promote hard.  I'll have to think about how to do
that.  This should fix build errors in Albany/LCM for the AD/UQ types it
uses.