Xpetra: MMM fix for Epetra when reusing the graph
Due to Epetra wrapper intricacies, we need to explicitly call
fillComplete on Xpetra matrix after Epetra multiply call. Specifically,
EpetraCrsMatrix only keeps an internal variable to check whether we are
in resumed state or not, but never touches the underlying Epetra object.
As such, we need to explicitly update the state of Xpetra matrix to that
of Epetra one afterwords.