Revert "STK_UTIL: Support for building on 32-bit systems (size_t is 32 bit, not 64-bit) [Nico Schlomer]"
This causes issues on a redhat-5 system using the intel-12 compiler.
Looking at the compiled source, it looks to be giving incorrect
results for the __SIZEOF_??__ symbols.  Reverting until
we have more time to figure out what is happening.

After further examination, on the intel-12 compiler on sierra002, the __SIZEOF_??__
symbols are not being defined, so __SIZEOF_INT__ == __SIZEOF_SIZE_T__ is
true (0 == 0) which is incorrect.  Will need to determine where those symbols
are defined.

Note that with this revert, building on 32-bit systems is again broken
(but hopefully running on rhel5 is unbroken)

This reverts commit 9f614ba63dd19e22f0824a823f478eb5c8c0b59c.

(From Sierra)