Sacado:  Fix scalar assignment for ViewFad.
View<...>::operator(...) = ... didn't work correctly when the
right-hand-side was a constant.  Since the derivative array in the view
cannot be resized to zero, we need to zero out the derivative components
in this case.  I added a unit test for this and then fixed it.  Thanks
to Eric Cyr for discovering this.