Add read and reset to the counter API.
The counter API did not provide a way to read or set the value of a
counter .  The only way to get the value was to increment or
decrement.  The only way to set a value was to increment or decrement
in a loop until the desired value was reached.  This commit adds
nthread_counter_read() which returns the current value and
nthread_counter_reset() which sets the counter value.