Disable always failing tests STK_stk_mesh_unit_tests_MPI_4 and STK_util_parallel_UnitTest_MPI_4 (Trilinos #6266)
It is not important that these tests pass in CI testing because they have been
failing for months and no one cares.  At least this way we mgiht be able to
get the people on the CDash regression mail list to remove their email filters
and respond to new failures.