Tpetra::MultiVector: Fixed View init to use intended execution space.
This commit only affects the Kokkos refactor version of Tpetra.

In Tpetra, it's important to template Views (and DualViews) on the
intended execution space, not on a memory space.  This ensures that
Tpetra will use exactly the specified execution space(s) and no
others.

This matters because View (and DualView) initialization is a parallel
Kokkos kernel.  If the View is templated on an execution space, Kokkos
uses that execution space (and only that execution space) to
initialize the View.  This is what we want.  If the View is templated
on a _memory_ space, Kokkos will use the memory space's default
_execution_ space to initialize.  This is not necessarily what we
want.  For example, if building Trilinos with OpenMP enabled, the
default execution space for host memory is Kokkos::OpenMP, even if the
user-specified DeviceType in Tpetra is Kokkos::Serial.  (This shows up
as a thrown exception, complaining that OpenMP has not yet been
initialized.)