Pamgen: Fixed assert with std::istream.
Pamgen was using code like the following to assert validity of an
std::istream:

std::istream in = ...;
assert (in != 0);

std::istream has a conversion to bool, but doesn't have a conversion to
int.  With Clang 3.5, the above code fails to compile.  I replaced the
"assert" line with "assert (in);" and it compiled.