TpetraKernels: Fixed build error with Kokkos::complex<T>.
This may result in build warnings, because I replaced the following test
for "equals -1" with an actual "s_a == -1.0":

s_a < 1 && s_a != 0

I had to replace it because s_a could be a complex number, and therefore
is not less-than comparable.  There could be other ways to do this, but
this should work for now.