Intrepid: Fixed minor build warning in test.
The "register" keyword is deprecated in C++11, says Clang 3.5 on Mac.
Compilers are pretty good these days at figuring out what belongs in
registers and what doesn't.  In this case, the "register" variable's
type is a template parameter, Scalar.  Scalar may be a type that doesn't
fit in registers, so it doesn't make sense to try to convince the
compiler to put it there.