Tpetra/Ifpack2:  Make return type of RowMatrix::getFrobeniusNorm() consistent.
The return type of childen of Tpetra::RowMatrix was not consistent
between the classic Tpetra and the Kokkos-Refactor
(ScalarTraits::magnitudeType for the former and ArithTraits::mag_type
for the latter).  For technical reasons, those two types are not the
same for the Stokhos UQ scalar types, which causes compiler errors.  I
made them consistent by adding a mag_type typedef to RowMatrix and used
that in the declaration/definition of getFrobeniousNorm() in all
RowMatrix derived classes.  This is one of many solutions to this
problem, and so it can be easily undone if we want to do something else.

I also fixed the implementation of getFrobeniousNorm() in
Tpetra::CrsMatrix and Ifpack2::LocalFilter to be correct for UQ scalar
tyeps.