Stokhos:  Fix weird build error when compiling with NVCC.
For some reason, compiling Stokhos with NVCC generates errors in
compiling Stokhos::CrsProductTensor relating to the internal sorting of
stochastic DOFs.  It is confusing the < in the compare functor with a
template argument list.  I have no idea why this started happening now,
but is clearly a bug in the compiler.  We don't use the sorting anymore,
so I commented it out with a note as to why.