Kokkos: Fixed some issues with 128bit atomics when CUDA is enabled.
While 128bit atomics are not supported on CUDA the signature (i.e. the
device variant) of the function needs to exist during parsing. I added
them and call a cuda_abort inside with a message that it was attempted
to call a 128bit atomic on CUDA. They are only supported on X86 right
now. Mayeb someone could test what happens if you try to compile it on
BGQ??