Update comment about check for sizes of dim() and NumGlobalElements64()
Now that NumGlobalElements64() is used instead of NumGlobalElements() (see
Tom's last commit), this will not overflow an int.  I updated the comment for
this change but argue the check is still not a bad idea and why.

Tom Evans reported that with just this one change, Exnihilo can use 64 bit
Epetra and run larger problems than an be run with just int.

NOTE: There are no tests in Thyra that test the 64 bit need.  It seems that it
is imposible to create a reasonble test because such as test would have to
create an Epetra_Map that NumGlobalElements64() return a number larger than
2^32/2.  You can't store even a single int array that big on one machine (it
would use up more memory than most single systems have).  I looked at the
implementatin of Epetra_BlockMap and it always stories an array of at least
the number of local elements.  Therefore, without mocking out Epetra_Map,
there is no way to test this without having a big machine with a lot of
processes and a lot of memory per process.

Build/Test Cases Summary
Enabled Packages:
Enabled all Packages
0) MPI_DEBUG => passed: passed=688,notpassed=0 (39.38 min)
1) SERIAL_RELEASE => passed: passed=605,notpassed=0 (12.56 min)
Other local commits for this build/test group: f1d85a9