Tpetra/Kokkos: Fixing an issue with isinf/isnan on some platforms.
Depending on the compiler/platform isinf/isnan can be found in the std::
namespace or in the global namespace. This change makes usage of both
functions falling the pattern:

using std::isinf;
return isinf(x);

This makes it unnecessary to check for compilers to figure out the correct
way to do it, and reduces the issues of compiling with one we didn't use yet.