MueLu: added reuse capability to Stratimikos adapters
When MueLu is being used through Stratimikos, we now check whether the
MueLu's parameter list contains "reuse: type" parameter. If it does,
then we are able to reuse existing preconditioner if it was created
prior to the call).

Note, that the whole thing is quite fragile at the moment, and we
heavily rely on a user to not do stupid things, like trying to reuse
with a different matrix graph, or switching parameter lists in time
after the first setup and before the second one.

So far, the code has only been tested through Drekar. That was quite
easy, as by default, current NOX' Thyra adapters automatically reuse the
existing preconditioner.

Build/Test Cases Summary
Enabled Packages: MueLu
Enabled all Forward Packages
0) MPI_DEBUG => Test case MPI_DEBUG was not run! => Does not affect push readiness! (-1.00 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)
2) MPI_SS_DEBUG => FAILED: passed=166,notpassed=5 => Not ready to push! (31.60 min)
WARNING: Forced the push!