Sacado:  Disable definition of acosh() etc. when C++11 is enabled.
Sacado provides a definition of acosh(), asinh(), and atanh() for when
these functions are not provided by the standard library (a legacy from
the original Fad/Rad).  There are configure tests to determine whether
the standard library defines them and disables these definitions if they
are defined.  However the tests don't always seem to work, so I am
adding an additional check if C++11 is enabled, in which case they
aren't defined.