Tpetra::MultiVector: Allow allocation without initialization.
MultiVector's most commonly used constructor takes a 'zeroOut' bool
argument, which governs whether or not to initialize the entries of the
MultiVector (i.e., fill all the entries with zeros) on allocation.  The
argument is true by default, which ensures first-touch allocation.

The new (Kokkos refactor) version of Tpetra used to ignore zeroOut.  It
would _always_ initialize the MultiVector's entries.  This commit fixes
that.  Thanks to Carter for explaining how to control initialization.

To make this easier, I moved all DualView allocation in MultiVector (all
instances of allocation, as far as I could tell) into a single private
function, allocDualView().  That helped, because I had to work around
the fact that Kokkos::DualView's constructor doesn't currently take the
same arguments as Kokkos::View's constructor.  I tried fixing that but
didn't get very far, so I decided just to work around it for now.

Thanks to Eric Phipps for reminding me of this issue.