Sacado:  Rename Sacado_Kokkos.hpp to Sacado.hpp.
I expect the vast majority of Sacado uses only use the Fad types, which
are the only types in Sacado that have been ported to Kokkos.  To
facilitate merging the phalanx-kokkos branch back into the master
branch, I renamed Sacado_Kokkos.hpp to Sacado.hpp and Sacado.hpp to
Sacado_No_Kokkos.hpp.  This should only affect users of Sacado types
that haven't been ported to Kokkos (Rad, DMFad, DVFad, Taylor,
ScalarFlopCounter).  Users of those classes should use the
Sacado_No_Kokkos.hpp include.  The reason for all of this is that those
classes cannot be compiled by nvcc when Cuda is enabled.  Note that both
headers cannot be used in the same compilation unit, as the latter
disables Cuda.