Tpetra:  Replace execution_space with memory_space in get1dCopy.
This fixes build errors on Cuda.  I think the proper way to do this
would be to use device_type to get the execution_space,memory_space
pair, but I'm not sure if that is the name that will be used going
forward (it is still marked deprecated in Kokkos::ViewTraits).