TpetraKernels: Fixed MultiVector::scale kernels for zero-scale case.
Tpetra::MultiVector has several variants of scale().  For all these
variants, if scaling (multiplying) by zero, the methods _should_
follow the BLAS convention of not actually multiplying.  This prevents
multiplication of zero by NaN or Inf, in case the values to be scaled
by zero haven't yet been initialized.  The kernels in question weren't
actually following this convention; this commit changes that.

At some point, we need to optimize the kernels for the case of scaling
by zero, so that the inner loops don't branch.