MueLu: fixed operator complexity output [Bug #6301]
The problem was the following line:

  double operatorComplexity =
  as<double>(std::accumulate(nnzPerLevel.begin(), nnzPerLevel.end(), 0)) / nnzPerLevel[0];

It was a double whammy:

  1. nnzPerLevel is a global_size_t (which defaults to size_t) array.
     std::accumulate accumulates the sum into a variable with a type of
     last argument, though, which is 0, which is set to int. So even if
     size_t is 64-bit, the sum is accumulated into int.

  2. Even if the last argument is of size_t, it can still overflow
     size_t, as it accumulates first, and transforms to double later.

The fix is trivial: use Hierarchy::GetOperatorComplexity(). The original
desire behind the accumulate code was to slightly speed things up: we
already construct nnzPerLevel() array, so lets skip the
GetOperatorComplexity() as it does few dynamic_cast transformations.

Live and learn.