Trilinos: Added 'quadmath' TPL.
GCC provides libquadmath, which implements GCC's "built-in" __float128
(128-bit floating-point) type.  I write "built-in" in quotes, because
a lot of pieces are missing.  For example, operator<<(std::ostream&,
__float128) doesn't exist.  On the other hand, libquadmath provides
many floating-point constants, transcendental functions, and other
useful things.

This commit adds 'quadmath' as a TPL.  The TPL requires that both the
"quadmath.h" header file and the "quadmath" library exist.  Both of
these are part of a GCC installation.

I added the TPL as EX (experimental), because (a) I don't have a
nightly test set up for it yet, and (b) I haven't committed any code
that depends on it yet.  The plan is for __float128 to replace
functionality previously provided by the QD TPL.  There's nothing
wrong with the QD TPL, but it's easier to let GCC install quadmath for
you, than it is to find, download, build, and install QD yourself.