Move Apub::Part_Change_Helper and execute_part_change_operations to Fmwk.
Paired with Manoj, Flint, Johnathan.

Moved to fmwk so that it can be called from within fmwk...

Added unit-testing for it, implemented the stk-mesh-parallel version underneath
so that calling-code doesn't need to change.
Needed to add special handling for recv-ghosts and shared-but-not-owned, since
stk by default doesn't allow changing parts on non-owned entities.

(From Sierra)