Tpetra,PyTrilinos: Removed deprecated function createVectorFromView.
Tpetra::createVectorFromView took a Teuchos::RCP<const Tpetra::Map>
and a Teuchos::ArrayRCP<Scalar>, and returned a
Teuchos::RCP<Tpetra::Vector> which views (not deep-copied) the data to
which the Teuchos::ArrayRCP<Scalar> points.

This function has been deprecated for quite a while now.  It cannot
work with the new (Kokkos refactor) version of Tpetra, because Tpetra
now uses Kokkos::View and Kokkos::DualView for its data structures.
This means that Tpetra can't safely persistently view data to which an
ArrayRCP points.  The input ArrayRCP also constrains data layout (why
should a Vector's local data always have constant stride?).  Finally,
the function is confusing and redundant, because Vector already has
constructors that view input data (as Kokkos::DualView).  Thus, I have
removed the deprecated function.

NOTE: I also updated one line in one of PyTrilinos' SWIG interface
files.  That line turns off some kind of warning for
"Tpetra::createVectorFromView".  Since that function no longer exists
as a result of this commit, I figured that it would be a bad idea to
refer to it in a SWIG interface file.