Removed unused tracing functionality from STK
This is a somewhat complex set of tracing and diagnostic output
functions that have fallen into disrepair in the past couple years since
nobody currently on the team can figure out how to use it.  So, the
decision has been made to rip it all out to streamline the code.

(From Sierra)