Removed tracking allocator from STK
This is potentially interesting functionality if used properly, but
nobody on the team has been properly using the tracking allocator for
the last couple years so that any reported data would be incorrect.
This was removed as a code clean-up and perhaps a teeny-tiny performance
improvement (untested) for cases of frequent memory allocation and
deallocation.

This functionality can always be resurrected later if desired, of course.

(From Sierra)