TpetraKernels: Added 5-arg dot() variant that selects columns of X and Y
Tpetra::MultiVector::dot would like to be able to select columns of
the input multivectors on which to compute a dot product.  This makes
dot() work correctly with so-called "noncontiguous views" (i.e.,
Tpetra views of multivector columns that are not contiguous).

This commit adds a 5-argument overload of KokkosBlas::dot() that lets
callers select the column of X (X_col) and the column of Y (Y_col) to
use.  It still accepts Kokkos Views of the same ranks as the
3-argument version of dot() -- rank 1 for the output array, and rank 2
for the input multivectors X and Y.  This makes the implementation of
Tpetra::MultiVector::dot shorter, since it doesn't have to create
subviews.