STK_IO: Fix problem with trying to overwrite restart file
If using hdf5 as the underlying file type for exodus/netcdf, it is
more picky about overwriting an existing file -- if the file is open,
then it will abort; it will only overwrite an existing file if it is
not open.  Since overwriting restart files (input/output) is a common
usecase, we need to check at this point whether there are any existing
input files with the same name as the file we are attempting to create
here. However, due to symbolic links and other junk, it is often
difficult to determine that the files are the same, so..., If
m_input_region refers to a file, just close it since we should be done
with it at this time...

(From Sierra)