Ifpack2: Added debug code to AdditiveSchwarz, LocalFilter, & RILUK.
When the HAVE_IFPACK2_DEBUG macro is defined, these three classes'
apply() methods check the 1-norm of both input and output, and throw
if it is Inf or NaN.  This is meant to track down the reason for the
failing AdditiveSchwarzRILUK test, that fails only when
Tpetra::MultiVector's constructor actually respects the zeroOut=false
case.  None of the exceptions trigger, yet Belos still reports a NaN.
This suggests that some preconditioner might be returning zero when it
isn't supposed to be.

NOTE: For some reason, HAVE_IFPACK2_DEBUG is NOT defined in a debug
build.  This is broken, but I don't have time to fix it today.