Tpetra: Fixed recent BlockCrsMatrix I/O commit.
I honestly have no idea how it compiled before.  The code was
assigning freely between Scalar and GO.  It looked like the code
wanted to use GO (the variables had names like "allMeshGids", despite
claiming to contain Scalar data), so I change Scalar to GO and hoped
for the best.

btw the best way to manifest compiler errors in this case would be to
ensure that Scalar = std::complex<double> is enabled.