Fix defect in C++11 disable logic (TriBITS #64)
I had accidentally used:

  ELSE (${PROJECT_NAME}_ENABLE_${CXX11_PACKAGE_NAME_IN})
    MESSAGE( FATAL_ERROR ... )

instead of:

  ELSEIF (${PROJECT_NAME}_ENABLE_${CXX11_PACKAGE_NAME_IN})
    MESSAGE( FATAL_ERROR ... )

This is the first example I have seen of this CMake feature where you can list
an unused clause in an ELSE(<clause>) statement that lead to a bug.  This is a
stupid language feature in my opinion.

The result is that if you explicitly turned off Kokkos or Tpetra with C++11
disabled, then you got a configure error.  This broke all of the Trilinos
nightly builds that had C++11 disabled and where Kokkos was explicitly
disabled.

All of this tested was manual testing and this was a test case that I forgot
to test.

Build/Test Cases Summary
Enabled Packages: Teuchos
Disabled Packages: Kokkos
0) MPI_DEBUG => passed: passed=110,notpassed=0 (5.60 min)
1) SERIAL_RELEASE => Test case SERIAL_RELEASE was not run! => Does not affect push readiness! (-1.00 min)