Panzer: Making STKMEFactory::buildLOWSFactory stateless
Previously there was only a non-const version of this method. This
was done to support usage in Drekar, however this is now outdated
and needs to be simplified and made more robust. This change supports
that.

Note that there is a new method that explicitly takes the stratimikos
parameter list and the teko request handler.