Remove failed experimental STK performance unit tests
This was an experiment to use valgrind (callgrind, specifically) to
gather CPU cycles and detailed memory stats, as an attempt to
reduce day-to-day timing noise in performance runs.  It turns out that
this replaces day-to-day noise with machine-to-machine noise.  Many
folks' blades give different results from each other, and certainly the
different HPC platforms are all different from each other despite
nominally the same architecture.  We ended up having to use such sloppy
tolerances for these tests that they became worthless.  So, this is
declared a failed experiment despite being quite interesting.

(From Sierra)