Trilinos: Add updateTrilinosDB.py script
This script generates a CSV file of software package database entries
that can be imported into the Drupal Trilinos web site.  This script
should be run for every new release of Trilinos, as part of the
workflow to update the Center 1400 software package database.  It can
also be run on older releases of Trilinos, to populate the database
with prior versions of Trilinos packages.