Tpetra::Details::FixedHashTable: Fix build error (assignment).
It's not legal to assign a FixedHashTable with one DeviceType, to a
FixedHashTable with a different DeviceType.  This test case only ever
gets exercised when multiple Node types are enabled in Tpetra, which
does not happen by default.

I fixed this by adding a templated copy constructor to FixedHashTable,
that does a deep copy if the Kokkos device types are different.
Tpetra::Map::clone now uses this.

I also added a unit test for the new copy constructor.  It only
actually tests anything if Kokkos enables more than one execution
space.  It can't _fail_ if Kokkos only enables one execution space,
but it does print a warning in that case.