Epetra: Fixed some build warnings.
GCC 4.8.2 has a hard time figuring out that EPETRA_CHK_ERR(-10) should
always return.  As a result, putting this at the end of a function
that returns int, makes the compiler think that the function might
possibly return void.  The easy fix is to insert an extra return after
use of that macro.