Trilinos: Extend updateTrilinosDB.py script to handle maturity
The script extracts the maturity level from the
Trilinos/PackagesList.cmake file and converts it to a longer,
human-readable string.  This string is now included in the output CSV
file.  Trilinos and TPLs are given a blank maturity string, as are
packages that are listed in Trilinos/PackagesList.cmake but not in
TrilinosPackageDependencies.xml