Fix SuperLUDist TPL configure checks
The SuperLUDist TPL file was attempting to do some configure time checks to see
what api was available. The check was not specifying the library for SuperLUDist
so the check was failing. This is now causing an issue on a version of
SuperLUDist that only has the one api that is being checked for. I suspect this
has always been silently failing, but the versions of SuperLUDist that we've
linked against have all had the fallback api so it was never noticed.