Tpetra: FixedHashTable now uses Kokkos to count buckets.
Tpetra::Details::FixedHashTable implements the global index - to - local
index hash table lookup in Tpetra::Map.  I'm working on a gradual
Kokkos-ization of Tpetra::Map.  This commit Kokkos-izes the initial pass
over the global indices ("keys" of the hash table) to count the number
of "buckets" (entries in each "row" of the FixedHashTable data
structure, which resembles a sparse graph with CSR storage).

The initial pass uses atomic updates in the Kokkos kernel.  This incurs
overhead even in the Kokkos::Serial case.  As a result, I only invoke
the kernel if the max number of hardware threads that the execution
space claims to have is greater than one (or if the execution space is
Kokkos::Cuda; see below).  It may be necessary to experiment to
determine the optimal minimum number of threads here, but using one as
the threshold ensures that existing MPI-only applications will not slow
down.

I did actually test this with the Kokkos::Cuda execution space.  That
made me discover that Kokkos::Cuda lacks the max_hardware_threads()
method that other Kokkos execution spaces have.  I had to resort to a
work-around: the method gets called on all other execution spaces, but I
assume that with Cuda, one always wants to run on the CUDA device.  That
may not be accurate but it works for now.