Tpetra: FixedHashTable now computes and exposes min & max values.
Tpetra::Details::FixedHashTable now computes min and max values, and
exposes them via the minVal() resp. maxVal() methods.  ("Value" here
refers to the result of get() on a key, NOT the result of calling the
hash function.)

The end goal is to give users a "Kokkos Map," so they can access a
subset of Tpetra::Map functionality in parallel kernels, in a
thread-safe and thread-scalable way.  Exposing the min and max values
lets FixedHashTable provide the same functionality as Tpetra::Map's
getMinLocalIndex and getMaxLocalIndex methods.

Adding this functionality to FixedHashTable has little cost compared
with the benefit of providing min and max local index functionality in
Kokkos kernels.  It does not increase the number of passes over the data
at initialization, and it only requires storing two additional ValueType
instances (in Tpetra, ValueType = LocalOrdinal (usually int) or int).
Furthermore, we can move the equivalent fields out of Tpetra::Map into
FixedHashTable (making the latter even more of a "local Map").

The only way in which _computing_ the min and max values takes any work
at all, is if invoking the FixedHashTable constructor that takes a list
of keys _and_ a list of values.  That constructor only gets invoked in
Tpetra::Directory, and only in the case of a "sparse" Map.  This case
matters for Zoltan2 partitioning, which often results in sparse Maps.
However, the main optimization for sparse Maps is the storage
optimization of using a hash table rather than a flat list.  I don't
mind two integers more storage for that application, with the benefit of
reusing FixedHashTable.

I also fixed a bug in the Tpetra::Details::FixedHashTable test.  I had
written a test for the "copy" constructor (between Device types) that
exercises the duplicate keys case.  However, the test wasn't being
built.  When I changed that, I noticed that the test was failing
unnecessarily.  I say unnecessarily, because for duplicate keys with
different values, FixedHashTable does not promise which value get()
returns.  I already had a hook for turning off testing values, so I used
that when I knew I was creating a table with duplicate keys.  This made
the test pass, yay!