Tpetra::Map::getLocalElement: Got rid of a function call.
Teuchos::as is a nice utility function, but it's still a function.  I
don't want any more levels of inlining in Tpetra::Map's getLocalElement
method.  It is performance critical and occurs often in users' code.
Thus, I've replaced calls to Teuchos::as with static_cast.  Hopefully
this may speed up build times a bit, both for Trilinos and for users.