Tpetra: FixedHashTable now has reverse map from values to keys.
Tpetra::Details::FixedHashTable now implements an optional "reverse map"
from values to keys, in its getKey() method (which is callable from a
Kokkos parallel kernel).  It does this via the "keepKeys" constructor
option I mentioned in a recent commit.  If keepKeys is true,
FixedHashTable keeps (a copy of) the input keys array.  If using the
constructor that assumes values occur in a contiguous sequence, this
allows looking up a key corresponding to a given value.  I also added to
the existing FixedHashTable test a test for getKey(), which passes.

This is part of making FixedHashTable into the "local" version of
Tpetra::Map.  Once we make FixedHashTable part of Tpetra's public
interface, we will rename getKey to getGlobalElement, for compatibility
with Tpetra::Map.  (I don't like the name "element," but we're stuck
with it for compatibility reasons.)