Tpetra: Added FixedHashTable ctors that take keys as a Kokkos::View.
Tpetra::Details::FixedHashTable now has constructors that take the input
array of keys as a 1-D Kokkos::View.  The public keys_type typedef in
FixedHashTable gives the desired type of this View.  I added tests for
the new constructors that illustrate how to create and fill a View of
keys, and use the View to create a FixedHashTable.

The two added constructors make a contiguous sequence of values.
Since we will always want to keep the keys (for the reverse map from
values to keys) in that case, these new constructors do NOT take a
keepKeys argument; they always keep the keys.