Tpetra: FixedHashTable now detects init sequence of contiguous keys.
Tpetra::Details::FixedHashTable now detects any initial sequence of
contiguous keys.  For the constructors that make a sequence of
contiguous values, FixedHashTable optimizes for the initial sequence
of contiguous keys, by not storing them in the hash table.

The (keys, vals) constructor (that takes lists of keys and of values)
is a special case.  In that case, we don't assume that the values are
contiguous.  A FixedHashTable instance remembers whether it was
created with that constructor.  If it was, it ignores the initial
sequence of contiguous keys when looking up a key.  If the run-time
overhead of checking that bool on every key lookup is of concern, we
could shift the lookup over to compile time, or have a separate lookup
function for that case.