[Phalanx] Add preprocessor guard around boost adjacency_list.hpp
The intel icpc compiler (v15) does not appear to compile the BGL
adjacency_list.hpp file, at least for boost versions between 105600
and 105800. This was documented / partially diagnosed in the Trilinos
bugzilla report [Bug 6343]. This commit sets up a guard around the
affected sections of Phalanx_Evaluator_Manager_Def.hpp for versions
of boost >= 105600 when icpc is used with:

// icpc cannot handle adjacency_list.hpp in BGL after boost 1.56.0
//#if (BOOST_VERSION<105600) || !defined(__INTEL_COMPILER)

This turns off the output and functioning of the evaluator graph
capabilities, printing a warning if the graph output function is
called.