Added shell support to the Elem-Elem graph
This basically involved grouping small collections of elements that
share a set of side nodes and filtering out the element types (shell vs.
volume element) that cannot "see" the type of element that we are
connecting to.  (e.g. Volume element connections are blocked if there
are intervening shell elements, and coincident shell elements can't see
each other.)

Tons of unit tests were added for volume-volume and volume-shell
combinations in serial and parallel.

The ElementGraph.test_element_death unit test was temporarily disabled
due to a face ID synchronization issue that still needs fixing in the
presence of shells.  A bit of code about face IDs was also commented out
in the ElemElemGraph class until we get a better understanding of what's
going on and can fix it.

Collaborators: Tolu Okusanya, Todd Coffey

(From Sierra)