Teuchos::Array: Finished class documentation.
I've been thinking about how to add memory tracking to the Teuchos
Memory Management classes.  The easiest way to do this in Array would be
to change the Allocator.

I was studying the class in order to see if this would work, and noticed
that the class had incomplete documentation ("ToDo: Finish
documentation!").  This commit fixes that.