Panzer: Modifying how scalar parameters are registered/initialized
Often this was being done in the evaluators constructor. This seemed
a bit strange and presented a considerable challenge depending on
how the model evaluator was being used. Now the use user of a thyra
ME uses the addParameter method to initialize the parameters.

Also this change had a slightly broader impact, so I propagated it
to the model evaluator factory and the (largely deprecated) epetra
model evaluator.

Also updated the unit tests