Teuchos:  Add config option for setting unittest repo global reduce default
This commit adds the configure option

-D Teuchos_GLOBALLY_REDUCE_UNITTEST_RESULTS=ON \

to set the default behavior of Teuchos::UnitTestRepository with regards
to globally reducing unittest pass/fail status.  The default is OFF
where no global reduction happens (which was the previous behavior),
which means the pass/fail status is only determined by the root
processor.  By turning this option on, developers can now investigate and
fix which of their tests now fail when the status of test results on non-root
processors are considered.  At some point in the future this will be
come the default.

This is for bug #6359.

Build/Test Cases Summary
Enabled Packages: Teuchos, TeuchosCore
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=1375,notpassed=0 (5.58 min)
1) SERIAL_RELEASE => passed: passed=1294,notpassed=0 (2.29 min)