Panzer: Modifying workset construction to not use physics blocks (optional)
This backward compatible change is a step toward being able to fully seperate
out the workset construction from the evaluator based portions of the
code.

This was done in a backward compatible way, however, the interfaces with "WorksetNeeds"
structure should be considered the primary functions to call in the future.

An issure remains to converting the workset container to only depend on
the needs. Basically, in order compute the orientations it needs a representative
field for a particular basis function.