PyTrilinos: Got all applicable Tpetra.Vector constructors to work
There is a memory deallocation problem with the Tpetra::Vector
constructor that takes a Teuchos::ArrayView, so I have removed it for
now.  I also removed the constructors that take a dual_view_type
argument, since this is not supported yet.

The copy constructors were a little tricky, because the concrete
classes are Vector_int, Vector_double, etc, and the factory function,
Vector(), should get the dtype information from the argument (as
opposed to a "dtype" keyword argument).  So now, objects that have a
C++ Scalar template (and are wrapped), also support a dtype() method,
which returns a string suitable for the numpy dtype() function (ints
return "i", floats return "f", etc.).  The Vector() factory function
tests this and then calls the appropriate concrete constructor.