EpetraExt,TrilinosCouplings: Fixed PETSc interface and example.
"PETSc interface" means an Epetra_RowMatrix subclass which wraps a PETSc
AIJ matrix.  The example creates an ML preconditioner using a PETSc AIJ
matrix (thusly wrapped).  It then runs an AztecOO solve and a PETSc KSP
solve, both using that preconditioner.

This change was tested with PETSc 3.6, the most recently released
version as of 09 June 2015.  The example in question builds but does not
actually run as a test.  Thus, it doesn't help to run the check-in test
script.  The example only builds if Trilinos enables the PETSc TPL.