Tpetra: Fixed failing Import_Union test (MUELU PERF. IMPLICATIONS?).
The Import_Union test has been failing for a while.  It turns out that
Import::setUnion (which the test exercises) was behaving correctly as
far as the test should be able to tell, but the test was too
restrictive.  This commit fixes the test so that it now passes.

See the following commit:

61d006cb74512c4d160c3c81845a9fd72db2d63d
Author: Mark Hoemmen <mhoemme@sandia.gov>
Date:   Thu Jun 25 14:14:52 2015 -0600

    Tpetra: Make Import_Union test actually show the failure that it should.

    Tpetra's Import_Union test (for Tpetra::Import::setUnion) should have
    been failing, but it wasn't, because the Teuchos unit test framework
    was ignoring output from processes other than Process 0.  This commit
    changes that.

    NOW THE TEST FAILS.  IT SHOULD HAVE FAILED.

    I also added more diagnostic output to the test.  setUnion is doing a
    funny thing, whereby the "same" IDs all get pushed into the "permute"
    IDs list.  The effect is that Import appears to work correctly, for
    the test.  I'm not sure if setUnion itself is correct, though.  I plan
    to add more tests to diagnose this.

I changed the test so that it compares Import objects for "functional"
equivalence, rather than exact equivalence.  "Functional" equivalence
means that the test groups "same IDs" and "permute IDs" together.
"Permute IDs" of one Import may include "same IDs" from the other
Import, as long as the sequence of same + permute IDs is the same on
both Import objects.

This test might have performance implications for MueLu, and for any
other code that relies on the sparse matrix-matrix multiply (which
uses Import::setUnion).  The test was failing before because the
result of setUnion put all the "same IDs" in the "permute IDs" list.
It still does this.  This makes the Import less efficient than it
could be.  Calling setUnion still saves setup time, but it might cost
something for later uses of Import.  We might want to fix this at some
point.