Tpetra: Fix for Bug 6377.
Tpetra::MatrixMarket::Writer::writeDense*, which writes
Tpetra::MultiVector or Tpetra::Vector in Matrix Market format to a
file (writeDenseFile) or output stream (writeDense), was not writing
values in scientific notation.  It was also not writing all of the
digits justified by the precision.  This commit fixes that.  Thanks to
Chris Luchini for the diagnosis and fix (for Bug 6377).