Tpetra: Fixed bug in test for Bug 6288.
The test for Bug 6288 had an exception sanity check for the size of a
communicator resulting from Teuchos::MpiComm::split.  The problem was
that the check only had the right size for the communicator on one
side of the split, not the other side (it was a two-color split).
This commit fixes that: the test now has the right size for both sides
of the split.

Build/Test Cases Summary
Enabled Packages: TpetraCore
Disabled Packages: CTrilinos,Didasko,FEApp,FEI,GlobiPack,Moertel,MOOCHO,Optika,OptiPack,PyTrilinos,RBGen,SEACAS,STK,Sundance,Shards
0) MPI_DEBUG => passed: passed=72,notpassed=0 (2.12 min)
1) SERIAL_RELEASE => passed: passed=54,notpassed=0 (0.59 min)
Other local commits for this build/test group: 799be4c