PyTrilinos: Fixed Epetra_*Vector wrappers so that all tests now pass
I added a copy() method to all Epetra_*Vector methods that returns the
results of the copy constructor.  Several MultiVector methods
implemented with %extend are now properly const.  I added a __call__()
method to MultiVector with %extend.  It behaves like operator()(int),
but the return type is more manageable (Epetra_Vector& instead of
Epetra_Vector*&).  I improved the logic of the MultiVector_getitem()
function, which gets assigned to be the MultiVector.__getitem__()
operator.