MueLu: add a new MatrixAnalysis factory.
The factory checks some basic properties of A on the current level and prints
some output and statistics on the screen which help to debug convergence problems
or to detect the occurrence of Nans or Inf entries along the simulation.

The factory only accepts A as input and pipes it through to the output without changes.
This way, the MatrixAnalysis can be plugged in everywhere in the multigrid setup layout.