Domi: Fixed dynamic type checking algorithm
The MDVector class has conversion methods such as
getEpetraVectorView(), which returns a Epetra_Vector whose memory
buffer is the same as the parent MDVector.  Since this uses view
semantics, it is only valid for the template parameter Scalar =
double.  You can't do template sepcialization on a single method (or a
subset of methods).  Rather, you have to specialize the entire class,
so the only way to implement this is to dynamically check the type of
Scalar.  If it is not double, then throw an exception.

My check for doing this was primitive, i.e., I was checking the
typeid(Scalar).name() against 'double', which, it turns out, doesn't
always work.  In PyTrilinos, I was getting typeid(Scalar).name() ==
'd' for some reason.  A better check is typeid(Scalar) ==
typeid(double), and this fix resulted in my DAP typemaps working in
PyTrilinos.