PyTrilinos: Ignore certain Epetra_CrsMatrix methods
There are certain Epetra_CrsMatrix methods (Multiply, Multiply1, and
Solve) that have two versions, one that takes Epetra_MultiVectors, and
one that takes Epetra_Vectors.  The ones that take Epetra_Vectors are
redundant, and they do not behave correctly with the new DAP typemaps.
What happens is the typemaps for Epetra_Vector arguments at first
reject an Epetra_MultiVector, but then accept it because it has a
__distarray__() method.  It then tries to force an Epetra_MultiVector
peg into an Epetra_Vector hole.

This change ignores the Epetra_Vector versions of the methods, which
prevents the problem from happening.  The Epetra_MultiVector versions
of the methods work just fine, and all tests pass.