Epetra: Improved Bug 6079 test & fixed build warning (unused variable).
I fixed an "unused variable" build warning in the test that claims to
be for Bug 6079.  The unused variable in question was a return code on
an Import.  I fixed this by doing an all-reduce over the return code
to ensure that the function actually claimed to succeed on all
processes.  Thus, the return code is actually used now.