Only generate Trilinos_version.h for SIERRA for Trilinos project only (Trilinos #6367)
The way it was, the file ${PROJECT_NAME}_version.h was getting generated for
all TriBITS projects.  This was messing up the creation of a tarball for SCALE
(see Trilinos #6367).  It is not clear why this was not messing up the
creation of the CASL VERA tarball as well.

Now, this file only gets generated back into the source tree when Trilinos is
the base TriBITS project (and when tarball generation is enabled, which it is
by default for Trilinos-based TriBITS projects).

 LocalWords:  unstage cmake

Build/Test Cases Summary
Enabled Packages: CTeuchos
0) MPI_RELEASE_DEBUG_SHARED => passed: passed=4,notpassed=0 (3.83 min)
Other local commits for this build/test group: c016f5e