Amesos2,Ifpack2: Disable LinearSolverFactory test if ETI OFF.
That's not exactly what I did, but it gets at the general idea.  What
I _actually_ did was disable the tests when
Trilinos_ENABLE_LINEAR_SOLVER_FACTORY_REGISTRATION is OFF.  This is ON
by default if and only if ETI (explicit template instantiation) is
enabled.  Thus, by default, the LinearSolverFactory test in these two
packages won't run unless ETI is ON.  LinearSolverFactory is really
only useful if ETI is ON, and the test takes too long to build
otherwise.