MueLu: bugfix in AddNonSerializableDataToHierarchy
Do not overwrite entries in FactoryManager entries with NoFactory for
"Nullspace" and "Coordinates".
Usually, one wants to set the "Nullspace" and "Coordinates" on the finest
level and uses the MueLu internal level transfer mechanism for these variables.
Overwriting FactoryManager data would only make sense if the user provides the
data for all levels. If the user wants to set some specific data on a certain level only,
the preferred way to do so would be to introduce a level-specific parameter list in
an advanced xml input file where NoFactory is declared as generating factory together with
the provided nonserializable data.