MueLu: add a SegregatedAFactory
Factory which accepts a Xpetra::Map which is supposed to be a submap of the
row map of A. Drops all off-diagonal entries (i,j) if i is in the provided
map and j is not in the provided map (and vice versa). This way we can segregate
a set of DOFs.
Note: we drop the entries and build a new matrix with a clean graph as the
CoalesceDropFactory works on the graph and does not take zero values into
consideration.