EXODUS: Modification to permit adding set to output database created via ex_copy
Modified ex_put_sets to work correctly when adding a set (node, edge,
face, element) to an output database that has had the mesh portion
created via a call to ex_copy.  There is a small kluge in this
proof-of-concept in that the id of the new set must be passed as the
negative of the desired set id.  Example usage is shown below:

   ex_init_params info;
   ex_get_init_ext(exoid, &info);
   info.num_node_sets++;
   ex_put_init_ext(exoid1, &info); /* Must be called before ex_copy */

   ex_copy (exoid, exoid1);

   ex_set new_set;
   new_set.id = -999;  /* Negative of desired set id */
   new_set.type = EX_NODE_SET;

   new_set.entry_list = NULL;
   new_set.extra_list = NULL;
   new_set.num_entry = 100;
   new_set.num_distribution_factor = 0;

   ex_put_sets(exoid1, 1, &new_set);
   ex_put_name(exoid1, EX_NODE_SET, -new_set.id, "NewNodeset");

(From Sierra)