Ifpack2::Details::Amesos2Wrapper now uses LinearSolverFactory inside.
The implementation has a temporary work-around for Bug 6392: It calls
Amesos2's LinearSolverFactory registration function, if it hasn't yet
been called for _some_ combination of template parameters.  This
doesn't guarantee that it has been called for the solver's combination
of template parameters.  However, users are always responsible for
handling this case on their own.  Furthermore, the whole point of this
commit is to avoid compiling all of Amesos2 (yet again, if ETI is
OFF).

NOTE: Ifpack2 still must depend on Amesos2, because of the
aforementioned temporary work-around for Bug 6392.  For now, if we
want to remove Ifpack2's dependency on Amesos2, we need to move any
Ifpack2 tests that invoke Amesos2 out of Ifpack2, and into a package
that depends on both of them.  This will let us invoke the
LinearSolverFactory registration functions in both Amesos2 and
Ifpack2.  Once we fix Bug 6392, we can then move the tests back into
Ifpack2.