Ifpack2::Details::LinearSolver now can wrap a custom Preconditioner.
We want to use LinearSolver as AdditiveSchwarz's inner solver.  This
commit will make that possible, but note that this commit doesn't do
that yet.

AdditiveSchwarz can take a custom Ifpack2::Preconditioner as its inner
solver.  We don't know that Preconditioner's type.  It could even be a
custom Preconditioner subclass that the user wrote.  As a result, we
can't use LinearSolverFactory or Ifpack2::Factory to (re)create it.
However, we don't need to, because it has a setMatrix() method
(AdditiveSchwarz requires that the input custom Preconditioner inherit
from Ifpack2::Details::CanChangeMatrix).

This commit gives LinearSolver a constructor that can take a custom
Ifpack2::Preconditioner.  It handles this case in the same way that
AdditiveSchwarz did, by calling its name "CUSTOM" and using
setMatrix() to change the matrix, rather than recreating the
preconditioner if the matrix changes.