Ifpack2: Fix ETI for ReorderFilter and SingletonFilter.
(ETI stands for explicit template instantiation.)

Whoever did ETI for ReorderFilter and SingletonFilter, only did so for
MatrixType = Tpetra::CrsMatrix.  AdditiveSchwarz wasn't even using
this; it was setting MatrixType = Tpetra::RowMatrix.  The only reason
that it worked was that AdditiveSchwarz included the _def.hpp files
for both ReorderFilter and SingletonFilter.  This was entirely wrong
and likely shows that AdditiveSchwarz was written before Ifpack2 had
ETI.  This commit fixes ETI for ReorderFilter and SingletonFilter, so
that they only get instantiated for MatrixType = Tpetra::RowMatrix.
This should strip some unused symbols out of the Ifpack2 library, and
improve build times.  (Tpetra::RowMatrix is a smaller class than
Tpetra::CrsMatrix.)