Panzer: Abstract an internal interface's WorksetDetails.
1. Workset inherits from WorksetDetails. If the workset is for an interface, it
contains another WorksetDetails for the elements on the other side of the
interface.

I've cleaned up the handling of this logic. There is no longer a
vector<RCP<WorksetDetails>>; instead, there is a single member
    RCP<WorksetDetails> other.

2. Workset::operator(int), details(int), numDetails(), and are available to
workset builders. These abstract the previous details[i]. details(int) is meant
as convenience so one can write
    workset->details(i)
instead of
    (*workset)(i).
Both return [const] WorksetDetails&.

These changes have the side benefit that Workset copy ctor and op= can remain
default; the explicit assignment of rcp(this, false) to details[0] is no longer
necessary and so does not need to be dealt with in non-default copy and op=.

These changes are propagated to the workset building code.

3. Evaluators may use WorksetDetailsAccessor as a convenient abstraction. Its
goal is to minimize textual change to evaluators and to maintain backwards
compatibility. I'm committing the changes to the evaluator DOF as an
example. Usage is as follows:

In _decl.hpp, declare wda:

  class DOF ... {
    ...
  private:
    ...
    WorksetDetailsAccessor wda;
  };

In _impl.hpp, the constructor initializes wda by passing it p:

  DOF::DOF(const Teuchos::ParameterList& p) :
    ...
    wda(p)
  { ... }

The parameter list may contain an entry "Details Index", and wda(p) deals with this.

Then in evaluateFields, the operator() method is used to abstract the details index:

  DOF::evaluateFields(workset) {
    // previously: workset.bases[basis_index]
    wda(workset).bases[basis_index]
  }