Panzer: Further removal of PhysicsBlocks dependency form workset container
Specifically fixed the workset container to use only the
lookupNeeds class inside. This is intended to make the workset
construction and maintenance a little more reliable internally

This resulted in the removal of the "lookupPhysicsBlock" method

Also workset container doesn't appear to use the workset size that
is being passed in. So the needs doesn't require one.