Tpetra::Map: Made the insides of some Map methods safe for Kokkos.
In the case of Tpetra, the methods remain host functions.  (Otherwise,
this would break Tpetra when the execution space is not Cuda, but CUDA
is enabled.)  The intent here is to make Tpetra's public interface
thread safe and thread scalable on the host.  At some point,
Tpetra::Map will make a "Kokkos object" available, which will provide
a subset of Map functionality on the CUDA device (if applicable).