Panzer: Introduce panzer::EvaluatorWithBaseImpl.
Phalanx provides PHX::EvaluatorWithBaseImpl. The new WorksetDetails affects many
evaluators. For convenience, insert a new class
  panzer::EvaluatorWithBaseImpl<Traits> : public PHX::EvaluatorWithBaseImpl<Traits>.
This class holds the WorksetDetailsAccessor. In the future, it could handle
other things that affect all evaluators.

To support this new class, provide macros that correspond to a subset of those
in Phalanx_Evaluator_Macros.hpp.

In addition, it was decided that the details index would be set using
setDetailsIndex rather than p.get<int>("Details Index"). WorksetDetailsAccessor
was modified accordingly.

A possible alternative to the WorksetDetailsAccessor is to consolidate its
functionality into panzer::EvaluatorWithBaseImpl. However, some evaluators call
non-class functions with panzer::Traits::EvalData (== panzer::Workset) as an
argument. The WorksetDetailsAccessor can be passed along with the workset. If
panzer::EvaluatorWithBaseImpl were to provide WorksetDetailsAccessor's
functionality without a wrapping class, then the details index would have to be
exposed to the evaluator, which I want to avoid.