Panzer: Make some manual changes in preparation for next commit.
The next commit will apply a script to Panzer to make a bunch of little changes
in many files.

In this commit:

1. Modify the workset utilities functions to take a
WorksetDetailsAccessor. wda(workset) will give the utility routine the correct
WorksetDetails. *Temporarily* provide the original signatures so that Charon and
other Panzer users don't have to be modified yet.

2. Apply the script to a few files with special-case usage and handle the
special cases manually.