Tpetra::Details::Lapack128: Remove unneeded use of {Arith,Scalar}Traits.
This will help fix Bug 6305.  We need to make Lapack128 the
implementation of the __float128 specialization of Teuchos::LAPACK, so
that downstream packages have access to it.  As a result, we need to
make it possible to move Lapack128 into the Numerics subpackage of
Teuchos.  This means that we need to minimize dependencies on things
outside of Teuchos.