decomp: change to note use launch if in an hpc allocation and reset openmpi mtl to ^psm
This is a direct response to a series of tickets that have
currently been consolidated into #13645.  Basically it was
found that decomp runs on the login nodes of TLCC2 machines
were failing with the exception of uno.  The difference
turned out to be that the transport was set to openib
even though the login nodes don't have that.  This resets
all of the openmpi variables and does not use launch unless
we are in a PBS/MOAB, SLURM, or LSF node allocation.

(From Sierra)