Tpetra::Map: Added test for Bug 6412 (it currently fails).
Thanks to Karen Devine for contributing a test for Bug 6412.
(Tpetra::Map::getRemoteIndexList, for a noncontiguous Map, does not
correctly handle repeated global indices on the same process in its
input list (NOT in the Map).)  I modified the test slightly to make it
use the Teuchos unit test framework.

The test builds, but fails.  We're supposed to be using test-driven
development, right?  First you add a failing test, then you make the
test pass.  I would rather do that in two commits, in case people need
to back out the Map changes.