Fix a field-data initialization bug in stk-mesh.
In the case where an entity is moved from a bucket where a field is not allocated
to a bucket where the field is allocated, the field-data for the entity was not
being explicitly initialized. After fixing this issue, ran all sierra tests and 1
aria test diffed. Checked with David Noble and the test is being reblessed in a
companion commit.

(From Sierra)