Change logic for when to do a full continuous build
Previously every 16 hours the continuous build did a full build when it
ran. This was intended to make the first build of the day be a full build
so the following builds could be incremental. However, due to the length
of the build now the first build of the day was moved from 7am to 1am to
ensure the first full build was done before the start of the day. The
continuous testing runs late into the day which allowed one of the later
builds to be a full build and caused the intended first build to be
incremental. Increasing this time along with adjusting the last build
to earlier in the day should prevent this from happening again.