Panzer: Redesign of the interface conditions infrastructure.
UniqueGlobalIndexer<LO,GO> now returns a ConnManagerBase<LO>.
EpetraLinearObjectFactory uses this in buildGhostedGraph instead of using the
now-removed method getElementAndAssociatedGIDs. This will be the pattern for
other similar situations.

Like the UniqueGlobalIndexer, the ConnManager must now be explicitly told to
enable the interface infrastructure. An example of doing this is in
Panzer_STK_ModelEvaluatorFactory_impl.hpp.

ScatterResidual no longer calls getElementAndAssociatedLIDs, and that method has
been removed. Instead, it uses workset.other, if non-null, to get the local cell
IDs across the interface.

Add a unit test to test the associated-element methods in STKConnManager. Rework
PoissonInterfaceExample's construction phase.