Tpetra::CrsMatrix::{replace,sumInto}LocalValues: Use Kokkos::View inside
For the common case (locally indexed matrix), replaceLocalValues and
sumIntoLocalValues now use Kokkos::View internally, when accessing the
current entries of the local row of the matrix.

The point is to avoid using Teuchos::ArrayView, which would make these
methods thread safe even in debug mode.  I'm not entirely done with
that, but I'm getting close.