Use observer for updating commspec and local-ids.
Rather than refreshing comspecs for each entity rank every time modification-end is called,
use an observer/notifier pattern to only refresh the commspec for a given rank when
something has changed in the mesh specific to that rank. This removes ~80% of the commspec fixup
cost for the primary test case we've been profiling.

Also update the Apub::ParallelComm data structures based on bucket and comm-info changes in
stk-parallel, using the same information.
This allows it to update less frequently. Previously it was updating after
each modification-end.

Paired with Todd Coffey

(From Sierra)