Tpetra: Move CMake logic for what types are enabled into top level.
The CMake logic for deciding what Scalar, GlobalOrdinal, and Node types
to enable (NOTE: this is independent of whether ETI (explicit template
instantiation) is enabled) used to live in tpetra/core/CMakeLists.txt.
Now it lives in tpetra/CMakeLists.txt.

This will make it easier to tie TpetraKernels' explicit instantiation
system to TpetraCore's, thus improving build times when ETI is ON and
preventing unneeded instantiations from being built (whether or not ETI
is ON).