PyTrilinos: Add Array(string) as a supported PyTrilinos ParameterList value
This required specializations of the following PyTrilinos functions:

  NumPy_TypeCode< std::string >()
  copyNumPyToTeuchosArray< std::string >(...)
  copyTeuchosArrayToNumPy< std::string >(...)

and upgrades in the

  setPythonParameter(...)
  getPythonParameter(...)

functions.  I also stubbed in specializations and parameter set/get
for Array(bool), but these are commented out because I get
compilations errors when I try to compile them.  It looks like this
scalar type is not tested for Teuchos::Array[View|RCP] and the
std::vector<bool> specialization causes problems.