Amesos2/tests:
The simple matrix used by the SolverFactory test with row and column map the
same and a diagonal causes trouble for lot of reasons. One example being the
parmetis call from Superlu_dist which will fail with a null array, and a
diagonal matrix results in a null array as parmetis doesn't like self edges.
We will use a simple tridiagonal matrix instead in our test. Work with
SuperLU_dist 2.5. Passing it on to Josh as he is debugging other issues with
SuperLU_dist > 2.5.