PyTrilinos: Debug DistArray Protocol for Tpetra (Multi)Vectors
The implementation of the DistArray Protocol for Tpetra Vectors is
tested by the exEpetra_DAP example script, which now constructs an
Epetra.LinearProblem using a LHS and RHS that are Tpetra.Vectors.  I
had to debug several components to get this to work.

The converters first create a DistArrayProtocol object from a
PyObject.  The object has to be a dictionary returned from the
__distarray__() method, but I was giving it a Tpetra.Vector, so the
fix was to call the method and pass the result.

During the debugging process, certain exceptions were not getting
caught by PyTrilinos.  The fix was to catch them in the PyTrilinos
functions and convert them to PyTrilinos::PythonExceptions, which were
already being properly caught by the wrappers.

In the __distarray__() method, the 'size' attribute was the local size
but needed to be the global size.  The 'start' and 'stop' attributes
were from 0...local size, rather than the local bounds.  these bugs
were fixed.

For the Tpetra.MultiVector getData() methods, I needed typemaps for
converting Teuchos::ArrayRCP objects to numpy arrays.  These typemaps
were added.