Tpetra::BlockMultiVector: Add "offset view" constructors (Bug 6450).
Tpetra::Experimental::BlockMultiVector now has two "offset view"
constructors.  They behave analogously to the offsetView and
offsetViewNonConst methods of Tpetra::MultiVector.  The constructors
view an existing BlockMultiVector with a different mesh Map, and an
optional local row offset from which to start the view on each process.
The offset is a mesh offset (it gets multiplied by the block size
internally in order to find the point offset).  The two constructors
differ only in that one lets you supply the new point Map, while the
other computes it for you.

I added a unit test in tpetra/core/test/Block/ExpBlockMultiVector.cpp
that exercises both constructors.

This is a partial fix for Bug 6450.  It is not a complete fix, because I
need to add the same constructors to BlockVector as well.