Added Penso unit tests of Krino-style and Percept-style rebalance needs
Both of these use-cases manage "parent" and "child" elements as a result
of refinement, but they do it in different ways.  Krino manages its own
data structures that hold these relationships and Percept uses
CONSTRAINT_RANK entities (called "family trees") to hold relations to
the parent and its child elements.

These use-cases require implementing a call-back function to manage
movement of the child elements to the same processor that parent
elements were moved to.  This is very similar in concept to the Fuego
particle use-case of moving particle elements to follow their parent.

A couple of disabled tests are also included which fail by allowing the
parent and child elements to be rebalanced independently, breaking up
the complete bundle of the parent and all of the children.

Collaborators: Tolu Okusanya, Dave Glaze, Jay Mosso

(From Sierra)