Tpetra::Crs{Graph,Matrix}: Fix Bug 6468.
Tpetra::CrsMatrix::{replace,sumInto}GlobalValues no longer allocate a
temporary array of local indices, if the matrix is locally indexed.
Instead, they call CrsGraph::transformGlobalValues directly in that
case.  CrsGraph::transformGlobalValues then uses the column Map to
convert each input global column index, one at a time, into a local
column index.  This avoids the intermediate allocation.  It should also
be thread safe.

NOTE: replaceGlobalValues and sumIntoGlobalValues are NOT currently
thread safe!  This is because they call CrsMatrix::getRowMap(), which
returns RCP by value.  That updates the RCP's reference count, which is
not currently thread safe.  We could probably fix this by pushing the
test into CrsGraph, so that it can dereference the RCP (rowMap_)
directly.  That does not change the RCP's reference count.

NOTE: This is not exactly the same patch as the one I posted to the
Bug 6468 discussion in Bugzilla.  This commit fixes some unused typedef
warnings.