Tribits: fix issue with download-cmake.py and proxies
A recent change to cmake.org moved all their files behind https. The
current way we were setting up proxies didn't work when SSL was in the
mix. Unfortunately it seems the deficiency was with the urllib that was
being used to do the communication with cmake.org. Thankfully urllib2
seems to handle SSL much better, though the code to download a file
is a bit more verbose with urllib2. Its better proxy handler system
solves our issue.

I also added an option for an https proxy since these are different
and setting them to the same thing doesn't seem to work at Sandia.

Manually tested with numerous options, both setting the proxies,
using the default proxies and setting the proxies to the empty
string to force the system to use the proxies from the env.